% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_acsendyear.R
\name{acsendyear}
\alias{acsendyear}
\title{helper - try to report the end year of the ACS 5-year survey as reported by current EJAM metadata for the package,
or try to infer the possible year}
\usage{
acsendyear(guess_as_of = Sys.Date(), guess_always = FALSE)
}
\arguments{
\item{guess_as_of}{optional alternative data to use if guessing what is available
as of this date, e.g., "2025-01-01" -- must be Date class like Sys.Date()}

\item{guess_always}{optional, set TRUE to ignore metadata and just guess at year.}
}
\value{
a single year like "2022" as character string, meaning ACS5 for 2018-2022
}
\description{
helper - try to report the end year of the ACS 5-year survey as reported by current EJAM metadata for the package,
or try to infer the possible year
}
\keyword{internal}
