% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_address.R
\name{address_from_table}
\alias{address_from_table}
\title{get USPS addresses from a table of that info}
\usage{
address_from_table(x)
}
\arguments{
\item{x}{data.frame with address info in column(s)}
}
\value{
vector of USPS addresses, 1 per row of x
}
\description{
get USPS addresses from a table of that info
}
