% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_run_EJAMejscreenapi.R
\name{app_run_EJAMejscreenapi}
\alias{app_run_EJAMejscreenapi}
\title{DRAFT - Launch EJAM ejscreenapi web app in RStudio}
\usage{
app_run_EJAMejscreenapi(
  onStart = NULL,
  options = list(),
  uiPattern = "/",
  enableBookmarking = NULL,
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{...}{arguments to pass to golem_opts
See \code{?golem::get_golem_options} for more details}
}
\description{
Launch EJAM ejscreenapi web app in RStudio
}
\details{
app_run_EJAM()                 is like run_app() for the EJAM shiny app run from the EJAM package

app_run_EJAMejscreenapi()      is like run_app() for the EJScreen API shiny app run from the EJAM package
}
\keyword{internal}
