% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{app_server - EJAM app server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.
DO NOT REMOVE.}
}
\description{
app_server - EJAM app server
}
\keyword{internal}
