% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{app_ui - The application User-Interface}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Internal parameter for \code{{shiny}}.
DO NOT REMOVE.}
}
\description{
app_ui - The application User-Interface
}
\keyword{internal}
