% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ask_number.R
\name{ask_number}
\alias{ask_number}
\alias{askradius}
\title{utility - interactive prompt in RStudio to ask user to specify number like radius}
\usage{
ask_number(
  default = 3,
  title = "Radius",
  message = "Within how many miles of each point?"
)
}
\arguments{
\item{default}{default value for the number to be provided and returned}

\item{title}{title of popup dialog box, like "Radius"}

\item{message}{question, like, "Within how many miles of each point?"}
}
\value{
a single number
}
\description{
same as askradius()
}
\seealso{
askYesNo()
}
\keyword{internal}
