% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata_add.R
\name{attributes2}
\alias{attributes2}
\title{utility to see all attributes other than row.names}
\usage{
attributes2(x)
}
\arguments{
\item{x}{R object}
}
\value{
attributes of x, excluding row.names
}
\description{
utility to see all attributes other than row.names
}
\keyword{internal}
