% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bgej.R
\name{bgej}
\alias{bgej}
\title{bgej (DATA) Summary Indexes for Census block groups}
\description{
bgej is a table of all blockgroups, with the raw scores of the Summary Indexes
and supplemental summary indexes for all the environmental indicators.
}
\details{
This file is not stored in the package, but is obtained via \code{\link[=dataload_from_pins]{dataload_from_pins()}}.

For documentation on the residential population and environmental data and indicators used,
see \url{https://web.archive.org/web/20250118193121/https://www.epa.gov/ejscreen/understanding-ejscreen-results}.

The column names are these:

\if{html}{\out{<div class="sourceCode">}}\preformatted{c("bgfips", "bgid", "ST", "pop", 
names_ej, 
names_ej_supp, 
names_ej_state,
names_ej_supp_state
)
}\if{html}{\out{</div>}}
}
