% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bgid2fips.R
\name{bgid2fips}
\alias{bgid2fips}
\title{bgid2fips (DATA) Census FIPS codes of block groups}
\description{
bgid2fips (DATA) Census FIPS codes of block groups
}
\details{
For documentation on EJScreen, see \href{https://www.epa.gov/ejscreen}{EJScreen documentation}

bgid2fips is a table of all census block groups, with their FIPS codes.

It also has a column
called \code{blockid} that can join it to other block datasets.

\if{html}{\out{<div class="sourceCode">}}\preformatted{  dataload_from_pins('bgid2fips')
  
  names(bgid2fips)
  dim(bgid2fips)
}\if{html}{\out{</div>}}
}
