% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_from_latlon.R
\name{bgid_from_blockid}
\alias{bgid_from_blockid}
\title{given vector of blockids, get bgid of each (the parent block group)}
\usage{
bgid_from_blockid(blockid)
}
\arguments{
\item{blockid}{vector of block ids like in blockwts data.table or blockpoints}
}
\value{
vector of bgid values
}
\description{
given vector of blockids, get bgid of each (the parent block group)
}
\keyword{internal}
