% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_bgpts.R
\docType{data}
\name{bgpts}
\alias{bgpts}
\title{bgpts (DATA) lat lon of popwtd center of blockgroup, and count of blocks per block group}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 242355 rows and 5 columns.
}
\usage{
bgpts
}
\description{
This is just a list of US block groups and how many blocks are in each. It also has the lat and lon roughly of each blockgroup
}
\details{
The point used for each bg is the Census 2020 population weighted mean of the internal points of the blocks. It gives an approximation of where people live and where each bg is, which is useful for some situations. Has all US States, DC, PR, but not  "AS" "GU" "MP" "VI" (and not U.S. Minor Outlying Islands FIPS 74 UM)
}
\keyword{datasets}
