% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_blockgroupstats.R
\docType{data}
\name{blockgroupstats}
\alias{blockgroupstats}
\title{blockgroupstats (DATA) residential population and environmental indicators for Census block groups}
\description{
The dataset of residential population data and environmental indicators originally from EJScreen.

For Summary Indexes, see \link{bgej}
}
\details{
\itemize{
\item Version 2.32.2 of EJAM
uses EPA environmental indicators and population indicators derived from
ACS data from `r a2018-2022 Each year this should be re-created as for the latest version.
See attributes(blockgroupstats)
}

It is a data.table of US Census blockgroups (not blocks).
With PR, and Island Areas

Column names include bgfips, bgid (for join to blockwt$bgid), pop, pctlowinc, etc.
}
