% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_blockpoints.R
\name{blockpoints}
\alias{blockpoints}
\title{blockpoints (DATA) Census blocks locations}
\description{
blockpoints (DATA) Census blocks locations
}
\details{
For documentation on EJScreen, see (https://www.epa.gov/ejscreen)

blockpoints is a table of all census blocks, with the lat, lon
providing the latitude and longitude of the Census Bureau-defined
internal point, like a centroid, of each block.

It also has a column
called blockid that can join it to other block datasets.

\if{html}{\out{<div class="sourceCode">}}\preformatted{dataload_from_pins('blockpoints')

names(blockpoints)
dim(blockpoints)
}\if{html}{\out{</div>}}
}
