% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_xls_from_ejam.R
\name{buffer_desc_from_sitetype}
\alias{buffer_desc_from_sitetype}
\title{helper to pick text phrase to use in excel report notes tab, for Locations analyzed: _____}
\usage{
buffer_desc_from_sitetype(sitetype, site_method)
}
\arguments{
\item{sitetype}{character string, one of "shp", "latlon", "fips" and to describe locations}

\item{site_method}{string used in filename for saved report and to describe locations
site_method can be SHP, latlon, FIPS, NAICS, FRS, EPA_PROGRAM, SIC, or MACT}
}
\value{
text string
}
\description{
helper to pick text phrase to use in excel report notes tab, for Locations analyzed: _____
}
\keyword{internal}
