% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_barplot_report.R
\name{build_barplot_report}
\alias{build_barplot_report}
\title{Generate HTML Page for Summary Barplot Report in shiny app}
\usage{
build_barplot_report(
  analysis_title,
  totalpop,
  locationstr,
  in_shiny = FALSE,
  filename = NULL
)
}
\arguments{
\item{analysis_title, }{title to use in header of report}

\item{totalpop, }{total population included in location(s) analyzed}

\item{locationstr, }{description of the location(s) analyzed}

\item{in_shiny, }{whether the function is being called in or outside of shiny - affects location of header}

\item{filename, }{path to file to save HTML content to; if null, returns as string (used in Shiny app)}
}
\description{
Creates header and footer of 1 page report to include a barplot on results for one site (to supplement the EJScreen Community Report)
}
\details{
For a related function for use in RStudio,
see \code{\link[=ejam2report]{ejam2report()}} which relies on \code{build_community_report()}
}
\keyword{internal}
