% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROXIMITY_FUNCTIONS.R
\name{calc_bgwts_bysite}
\alias{calc_bgwts_bysite}
\title{aggregate blockwt values by blockgroup, by site they are in or near}
\usage{
calc_bgwts_bysite(sites2blocks)
}
\arguments{
\item{sites2blocks}{like output of \code{\link[=getblocksnearby]{getblocksnearby()}}
or input to \code{\link[=doaggregate]{doaggregate()}} or \code{\link[=custom_doaggregate]{custom_doaggregate()}}}
}
\value{
data.table, 1 row per site-bg pair.
May have same bgid or bgfips in 2,3, more rows
since it is here once per site that the bg is near.
It is like a sites2blockgroups table.
}
\description{
aggregate blockwt values by blockgroup, by site they are in or near
}
\seealso{
\code{\link[=custom_doaggregate]{custom_doaggregate()}}
}
\keyword{internal}
