% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calc_ejam.R
\name{calc_byformula}
\alias{calc_byformula}
\title{DRAFT utility to use formulas provided as text, to calculate indicators}
\usage{
calc_byformula(mydf, formulas = NULL, keep = formula_varname(formulas))
}
\arguments{
\item{mydf}{data.frame of indicators or variables to use}

\item{formulas}{text strings of formulas}

\item{keep}{useful if some of the formulas are just interim steps
creating evanescent variables created only for use in later formulas
and not needed after that}
}
\value{
data.frame of results, but
if mydf was a data.table, returns a data.table
}
\description{
DRAFT utility to use formulas provided as text, to calculate indicators
}
