% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_censusplaces.R
\docType{data}
\name{censusplaces}
\alias{censusplaces}
\title{censusplaces (DATA) Census FIPS and other basic info on roughly 40,000 cities/towns/places}
\format{
An object of class \code{data.frame} with 38195 rows and 7 columns.
}
\usage{
censusplaces
}
\description{
Table of US cities and other Census Designated Places, Incorporated Places, and County Subdivisions
}
\details{
from \url{https://www2.census.gov/geo/docs/reference/codes/PLACElist.txt}
Column names: eparegion, ST, stfips, countyname, countyfips, placename, fips
}
\keyword{datasets}
