% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.summarize.helpers.R
\name{change.fieldnames}
\alias{change.fieldnames}
\title{change.fieldnames
one way to convert variable names (colnames) to friendlier versions}
\usage{
change.fieldnames(allnames, oldnames, newnames, file = NA, sort = FALSE)
}
\arguments{
\item{allnames}{a vector of all the original fieldnames,}

\item{oldnames}{a vector of just the fieldnames to be changed, and}

\item{newnames}{a vector of what those should be change to}

\item{file}{path to a csv file with two columns: oldnames, newnames (instead of passing them to the function as parameters)}

\item{sort}{a logical (default is FALSE). If FALSE, return new fieldnames. If sort=TRUE, return vector of indexes giving new position of given field, based on sort order of oldnames.}
}
\value{
new versions of names as vector
}
\description{
change.fieldnames
one way to convert variable names (colnames) to friendlier versions
}
\keyword{internal}
