% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colcounter_xyz.R
\name{colcounter_summary_cum}
\alias{colcounter_summary_cum}
\title{Summarize how many rows have AT LEAST N columns at or above (or below) various thresholds}
\usage{
colcounter_summary_cum(
  x,
  thresholdlist,
  or.tied = TRUE,
  na.rm = TRUE,
  below = FALSE,
  one.cut.per.col = FALSE
)
}
\arguments{
\item{x}{Data.frame or matrix of numbers to be compared to threshold value,
like percentiles for example.}

\item{thresholdlist}{vector of numeric threshold values to compare to}

\item{or.tied}{if TRUE, include ties (value in x equals threshold)}

\item{na.rm}{if TRUE, used by colcounter to count only the non-NA columns in given row}

\item{below}{if TRUE, count x below threshold not above threshold}

\item{one.cut.per.col}{if FALSE, compare each threshold to all of x.
If TRUE, specify one threshold to use for each column.}
}
\description{
See colcounter_summary() for details and examples
}
\seealso{
\code{\link[=colcounter_summary_all]{colcounter_summary_all()}} \code{\link[=colcounter_summary]{colcounter_summary()}} \code{\link[=colcounter_summary_cum]{colcounter_summary_cum()}} \code{\link[=colcounter_summary_pct]{colcounter_summary_pct()}} \code{\link[=colcounter_summary_cum_pct]{colcounter_summary_cum_pct()}}
}
\keyword{internal}
