% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colcounter_xyz.R
\name{colcounter_summary_pct}
\alias{colcounter_summary_pct}
\title{Summarize what percent of rows have N columns at or above (or below) various thresholds}
\usage{
colcounter_summary_pct(x, thresholdlist, ...)
}
\arguments{
\item{x}{Data.frame or matrix of numbers to be compared to threshold value,
like percentiles for example.}

\item{thresholdlist}{vector of numeric threshold values to compare to}

\item{...}{passed to colcounter_summary()
like or.tied=TRUE, na.rm=TRUE, below=FALSE, one.cut.per.col=FALSE}
}
\description{
See colcounter_summary() for details and examples
}
\seealso{
\code{\link[=colcounter_summary_all]{colcounter_summary_all()}} \code{\link[=colcounter_summary]{colcounter_summary()}} \code{\link[=colcounter_summary_cum]{colcounter_summary_cum()}} \code{\link[=colcounter_summary_pct]{colcounter_summary_pct()}} \code{\link[=colcounter_summary_cum_pct]{colcounter_summary_cum_pct()}}
}
\keyword{internal}
