% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.summarize.helpers.R
\name{colMaxs2}
\alias{colMaxs2}
\title{Get maximum of each column of data.frame}
\usage{
colMaxs2(df, na.rm = TRUE)
}
\arguments{
\item{df}{data.frame or matrix.
Can include numbers stored as character or as factors.}

\item{na.rm}{default is TRUE. not tested for FALSE}
}
\value{
named vector of numbers
}
\description{
Get maximum of each column of data.frame
}
\examples{
df <- rbind(NA, data.frame(
  n1 = c(0, 0:8), n2 = c(0.1 + (0:9)), n3 = c(1:10), 
  allnas = c(rep(NA, 10)),
  logic = TRUE, 
  factortxt = factor('factor'), 
  txt = 'words',
  numberlike = as.character(6:15),
  numberlikefact = factor(as.character(6:15)),
  stringsAsFactors = FALSE))
  
df
colMaxs2(df)
}
\seealso{
\code{\link[=colMins2]{colMins2()}}
}
\keyword{internal}
