% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROXIMITY_FUNCTIONS.R
\name{countpoints_after_getpoints}
\alias{countpoints_after_getpoints}
\title{utility - count topoints near each frompoint, AFTER getpointsnearby() or getfrsnearby() or getblocksnearby() was already run}
\usage{
countpoints_after_getpoints(
  sites2points,
  frompoints_id_colname = "ejam_uniq_id",
  topoints_id_colname = "blockid",
  radius = NULL
)
}
\arguments{
\item{sites2points}{output of function like \code{\link[=getpointsnearby]{getpointsnearby()}} or \code{\link[=getfrsnearby]{getfrsnearby()}} or \code{\link[=getblocksnearby]{getblocksnearby()}}}

\item{frompoints_id_colname}{character string name of column in sites2points that is the unique ID of frompoints}

\item{topoints_id_colname}{character string name of column in topoints that is the unique ID of counted points}

\item{radius}{optional, should be less than or equal to radius originally used to create
sites2points. If radius is provided here, this function counts only the topoints that are at a distance of
less than or equal to this radius (which is likely only a subset of all points within original radius used
to create sites2points). You can run for example,

s2s <- getpointsnearby(frompoints = testpoints_10\link{1,}, topoints = frs_from_naics("cement"), radius = 30)
countpoints_after_getpoints(s2s)
countpoints_after_getpoints(s2s, radius = 20)
countpoints_after_getpoints(s2s, radius = 10)
countpoints_after_getpoints(s2s, radius = 5)
countpoints_after_getpoints(s2s, radius = 3)}
}
\value{
counts data.table with column N for count, and a column named via frompoints_id_colname
}
\description{
utility - count topoints near each frompoint, AFTER getpointsnearby() or getfrsnearby() or getblocksnearby() was already run
}
\examples{
countpoints_after_getpoints(testoutput_getblocksnearby_10pts_1miles)

}
\seealso{
\code{\link[=countpoints_nearby]{countpoints_nearby()}}
}
\keyword{internal}
