% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROXIMITY_FUNCTIONS.R
\name{countpoints_nearby}
\alias{countpoints_nearby}
\title{count the number of topoints within a specified radius of each frompoint}
\usage{
countpoints_nearby(frompoints, topoints, radius = 3)
}
\arguments{
\item{frompoints}{table with lat,lon colnames}

\item{topoints}{table with lat,lon colnames}

\item{radius}{search radius in miles from each of frompoints}
}
\value{
table or vector??  1 row per frompoints row, column
attribute called "unique" will store the count of unique topoints overall
}
\description{
count the number of topoints within a specified radius of each frompoint
}
\examples{
# countpoints_nearby(pts_sites, pts_features, radius = 3)



}
\keyword{internal}
