% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filename.R
\name{create_filename}
\alias{create_filename}
\title{create_filename - construct custom filename for downloads in EJAM app}
\usage{
create_filename(
  file_desc = "",
  title = "",
  buffer_dist = 0,
  site_method = "",
  with_datetime = TRUE,
  ext = NULL
)
}
\arguments{
\item{file_desc}{file description, such as "short report", "long report", "results_table"}

\item{title}{analysis title}

\item{buffer_dist}{buffer distance}

\item{site_method}{site selection method, such as NAICS, FRS, SHP, latlon}

\item{with_datetime}{boolean to include date and time}

\item{ext}{optional file extension. will check for '.' and add if not provided}
}
\value{
Returns string of pasted filename with specified components
}
\description{
Builds filename from file description, analysis title, buffer distance, and site selection method. Values are pulled from Shiny app if used there.
}
\examples{
# specify all arguments
create_filename(file_desc = 'results_table', title = 'My Title', buffer_dist = 1, site_method = 'NAICS', with_datetime=TRUE, ext = '.xlsx')
# specify title only
create_filename(title = 'Summary of EJ Analysis')

}
\keyword{internal}
