% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interactive_table.R
\name{create_interactive_table}
\alias{create_interactive_table}
\title{create_interactive_table - function that creates interactive html version of site by site table in app}
\usage{
create_interactive_table(data_processed, testing)
}
\arguments{
\item{data_processed}{processed data from app_server}

\item{testing}{variable to indicate if analysis is being tested}
}
\description{
Builds site by site table after an analysis in EJAM app. Pulls in uploaded and analyzed data to create table
}
\keyword{internal}
