% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROXIMITY_FUNCTIONS.R
\name{custom_doaggregate}
\alias{custom_doaggregate}
\title{custom version of doaggregate(), to calculate user-provided indicators}
\usage{
custom_doaggregate(
  sites2blocks,
  custom_blockgroupstats = blockgroupstats,
  countcols = "pop",
  popmeancols = names_these,
  wtcols = "pop",
  custom_formulas = NULL,
  custom_cols = NULL,
  custom_map_headernames = map_headernames
)
}
\arguments{
\item{sites2blocks}{see \code{\link[=doaggregate]{doaggregate()}}}

\item{custom_blockgroupstats}{like blockgroupstats but with custom
indicators, one value per block group, with colnames bgid, bgfips, pop}

\item{countcols}{vector of colnames in custom_blockgroupstats to be
aggregated as sums of counts, like population counts}

\item{popmeancols}{vector of colnames in custom_blockgroupstats to be
aggregated as weighted means, population weighted or with other weights}

\item{wtcols}{vector of colnames to use as the weights for wtd means,
same length as popmeancols, but not used yet}

\item{custom_formulas}{like formulas_all,  not used yet}

\item{custom_cols}{not used yet}

\item{custom_map_headernames}{like map_headernames but for the
custom indicators}
}
\value{
list of tables similar to what \code{\link[=doaggregate]{doaggregate()}} returns
}
\description{
custom version of doaggregate(), to calculate user-provided indicators
}
