% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload_from_entirefolder.R
\name{dataload_from_entirefolder}
\alias{dataload_from_entirefolder}
\title{dataload_from_entirefolder}
\usage{
dataload_from_entirefolder(folder = "./data")
}
\arguments{
\item{folder}{path}
}
\value{
nothing. just loads to global envt
}
\description{
Loads into global environment all .rda files found in specified folder
}
\keyword{internal}
