% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload_from_local.R
\name{dataload_from_local}
\alias{dataload_from_local}
\title{Load datasets from local disk folder}
\usage{
dataload_from_local(
  varnames = .arrow_ds_names[1:3],
  envir = globalenv(),
  folder_local_source = NULL,
  justchecking = FALSE,
  testing = FALSE,
  silent = FALSE,
  return_data_table = TRUE
)
}
\arguments{
\item{varnames}{use defaults, or vector of names like "bgej" or use "all" to get all available}

\item{envir}{use defaults. see \code{\link[=dataload_from_pins]{dataload_from_pins()}}}

\item{folder_local_source}{Your local folder path. see \code{\link[=dataload_from_pins]{dataload_from_pins()}}}

\item{justchecking}{use defaults. see \code{\link[=dataload_from_pins]{dataload_from_pins()}}}

\item{testing}{use defaults}

\item{silent}{set to TRUE to stop cat() printing to console like when running tests}

\item{return_data_table}{whether the \code{\link[=read_ipc_file]{read_ipc_file()}} should return a data.table (T, the default), or arrow (F)}
}
\value{
vector of paths to files (as derived from varnames) that were
actually found in folder_local_source,
but only for those not already in memory, so it is
just the ones loaded from disk because not already in memory and found on disk locally.
}
\description{
Utility for analysts / developers to store large block /
other data locally instead of redownloading
}
\details{
See \code{\link[=dataload_from_pins]{dataload_from_pins()}} also.

rm(bgid2fips, blockid2fips, blockpoints, blockwts, quaddata)

dataload_from_local(folder_local_source = '.')
}
