% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrite_to_aws.R
\name{datawrite_to_aws}
\alias{datawrite_to_aws}
\title{Utility NOT YET WORKING - AccessDenied}
\usage{
datawrite_to_aws(
  varnames = c("bgid2fips", "blockid2fips", "blockpoints", "blockwts", "quaddata"),
  ext = c(".arrow", ".rda")[2],
  fun = c("arrow::write_ipc_file", "save")[2],
  mybucket = "dmap-data-commons-oa",
  mybucketfolder = "EJAM",
  justchecking = TRUE
)
}
\arguments{
\item{varnames}{vector of object names to upload}

\item{ext}{file .extension appropriate to the format and fun, like ".rda" or ".arrow"}

\item{fun}{function to use, but as a character string, like "arrow::write_ipc_file"
but fun is ignored if ext=".rda" since it then just uses s3save()}

\item{mybucket}{do not need to change}

\item{mybucketfolder}{do not need to change}

\item{justchecking}{set this to FALSE to actually upload instead of
just viewing in console the commands to be used, to test/check this}
}
\value{
the paths of the objects on server
}
\description{
Utility to write object(s) like a dataset to DMAP Data Commons, formatted as .arrow or .rda
}
\details{
mybucket <-  'dmap-data-commons-oa' #

bucket_contents <- data.table::rbindlist(

\if{html}{\out{<div class="sourceCode">}}\preformatted{get_bucket(bucket = mybucket, prefix = "EJAM"),

fill = TRUE
  
)
}\if{html}{\out{</div>}}

bucket_contents
}
\keyword{internal}
