% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrite_to_local.R
\name{datawrite_to_local}
\alias{datawrite_to_local}
\title{Save datasets during package development}
\usage{
datawrite_to_local(
  varnames = .arrow_ds_names,
  fnames = paste0(varnames, ".arrow"),
  localfolder = "~/../Downloads",
  justchecking = F,
  overwrite = FALSE
)
}
\arguments{
\item{varnames}{vector of object names}

\item{fnames}{optional vector of file names including extension ".rda" or ".arrow", but without path.
Default is .arrow}

\item{localfolder}{path to local folder without slash at end}

\item{justchecking}{set this to FALSE to actually save instead of
just seeing in console info or the commands to be used, to test/check this}

\item{overwrite}{Set to TRUE to overwrite file if it exists already, with new copy.}
}
\value{
the paths of the objects as requested to be saved whether or not actually done
}
\description{
Utility to write large object(s) like EJAM datasets to local disk for convenience during app/pkg development,
formatted as .arrow or .rda
}
\examples{
  x = 1
  datawrite_to_local("x", localfolder = ".", justchecking = T) 
  
}
\seealso{
\code{\link[=dataload_from_local]{dataload_from_local()}}
}
\keyword{internal}
