% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawrite_to_pins.R
\name{datawrite_to_pins}
\alias{datawrite_to_pins}
\title{utility - write data objects to pins board in .arrow format}
\usage{
datawrite_to_pins(
  varnames = .arrow_ds_names,
  type = "arrow",
  justchecking = FALSE,
  silent = FALSE,
  access_type = "all",
  boardfolder = "Mark",
  auth = "auto",
  server = "https://rstudio-connect.dmap-stage.aws.epa.gov"
)
}
\arguments{
\item{varnames}{vector of quoted names of datasets, like c("bgej", "bgid2fips")}

\item{type}{"arrow"}

\item{justchecking}{can set to TRUE to just see a list of what pins are stored in that board}

\item{silent}{set to TRUE to suppress cat() msgs to console}

\item{access_type}{"all" means public (if Posit Connect instance allows that)}

\item{boardfolder}{if needed to specify a different folder than default}

\item{auth}{See help documentation for \code{\link[pins:board_connect]{pins::board_connect()}}}

\item{server}{if needed to specify a server other than default (which might be
stored in envt variable CONNECT_SERVER or be registered via the rsconnect package).
Note if auth = "envvar" then it looks for CONNECT_SERVER to get name of server which
needs to be the full url starting with https:// - see help for board_connect}
}
\value{
varnames vector
}
\description{
utility - write data objects to pins board in .arrow format
}
\keyword{internal}
