% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_default_points_shown_at_startup.R
\docType{data}
\name{default_points_shown_at_startup}
\alias{default_points_shown_at_startup}
\title{Default Points Shown at Startup
keywords internal}
\format{
An object of class \code{data.frame} with 2 rows and 6 columns.
}
\usage{
default_points_shown_at_startup
}
\description{
Default Points Shown at Startup
keywords internal
}
\keyword{datasets}
