% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames_etc.R
\name{dependencies_of_ejam}
\alias{dependencies_of_ejam}
\title{utility for developing package, see what pkgs it depends on, recursively (i.e., downstream ones too)
Reminder of ways to check this is printed to console.}
\usage{
dependencies_of_ejam(
  localpkg = "EJAM",
  depth = 6,
  ignores_grep = "0912873410239478"
)
}
\arguments{
\item{localpkg}{"EJAM" or another installed package}

\item{depth}{would be used if using the deepdep package and function}

\item{ignores_grep}{would be used if using the deepdep package and function}
}
\description{
utility for developing package, see what pkgs it depends on, recursively (i.e., downstream ones too)
Reminder of ways to check this is printed to console.
}
\keyword{internal}
