% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_linesofcode_.R
\name{dir2}
\alias{dir2}
\title{utility - quick way to do dir() but for query pattern like  *.xlsx
useful if you forget \code{\link[=glob2rx]{glob2rx()}} is available}
\usage{
dir2(
  query_glob = "*.*",
  ignore.case = TRUE,
  recursive = FALSE,
  silent = FALSE,
  ...
)
}
\arguments{
\item{query_glob}{optional, something like "\emph{.r}" or "\emph{.xlsx" or "myfile.}" or "a*.*",
so it is like \code{dir(pattern = glob2rx(query_glob))}}

\item{ignore.case}{passed to \code{\link[=dir]{dir()}}}

\item{recursive}{passed to \code{\link[=dir]{dir()}}, often useful to set TRUE}

\item{silent}{whether to avoid printing to console}

\item{...}{passed to \code{\link[=dir]{dir()}}}
}
\value{
vector of paths
}
\description{
utility - quick way to do dir() but for query pattern like  *.xlsx
useful if you forget \code{\link[=glob2rx]{glob2rx()}} is available
}
\examples{
dontrun{
dir2()
dir2("*.zip", recursive = T)
dir2("*.y*",  recursive = T)

# if recursive=T, left-aligned view of paths is shown
dir2("*datacreate*.*", recursive = T)
# for right-aligned view:
data.frame(hit=dir2("*address*.*", recursive = T))

dir2("*.csv*", path = testdatafolder(installed = FALSE), recursive = T)
}

}
\keyword{internal}
