% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_near_eachother.R
\name{distance_near_eachother}
\alias{distance_near_eachother}
\title{which points are near any of the others in a list?}
\usage{
distance_near_eachother(lon, lat, distance, or_tied = FALSE)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{distance}{distance between points in miles to check}

\item{or_tied}{if TRUE, checks if less than or equal to distance, otherwise if less than}
}
\value{
logical vector the length of lon or lat, telling if the point is
within distance of any other point in list for example,
which sites have residents that might also be near others sites?
}
\description{
which points are near any of the others in a list?
}
\examples{
mapfast(testpoints_500[distance_near_eachother(
  lon = testpoints_500$lon, 
  lat = testpoints_500$lat, 
  3.1), ], radius = 3.1)
   
}
\keyword{internal}
