% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejamit_compare_distances.R
\name{distance_trends}
\alias{distance_trends}
\title{Which indicators fall most as proximity does? (i.e., are higher if closer to site)
Which variables have strongest trend with distance based on slope of linear fit}
\usage{
distance_trends(
  results_bydistance,
  myvars = names_d_subgroups_ratio_to_state_avg,
  radii,
  n = 1
)
}
\arguments{
\item{results_bydistance}{data.frame of a few indicators, no other columns, taken from output of \code{\link[=ejamit_compare_distances]{ejamit_compare_distances()}}}

\item{myvars}{optional, vector of some colnames of results_bydistance}

\item{radii}{optional vector - taken from results_bydistance$radius.miles}

\item{n}{optional number of indicators to list. n=3 would mean show the top 3.}
}
\value{
vector of text names of indicators
}
\description{
Which indicators fall most as proximity does? (i.e., are higher if closer to site)
Which variables have strongest trend with distance based on slope of linear fit
}
\details{
Used by \code{\link[=ejamit_compare_distances2plot]{ejamit_compare_distances2plot()}} which is used by \code{\link[=ejamit_compare_distances]{ejamit_compare_distances()}}
}
\examples{
distance_trends(ejamit_compare_distances(testpoints_10, radii = c(1,3)))
}
