% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_via_surfacedistance.R
\name{distance_via_surfacedistance}
\alias{distance_via_surfacedistance}
\title{Convert surface distance to actual distance}
\usage{
distance_via_surfacedistance(x)
}
\arguments{
\item{x}{surface distance in miles}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{\preformatted{
    Just a simple formula:
   earthRadius_miles <- 3959
   angle_rad <- x/earthRadius_miles
   # Calculate  radius * cord length
   return( earthRadius_miles * 2*sin(angle_rad/2) )
   }
}\if{html}{\out{</div>}}
}
\keyword{internal}
