% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROXIMITY_FUNCTIONS.R
\name{doaggregate_newscores}
\alias{doaggregate_newscores}
\title{For user-provided indicators and formulas, aggregate at each site and overall}
\usage{
doaggregate_newscores(
  sites2blocks,
  userstats,
  formulas = NULL,
  sites2states_or_latlon = NA,
  radius = NULL,
  countcols = "pop",
  popmeancols = NULL,
  calculatedcols = NULL,
  varsneedpctiles = NULL,
  usastats_newscores = NULL,
  statestats_newscores = NULL,
  ...
)
}
\arguments{
\item{sites2blocks}{output of \code{\link[=getblocksnearby]{getblocksnearby()}}, as for \code{\link[=doaggregate]{doaggregate()}}}

\item{userstats}{like blockgroupstats but data.frame or data.table of all US
blockgroups and one or more columns of user provided raw indicator scores
and any other variables needed for formulas to aggregate indicators
across block groups in each site.}

\item{formulas}{a character vector of formulas in R code (see formulas_d for
an example), that use variables in userstats to calculate any
derived indicators or aggregated ones, for cases where just a sum or a
population weighted mean is not the right way to aggregate some indicator.
Formulas can include intermediate steps, or can aggregate across all places.

For example one formula might be

"pctover64 = ifelse(pop == 0, 0, over64 / pop)"}

\item{sites2states_or_latlon}{see \code{\link[=doaggregate]{doaggregate()}}}

\item{radius}{see \code{\link[=doaggregate]{doaggregate()}}}

\item{countcols}{see \code{\link[=doaggregate]{doaggregate()}}}

\item{popmeancols}{see \code{\link[=doaggregate]{doaggregate()}}}

\item{calculatedcols}{see \code{\link[=doaggregate]{doaggregate()}}}

\item{varsneedpctiles}{see \code{\link[=doaggregate]{doaggregate()}}}

\item{usastats_newscores}{calculated if not provided}

\item{statestats_newscores}{calculated if not provided}

\item{...}{not used}
}
\value{
see \code{\link[=doaggregate]{doaggregate()}}
}
\description{
Like doaggregate() but for any user-provided indicator available for all
block groups (nationwide so that US percentiles make sense, or at least statewide)
}
\keyword{internal}
