% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_latest_arrow_data.R
\name{download_latest_arrow_data}
\alias{download_latest_arrow_data}
\title{Download latest versions of arrow datasets if user doesn't have them already}
\usage{
download_latest_arrow_data(
  varnames = .arrow_ds_names,
  repository = NULL,
  envir = globalenv()
)
}
\arguments{
\item{varnames}{use defaults, or vector of names like "bgej" or use "all" to get all available}

\item{repository}{repository name such as https://usepa.github.io/EJAM-open or "XYZ/ejamdata"
(wherever the ejamdata repo is hosted, as specified in the DESCRIPTION file of this package)}

\item{envir}{if needed to specify environment other than default, e.g., globalenv() or parent.frame()}
}
\description{
Used when EJAM package is attached
}
\details{
Checks to see what is the latest version of datasets available according to a repository's latest release tag.
Compares that to what version was last saved locally (as stored in the installed package's
ejamdata_version.txt file).

Relies on \code{\link[piggyback:pb_releases]{piggyback::pb_releases()}} to download data files from a specific release (version) of the package.
}
\keyword{internal}
