% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames_etc.R
\name{dupenames}
\alias{dupenames}
\title{UTILITY - check conflicting exported function or data names}
\usage{
dupenames(
  pkg = EJAM::ejampackages,
  sortbypkg = FALSE,
  compare.functions = TRUE
)
}
\arguments{
\item{pkg}{one or more package names as vector of strings.
If "all" it checks all installed pkgs, but takes very very long potentially.}

\item{sortbypkg}{If TRUE, just returns same thing but sorted by package name}

\item{compare.functions}{If TRUE, sends to console inf about whether body and formals
of the functions are identical between functions of same name from different packages.
Only checks the first 2 copies, not any additional ones (where 3+ pkgs use same name)}
}
\value{
data.frame with columns Package, Object name (or NA if no dupes)
}
\description{
See what same-named objects (functions or data) are exported by some (installed) packages
}
\details{
utility to find same-named exported objects (functions or datasets) within source code
of 2+ packages, and see what is on search path, for dev renaming / moving functions/ packages

See \code{\link[=dupeRfiles]{dupeRfiles()}} for files supporting a shiny app that is not a package, e.g.

See \code{\link[=dupenames]{dupenames()}} for objects that are in R packages.

See \code{\link[=functions_in_pkg]{functions_in_pkg()}}

See \code{\link[=datapack]{datapack()}}
}
\keyword{internal}
