% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames_etc.R
\name{dupeRfiles}
\alias{dupeRfiles}
\title{UTILITY - check conflicting sourcefile names}
\usage{
dupeRfiles(folder1 = "../EJAM/R", folder2 = "./R")
}
\arguments{
\item{folder1}{path to other folder with R source files}

\item{folder2}{path to a folder with R source files, defaults to "./R"}
}
\description{
See what same-named .R files are in 2 sourcecode folders
}
\details{
See \code{\link[=dupeRfiles]{dupeRfiles()}} for files supporting a shiny app that is not a package, e.g.

See \code{\link[=dupenames]{dupenames()}} for objects that are in R packages.

See \code{\link[=functions_in_pkg]{functions_in_pkg()}}

See \code{\link[=datapack]{datapack()}}
}
\keyword{internal}
