% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flagged_FUNCTIONS.R
\name{ejam2areafeatures}
\alias{ejam2areafeatures}
\title{simple way to see the table of summary stats on special areas and features like schools}
\usage{
ejam2areafeatures(ejamitout)
}
\arguments{
\item{ejamitout}{output from ejamit()}
}
\value{
a data frame with the summary of flagged areas
}
\description{
simple way to see the table of summary stats on special areas and features like schools
}
\details{
In this table, summary stats mean the following:
\itemize{
\item The "flag" or "yesno" indicators here are population weighted sums, so they show
how many people from the analysis live in blockgroups
that overlap with the given special type of area, such as
non-attainment areas under the Clean Air Act.
\item The "number" indicators are counts for each site in the
\code{ejamit()$results_overall} table, but here are summarized as
what percent of residents overall in the analysis have
AT LEAST ONE OR MORE of that type of site in
the blockgroup they live in.
\item The "pctno" or \% indicators are summarized as what \% of the
residents analyzed lack the critical service.
}
}
\seealso{
\code{\link[=ejam2barplot_areafeatures]{ejam2barplot_areafeatures()}} \code{\link[=batch.summarize]{batch.summarize()}}
}
