% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_flagged_FUNCTIONS.R
\name{ejam2barplot_areafeatures}
\alias{ejam2barplot_areafeatures}
\title{barplot of summary stats on special areas and features at the sites}
\usage{
ejam2barplot_areafeatures(
  ejamitout,
  main =
    "\% of analyzed population that lives in blockgroups with given features or that overlap given area type",
  ylab = "Ratio of Indicator in Analyzed Locations / in US Overall",
  shortlabels = NULL
)
}
\arguments{
\item{ejamitout}{output from ejamit()}

\item{main}{optional title for plot}

\item{ylab}{optional y axis label}

\item{shortlabels}{optional alternative labels for the bars}

\item{caption}{left blank to avoid default caption used in \code{\link[=plot_barplot_ratios]{plot_barplot_ratios()}}}
}
\value{
ggplot2 plot
}
\description{
Summary of whether residents at the analyzed locations are more likely to have
certain types of features (schools) or special areas (Tribal, nonattainment, etc.)
}
\details{
See
\code{varinfo(c(names_featuresinarea, names_flag, names_criticalservice))[,c("longname", "varlist")]}

These are the indicator summary stats shown:
\itemize{
\item \link{1,} "Number of Schools"
\item \link{2,} "Number of Hospitals"
\item \link{3,} "Number of Worship Places"
\item \link{4,} "Flag for Overlapping with Tribes"
\item \link{5,} "Flag for Overlapping with Non-Attainment Areas"
\item \link{6,} "Flag for Overlapping with Impaired Waters"
\item \link{7,} "Flag for Overlapping with CJEST Disadvantaged Communities"
\item \link{8,} "Flag for Overlapping with EPA IRA Disadvantaged Communities"
\item \link{9,} "Flag for Overlapping with Housing Burden Communities"
\item \link{10,} "Flag for Overlapping with Transportation Disadvantaged Communities"
\item \link{11,} "Flag for Overlapping with Food Desert Areas"
\item \link{12,} "\% Households without Broadband Internet"
\item \link{13,} "\% Households without Health Insurance"
}
}
\examples{
out <- testoutput_ejamit_1000pts_1miles
ejam2barplot_areafeatures(out)

shortlabels = EJAM:::flagged_areas_shortlabels_from_ejam(out)
ejam2barplot_areafeatures(out, shortlabels = shortlabels)

}
\seealso{
\code{\link[=ejam2areafeatures]{ejam2areafeatures()}} \code{\link[=batch.summarize]{batch.summarize()}}
}
