% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2barplot_indicators.R
\name{ejam2barplot_indicators}
\alias{ejam2barplot_indicators}
\title{Create facetted barplots of groups of indicators}
\usage{
ejam2barplot_indicators(
  ejamitout,
  indicator_type = "Demographic",
  data_type = "raw"
)
}
\arguments{
\item{ejamitout}{output from running an EJAM analysis, with ejamit or the EJAM shiny app}

\item{indicator_type}{group of indicators to display, such as 'Demographic', 'Environmental','EJ', or 'EJ supplemental'}

\item{data_type}{form to display data in: 'raw' or 'ratio'}
}
\value{
ggplot object with facets for each indicator and 3 bars
}
\description{
Create facetted barplots of groups of indicators
}
\examples{
ejam2barplot_indicators(testoutput_ejamit_1000pts_1miles, 'Demographic', 'raw')
}
