% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2barplot_sites.R
\name{ejam2barplot_sitegroups}
\alias{ejam2barplot_sitegroups}
\title{Barplot comparing groups of sites on 1 indicator, for output of ejamit_compare_types_of_places()
easy high-level function for getting a quick look at top few groups of sites}
\usage{
ejam2barplot_sitegroups(
  ejamitout,
  varname = "pctlowinc",
  names.arg = NULL,
  main = "Sites by Type",
  xlab = "Groups or Types of Sites",
  ylab = NULL,
  sortby = NULL,
  topn = 10,
  ...
)
}
\arguments{
\item{ejamitout}{list that is output of ejamit_compare_types_of_places(), where one element is
a table named results_bytype}

\item{varname}{name of a column in results_bytype, bar height}

\item{names.arg}{optional vector of labels on the bars, like the types of sites represented by each group}

\item{main}{optional, for barplot}

\item{xlab}{optional, for barplot}

\item{ylab}{optional, for barplot, plain English version of varname, indicator that is bar height}

\item{sortby}{set to FALSE if you want to have no sorting, or to an increasing vector
that provides the sort order}

\item{topn}{optional, show only the top n groups (site types) -- Does not show all
by default -- only shows top n groups.}

\item{...}{passed to barplot()}
}
\value{
same as \code{\link[=barplot]{barplot()}}
}
\description{
Barplot comparing groups of sites on 1 indicator, for output of ejamit_compare_types_of_places()
easy high-level function for getting a quick look at top few groups of sites
}
\details{
see \code{\link[=ejamit_compare_types_of_places]{ejamit_compare_types_of_places()}} for more examples
}
\examples{
 out <- ejamit_compare_types_of_places(testpoints_10[1:4, ], 
   typeofsite <- c("A", "B", "B", "C"))
   cbind(Rows_or_length = sapply(out, NROW))
  
 ejam2barplot_sitegroups(out, "sitecount_unique", topn=3, sortby = F)
 
}
\seealso{
\code{\link[=plot_barplot_sitegroups]{plot_barplot_sitegroups()}} \code{\link[=ejamit_compare_distances2plot]{ejamit_compare_distances2plot()}}
}
