% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot_ratios.R
\name{ejam2boxplot_ratios}
\alias{ejam2boxplot_ratios}
\title{Make boxplot of ratios to US averages}
\usage{
ejam2boxplot_ratios(
  ejamitout,
  radius,
  varnames = c(names_d, names_d_subgroups),
  main = NULL,
  maxratio = 5
)
}
\arguments{
\item{ejamitout}{output from an EJAM analysis, like from \code{\link[=ejamit]{ejamit()}}}

\item{radius}{buffer radius used for an analysis}

\item{varnames}{currently only works with names_d and names_d_subgroups}

\item{main}{can specify a main title to use instead of default}

\item{maxratio}{largest ratio to plot in case of outliers, so plot looks better}
}
\value{
ggplot object
}
\description{
Make boxplot of ratios to US averages
}
\details{
IMPORTANT:
NOTE this uses the ratio at each site
USING THE AVERAGE RESIDENT AT THAT SITE,
SO A BOXPLOT SHOWS ONE DOT PER SITE AND THE BOXPLOT IS NOT POP WTD
MEANING IT SHOWS THE MEDIAN AND 75TH PERCENTILE SITE NOT RESIDENT, ETC.
}
\examples{
ejam2boxplot_ratios(testoutput_ejamit_1000pts_1miles, radius=1)

out <- testoutput_ejamit_100pts_1miles
ejam2boxplot_ratios(out, radius=1)
ejam2boxplot_ratios(out)
## not ## ejam2boxplot_ratios(out$results_bysite)

}
