% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2excel.R
\name{ejam2excel}
\alias{ejam2excel}
\title{Save EJAM results in a spreadsheet}
\usage{
ejam2excel(
  ejamitout,
  fname = NULL,
  save_now = TRUE,
  overwrite = TRUE,
  launchexcel = FALSE,
  interactive_console = TRUE,
  ok2plot = TRUE,
  in.testing = FALSE,
  in.analysis_title = "EJAM analysis",
  react.v1_summary_plot = NULL,
  radius_or_buffer_in_miles = NULL,
  buffer_desc = "Selected Locations",
  radius_or_buffer_description =
    "Miles radius of circular buffer (or distance used if buffering around polygons)",
  hyperlink_colnames = c("EJScreen Report", "EJScreen Map", "ECHO report"),
  site_method = "",
  mapadd = FALSE,
  ...
)
}
\arguments{
\item{ejamitout}{output of \code{\link[=ejamit]{ejamit()}}}

\item{fname}{optional name or full path and name of file to save locally, like "out.xlsx"}

\item{save_now}{optional logical, whether to save as a .xlsx file locally or just return workbook object
that can later be written to .xlsx file using \code{\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook()}}}

\item{overwrite}{optional logical, passed to \code{\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook()}}}

\item{launchexcel}{optional logical, passed to \code{\link[=table_xls_format]{table_xls_format()}}, whether to launch browser to see spreadsheet immediately}

\item{interactive_console}{optional - should set to FALSE when used in code or server. If TRUE,
prompts RStudio user interactively asking where to save the downloaded file}

\item{ok2plot}{optional logical, passed to  \code{\link[=table_xls_format]{table_xls_format()}}, whether safe to try and plot or set FALSE if debugging plot problems}

\item{in.testing}{optional logical}

\item{in.analysis_title}{optional title as character string}

\item{react.v1_summary_plot}{optional - a plot object}

\item{radius_or_buffer_in_miles}{optional radius in miles}

\item{buffer_desc}{description of location to use in labels, like "Selected Locations"}

\item{radius_or_buffer_description}{optional text phrase describing places analyzed}

\item{hyperlink_colnames}{optional names of columns with URLs}

\item{site_method}{site selection method, such as NAICS, FRS, SHP, latlon,
optional site method parameter used to create a more specific title with create_filename}

\item{mapadd}{logical option for including a map of the points}

\item{...}{optional additional parameters passed to \code{\link[=table_xls_format]{table_xls_format()}}, such as
heatmap_colnames, heatmap_cuts, heatmap_colors, etc.}
}
\value{
returns a workbook object for use by openxlsx::saveWorkbook(wb_out, pathname)
or returns just the full path/file name of where it was saved if save_now = TRUE
}
\description{
ejam2excel() takes the output of something like ejamit() and
creates a spreadsheet with an overall summary tab, a site by site table tab,
as well as other tabs such as map, plot, notes, etc.
}
\examples{
\dontrun{
# Add purple to flag indicators at 99th percentile
ejam2excel(testoutput_ejamit_10pts_1miles, 
  # View spreadsheet 1st without saving it as a file
  launchexcel = T, save_now = F, 
  heatmap_cuts = c(80, 90, 95, 99), 
  heatmap_colors  = c("yellow", "orange", "red", "purple"), 
  # Apply heatmap to only a few of the ratio columns
  heatmap2_colnames = names_d_ratio_to_state_avg)
}

}
