% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2histogram.R
\name{ejam2histogram}
\alias{ejam2histogram}
\title{Histogram of single indicator from EJAM output}
\usage{
ejam2histogram(
  ejamitout,
  varname,
  distn_type = "Sites",
  data_type = "raw",
  n_bins = 30,
  sitetype = ejamitout$sitetype
)
}
\arguments{
\item{ejamitout}{output of an EJAM analysis, like from \code{\link[=ejamit]{ejamit()}}}

\item{varname}{indicator name, such as 'Demog.Ind' or 'pctlowinc'}

\item{distn_type}{group to show distribution across, either 'Sites' or 'People'}

\item{data_type}{type of values to show for the indicator, either 'raw' or 'pctile'}

\item{n_bins}{number of bins}

\item{sitetype}{what type of sites were analyzed, like \code{\link[=ejamit]{ejamit()}} 'sitetype'. Examples are 'latlon', 'FIPS', 'SHP'}

\item{...}{passed to \code{\link[=plot_barplot_ratios]{plot_barplot_ratios()}}}
}
\description{
Histogram of single indicator from EJAM output
}
\examples{
ejam2histogram(testoutput_ejamit_1000pts_1miles, 'Demog.Index', distn_type='Sites', data_type='raw')
}
