% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2map.R
\name{ejam2map}
\alias{ejam2map}
\title{Show EJAM results as a map of points}
\usage{
ejam2map(ejamitout, radius = NULL, column_names = "ej", browse = FALSE)
}
\arguments{
\item{ejamitout}{output of ejamit()}

\item{radius}{radius in miles}

\item{column_names}{can be "ej", passed to \code{\link[=mapfast]{mapfast()}}}

\item{browse}{logical optional whether to open the web browser to view the map}
}
\value{
like what \code{\link[=mapfastej]{mapfastej()}} returns
}
\description{
Takes the output of ejamit() and uses \code{\link[=mapfastej]{mapfastej()}} to
create a map of the points.
}
\details{
Gets radius by checking ejamitout$results_overall$radius.miles
You can use browse=TRUE to save it as a shareable .html file
and see it in your web browser.
}
\examples{
pts = testpoints_100
mapfast(pts)

# out = ejamit(pts, radius = 1)
out = testoutput_ejamit_100pts_1miles

# See in RStudio viewer pane
ejam2map(out)
mapfastej(out$results_bysite[c(12,31),])
\dontrun{

# See in local browser instead
ejam2map(out, browse = T)

# Open folder where interactive map
#  .html file is saved, so you can share it:
x = ejam2map(out)
fname = map2browser(x)
# browseURL(dirname(fname)) # to open the temp folder
# file.copy(fname, "./map.html") # to copy map file to working directory

out <- testoutput_ejscreenapi_plus_5
mapfastej(out)
}
}
