% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2means.R
\name{ejam2means}
\alias{ejam2means}
\title{ejam2means - quick look at averages, via ejamit() results}
\usage{
ejam2means(ejamitout, vars = names_these)
}
\arguments{
\item{ejamitout}{as from ejamit()}

\item{vars}{all or some of colnames in ejamitout$results_overall}
}
\value{
means in a useful format
}
\description{
ejam2means - quick look at averages, via ejamit() results
}
\examples{
out <- testoutput_ejamit_100pts_1miles
ejam2means(out, vars = names_e_ratio_to_state_avg)

#' # these should tell you the same thing:
out$results_summarized$keystats[
  rownames(out$results_summarized$keystats) \%in\% names_e_ratio_to_state_avg, 
]
ejam2means(out, vars = names_e_ratio_to_state_avg)

}
