% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2means.R
\name{ejam2quantiles}
\alias{ejam2quantiles}
\title{DRAFT - get quantiles of indicator values among all people analyzed
See \code{ejamit()} (or \code{batch.summarize()}) instead of this, for percentiles!}
\usage{
ejam2quantiles(
  ejamitout,
  vars = names_e_state_pctile,
  w = NULL,
  probs = c(50, 90)/100,
  topx_insteadof_pctile = TRUE
)
}
\arguments{
\item{ejamitout}{as from ejamit()}

\item{vars}{colnames in ejamitout$results_bybg_people}

\item{w}{weights}

\item{probs}{as written is always the pctile, but default is to report on the top X\%, or 100-probs}

\item{topx_insteadof_pctile}{logical}
}
\value{
info about quantiles of places or people
}
\description{
DRAFT - get quantiles of indicator values among all people analyzed
See \code{ejamit()} (or \code{batch.summarize()}) instead of this, for percentiles!
}
\details{
DRAFT -- CANNOT YET ACTUALLY PROVIDE QUANTILES ACROSS UNIQUE PEOPLE if using results_bysite
AND DIFFERENT WEIGHTS WOULD BE NEEDED FOR DIFFERENT INDICATORS if using results_bybg_people
to avoid overlaps between sites to get unique people.

See \code{ejamit()} (or \code{batch.summarize()}) instead of this, for percentiles!
}
\examples{

# pctiles
  ejam2quantiles(out,  vars = names_e_state_pctile, topx_insteadof_pctile = F) 
# ratios  
  ejam2quantiles(out, vars = names_e_ratio_to_state_avg,  topx_insteadof_pctile = T)
# see the quantiles and mean together
 t(rbind(
  AvgPerson = ejam2means(out, vars = names_e_ratio_to_state_avg),
  ejam2quantiles(
     out, 
     vars = names_e_ratio_to_state_avg, 
     probs = 0.5, topx_insteadof_pctile = T
  ) 
  ))

}
\keyword{internal}
