% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2table_tall.R
\name{ejam2table_tall}
\alias{ejam2table_tall}
\title{Simple quick look at results of ejamit() in RStudio console}
\usage{
ejam2table_tall(ejamitout, sitenumber)
}
\arguments{
\item{ejamitout}{like from ejamit() or doaggregate()}

\item{sitenumber}{if omitted, shows results_overall, and
if an integer, identifies which site (which row) to show from results_bysite}
}
\value{
data.frame with one indicator per row
}
\description{
Simple quick look at results of ejamit() in RStudio console
}
