% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2tableviewer.R
\name{ejam2tableviewer}
\alias{ejam2tableviewer}
\title{See ejamit()$results_bysite in interactive table in RStudio viewer pane}
\usage{
ejam2tableviewer(
  out,
  fname = "automatic",
  maxrows = 1000,
  launch_browser = TRUE
)
}
\arguments{
\item{out}{output of ejamit(), or one table like ejamit()$results_overall,
or subset like ejamit()$results_bysite\link{7,}}

\item{fname}{optional. path and filename of the html file to save the table to,
or it uses tempdir() if not specified. Set it to NULL to prevent saving a file.}

\item{maxrows}{only load/ try to show this many rows max.}

\item{launch_browser}{set TRUE to have it launch browser and show report.
Ignored if not interactive() or if fname is set to NULL.}
}
\value{
a datatable object using \code{\link[DT:datatable]{DT::datatable()}}
that can be printed to the console or shown in the RStudio viewer pane
}
\description{
See ejamit()$results_bysite in interactive table in RStudio viewer pane
}
\examples{
ejam2tableviewer(testoutput_ejamit_10pts_1miles)
  
}
