% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejamit_compare_distances.R
\name{ejamit_compare_distances}
\alias{ejamit_compare_distances}
\title{Compare EJAM results overall for more than one radius
Run ejamit() once per radius, get a summary table with a row per radius}
\usage{
ejamit_compare_distances(
  sitepoints,
  radii = c(1, 2, 3),
  donuts_not_cumulative = FALSE,
  quiet = TRUE,
  silentinteractive = TRUE,
  plot = TRUE,
  myvars = names_d_subgroups_ratio_to_state_avg,
  ylab = "Ratio of Avg. within X miles to Avg. Statewide or Nationwide",
  ylim = c(0, 5),
  n = 1,
  ...
)
}
\arguments{
\item{sitepoints}{like for \code{\link[=ejamit]{ejamit()}}}

\item{radii}{optional, vector of radius values like 1:3 for \code{\link[=ejamit]{ejamit()}}}

\item{donuts_not_cumulative}{optional, when implemented, if set TRUE,
would return results on areas in each "donut" or ring that is a distance bin,
such as for

0 < R <= radii\link{1}

radii\link{1} < R <= radii\link{2}

etc.}

\item{quiet}{optional, passed to \code{\link[=ejamit]{ejamit()}}}

\item{silentinteractive}{optional, passed to \code{\link[=ejamit]{ejamit()}}}

\item{plot}{optional logical, set FALSE to avoid plotting}

\item{myvars}{optional, for plot, see default value}

\item{ylab}{optional, for plot, see default value}

\item{ylim}{optional, for plot, see default value}

\item{n}{optional, how many of the indicators to report on (printed to console),
when reporting which indicators most strongly increase as radius decreases.}

\item{...}{optional, passed to \code{\link[=ejamit]{ejamit()}}}
}
\value{
data.table you can call results_bydistance,
like ejamit()$results_overall but with one row per radius
}
\description{
Compare EJAM results overall for more than one radius
Run ejamit() once per radius, get a summary table with a row per radius
}
\examples{
  radii <- c(convert_units(5,"km","miles"), convert_units(50,"km","miles"))
  radii <- 1:10
  radii <- c(1, 10)
  pts <- testpoints_100
  pts <- testpoints_10
  
  bydist <- ejamit_compare_distances(pts, radii = radii)
  ejamit_compare_distances2plot(bydist, myvars = c(
    "ratio.to.avg.pctlowinc", "ratio.to.avg.pcthisp", "ratio.to.avg.pctnhba"))
  
  names(bydist) <- fixcolnames(names(bydist), "r", "shortlabel")
  
  
}
\seealso{
\code{\link[=plot_distance_by_pctd]{plot_distance_by_pctd()}}, \code{\link[=distance_by_group]{distance_by_group()}},
and \code{\link[=ejamit_compare_distances_fulloutput]{ejamit_compare_distances_fulloutput()}}
}
