% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejamit_compare_types_of_places.R
\name{ejamit_compare_types_of_places}
\alias{ejamit_compare_types_of_places}
\title{Compare subsets (types) of places that are all from one list}
\usage{
ejamit_compare_types_of_places(
  sitepoints,
  typeofsite = NULL,
  shapefile = NULL,
  fips = NULL,
  silentinteractive = TRUE,
  ...
)
}
\arguments{
\item{sitepoints}{see \code{\link[=ejamit]{ejamit()}}}

\item{typeofsite}{vector of length same as NROW(sitepoints), where
each unique value defines a group of sites}

\item{shapefile}{see \code{\link[=ejamit]{ejamit()}}}

\item{fips}{see \code{\link[=ejamit]{ejamit()}}}

\item{silentinteractive}{passed to \code{\link[=ejamit]{ejamit()}}}

\item{...}{see \code{\link[=ejamit]{ejamit()}}}
}
\value{
similar to ejamit output but results_overall has one row per unique typeofsite
}
\description{
*** DRAFT - May change but works as currently drafted.
e.g., change output formats of results_bytype vs results_overall
}
\examples{
  out <- ejamit_compare_types_of_places(testpoints_10[1:4, ], 
    typeofsite = c("A", "B", "B", "C"))
  cbind(Rows_or_length = sapply(out, NROW))
  
  ejam2barplot_sitegroups(out, names_these_ratio_to_avg[1], topn = 3)
  
  ejam2barplot_sitegroups(out, "sitecount_unique", topn=3, sortby = F)
  
  ejam2barplot_sitegroups(out, "pop", topn = 3, sortby = F)
  
  # use calculated variable not in original table
  df <- out$results_bytype
  df$share <- df$pop / sum(df$pop)
  df$pop_per_site <- df$pop / df$sitecount_unique
  
  plot_barplot_sites(df,
    "share", ylab = "Share of Total Population",
    topn = 3, names.arg = out$types , sortby = F)
    
  plot_barplot_sites(df,
    "pop_per_site", ylab = "Pop. at Avg. Site in Group",
    topn = 3, main = "Nearby Residents per Site, by Site Type",
    names.arg = out$types , sortby = F)
  
  \dontrun{
    
  # Analyze by EPA Region
  
  pts <- data.frame(testpoints_1000)
  
  # Get State and EPA Region of each point from lat/lon
  
   x <- state_from_latlon(lat = pts$lat, lon = pts$lon)
   pts <- data.frame(pts, x)
   
   out_byregion <- ejamit_compare_types_of_places(
     pts, typeofsite = pts$REGION)
   
   dvarname <- names_d[3]
   ejam2barplot_sitegroups(out_byregion, dvarname)
   abline(h = usastats_means(dvarname))
   
   ejam2barplot_sitegroups(out_byregion, "ratio.to.avg.pctmin",
      main = "By EPA Region", ylim = c(0, 2))
   abline(h = 1)
     
   # Analyze by State (slow)
   
   out_bystate <- ejamit_compare_types_of_places(pts, typeofsite = pts$ST)
   
   ejam2barplot_sitegroups(out_bystate, "sitecount_unique", 
     names.arg = out_bystate$types, topn = 52, cex.names = 0.5,
     main = "Sites by State")

  }
  
}
