% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejamit_sitetype_from_.R
\name{ejamit_sitetype_from_input}
\alias{ejamit_sitetype_from_input}
\title{helper to infer what type of sites were analyzed by looking at params given as INPUT to ejamit()
used by ejamit() and ejamit_compare_types_of_places()}
\usage{
ejamit_sitetype_from_input(sitepoints = NULL, fips = NULL, shapefile = NULL)
}
\arguments{
\item{sitepoints}{parameter as was passed to \code{\link[=ejamit]{ejamit()}}}

\item{fips}{parameter as was passed to \code{\link[=ejamit]{ejamit()}}}

\item{shapefile}{parameter as was passed to \code{\link[=ejamit]{ejamit()}}}
}
\value{
either "latlon", "fips", or "shp",
or errors if 2 or 3 types were specified at once
}
\description{
helper to infer what type of sites were analyzed by looking at params given as INPUT to ejamit()
used by ejamit() and ejamit_compare_types_of_places()
}
\details{
Note \code{sitetype} is not quite the same as the \code{site_method} parameter used in building reports.
\code{site_method} can be one of these: SHP, latlon, FIPS, NAICS, FRS, EPA_PROGRAM, SIC, MACT
\code{sitetype} can be latlon, fips, or shp as returned by \code{\link[=ejamit]{ejamit()}}, but can include
lowercase versions of site_method values too within server code and some function parameters!
}
\keyword{internal}
