% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ejampackages.R
\docType{data}
\name{ejampackages}
\alias{ejampackages}
\title{ejampackages (DATA) list of names of key EJAM-related R packages}
\format{
An object of class \code{character} of length 1.
}
\usage{
ejampackages
}
\description{
ejampackages (DATA) list of names of key EJAM-related R packages
}
\keyword{internal}
