% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen_vs_ejam_.R
\name{ejscreen_vs_ejam_1var}
\alias{ejscreen_vs_ejam_1var}
\title{EJAM/EJSCREEN comparisons - see results at every site, for 1 INDICATOR after using ejscreen_vs_ejam()}
\usage{
ejscreen_vs_ejam_1var(
  vs,
  varname = "pop",
  info = c("EJSCREEN", "EJAM", "ratio", "diff", "absdiff", "pctdiff", "abspctdiff")[c(1,
    2, 5, 6)]
)
}
\arguments{
\item{vs}{output of ejscreen_vs_ejam() or ejscreen_vs()}

\item{varname}{}

\item{info}{}
}
\value{
data frame with colnames like
EJSCREEN, EJAM, EJSCREEN_shown, EJAM_shown, same_shown, ratio,
diff, absdiff, pctdiff, etc.

and rows represent sites analyzed.
}
\description{
EJAM/EJSCREEN comparisons - see results at every site, for 1 INDICATOR after using ejscreen_vs_ejam()
}
\seealso{
\code{\link[=ejscreen_vs_ejam_1var_cdf]{ejscreen_vs_ejam_1var_cdf()}} \code{\link[=ejscreen_vs]{ejscreen_vs()}}
}
\keyword{internal}
