% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen_vs_ejam_.R
\name{ejscreen_vs_ejam_1var_bysite}
\alias{ejscreen_vs_ejam_1var_bysite}
\title{EJAM/EJSCREEN comparisons - see 1 site/row, check estimates for 1 indicator}
\usage{
ejscreen_vs_ejam_1var_bysite(vs, pts, varname = "blockcount_near_site")
}
\arguments{
\item{vs}{results of \code{\link[=ejscreen_vs_ejam]{ejscreen_vs_ejam()}}}

\item{pts}{data.frame with lat, lon that was input to \code{\link[=ejscreen_vs_ejam]{ejscreen_vs_ejam()}}}

\item{varname}{default is "blockcount_near_site" but can be "pop", "Demog.Index", etc.}
}
\value{
table
}
\description{
EJAM/EJSCREEN comparisons - see 1 site/row, check estimates for 1 indicator
}
\seealso{
\code{\link[=ejscreen_vs]{ejscreen_vs()}}
}
\keyword{internal}
