% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen_vs_ejam_.R
\name{ejscreen_vs_ejam_see1}
\alias{ejscreen_vs_ejam_see1}
\title{EJAM/EJSCREEN comparisons - see results for 1 site after using ejscreen_vs_ejam()}
\usage{
ejscreen_vs_ejam_see1(
  vs,
  myvars = c("ejam_uniq_id", "pop", names_d),
  mysite = 1
)
}
\arguments{
\item{vs}{output of ejscreen_vs_ejam() or ejscreen_vs()}

\item{myvars}{optional to check just a subset of the colnames found in vs$EJAM and vs$EJSCREEN,
such as these possible values:

myvars = "all"    # all the indicators in the output tables, i.e., colnames(vs$EJAM)

myvars = "inboth" # just the ones in both (not NA values because EJAM or EJSCREEN did not report the indicator)

myvars = "bad"    # just the ones in both where EJAM_shown and EJSCREEN_shown disagree

myvars = c(names_d, names_d_subgroups) or

myvars = grep("pctile", colnames(vs$EJAM), value = T)}

\item{mysite}{rownumber corresponding to site of interest, among 1:nrow(vs$EJAM)}
}
\value{
a table showing one row per indicator, and columns like EJSCREEN, EJAM, ratio, etc.,
but see str() because it is a list in matrix form
}
\description{
EJAM/EJSCREEN comparisons - see results for 1 site after using ejscreen_vs_ejam()
}
\details{
see also ejscreen_vs_ejam_1var_bysite() and ejscreen_vs_explain()
}
\examples{
  mysite <- 9
  \dontrun{
  vs <- ejscreen_vs_ejam(testpoints_10, radius = 3)
  ejscreen_vs_ejam_see1(vs, mysite = mysite, myvars = colnames(vs$EJAM))[!is.na(vs$EJSCREEN[mysite, ]) , 1:2]
  }

}
\keyword{internal}
