% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen_vs_ejam_.R
\name{ejscreen_vs_ejam_summary_quantiles}
\alias{ejscreen_vs_ejam_summary_quantiles}
\title{EJAM/EJSCREEN comparisons - see quantiles over tested sites, of a stat like ratio of EJAM/EJSCREEN}
\usage{
ejscreen_vs_ejam_summary_quantiles(
  vs,
  mystat = c("ratio", "diff", "absdiff", "pctdiff", "abspctdiff")[1],
  myvars = c("pop", names_these),
  probs = (0:20)/20,
  na.rm = TRUE,
  digits = 4
)
}
\arguments{
\item{vs}{output of ejscreen_vs_ejam()}

\item{mystat}{just one (not more than 1) of these: "ratio", "diff", "absdiff", "pctdiff", "abspctdiff"}

\item{myvars}{optional vector of indicators, to check just a subset of the colnames found in vs$EJAM and vs$EJSCREEN,
such as myvars = c(names_d, names_d_subgroups)
or myvars = grep("pctile", colnames(vs$EJAM), value = T)}

\item{probs}{optional vector of probabilities 0 to 1 to pass to quantile()}

\item{na.rm}{optional, leave it as default TRUE}

\item{digits}{round results to this many digits}
}
\value{
table of percentiles across analyzed locations, of a stat like the ratio of
EJAM estimates over EJSCREEN estimates, for specified list of indicators like in names_d
}
\description{
EJAM/EJSCREEN comparisons - see quantiles over tested sites, of a stat like ratio of EJAM/EJSCREEN
}
\keyword{internal}
