% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejscreen_vs_ejam_.R
\name{ejscreen_vs_explain}
\alias{ejscreen_vs_explain}
\title{EJAM/EJSCREEN comparisons - loop or interactively step through sites to see which blocks may explain difference in population count}
\usage{
ejscreen_vs_explain(
  vs,
  ejam_uniq_id = NULL,
  pause_on_each_site = TRUE,
  onlyifpopdiffers = TRUE,
  map = TRUE
)
}
\arguments{
\item{vs}{The output of \code{\link[=ejscreen_vs_ejam]{ejscreen_vs_ejam()}} or \code{\link[=ejscreen_vs]{ejscreen_vs()}}}

\item{pause_on_each_site}{set FALSE to avoid tapping key to advance to each next one}

\item{onlyifpopdiffers}{set FALSE to not map and show table for the ones where
population counts agreed to zero decimal places}

\item{map}{set to FALSE to suppress drawing map but still return block-level info}
}
\value{
same
}
\description{
EJAM/EJSCREEN comparisons - loop or interactively step through sites to see which blocks may explain difference in population count
}
\seealso{
uses \code{\link[=ejscreen_vs_explain_summary]{ejscreen_vs_explain_summary()}} to show summary
}
\keyword{internal}
