% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part1.R
\name{ejscreenapi_online}
\alias{ejscreenapi_online}
\title{utility - check if EJScreen API seems to be online or offline}
\usage{
ejscreenapi_online(url = "https://ejscreen.epa.gov")
}
\arguments{
\item{url}{DEFAULT is the 2024 base url of the EJScreen API}
}
\value{
TRUE or FALSE (but NA if no internet connection seems to be available at all)
}
\description{
utility - check if EJScreen API seems to be online or offline
}
\keyword{internal}
