% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ejscreenRESTbroker2table_na_filler.R
\docType{data}
\name{ejscreenRESTbroker2table_na_filler}
\alias{ejscreenRESTbroker2table_na_filler}
\title{used to fill in NA values in correctly-formatted output of \code{\link[=ejscreenRESTbroker2table]{ejscreenRESTbroker2table()}} when no data are available}
\format{
An object of class \code{data.frame} with 1 rows and 291 columns.
}
\usage{
ejscreenRESTbroker2table_na_filler
}
\description{
used to fill in NA values in correctly-formatted output of \code{\link[=ejscreenRESTbroker2table]{ejscreenRESTbroker2table()}} when no data are available
}
\keyword{datasets}
