% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_epa_programs_defined.R
\docType{data}
\name{epa_programs_defined}
\alias{epa_programs_defined}
\title{Full names and definitions for acronyms of EPA programs in Facility Registry Services (FRS)}
\format{
An object of class \code{data.frame} with 99 rows and 3 columns.
}
\usage{
epa_programs_defined
}
\description{
Full names and definitions for acronyms of EPA programs in Facility Registry Services (FRS)
}
\seealso{
\link{epa_programs}
}
\keyword{datasets}
