% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{fill_tbl_full_ej}
\alias{fill_tbl_full_ej}
\title{Create full EJ + EJ supp index HTML table of indicator rows}
\usage{
fill_tbl_full_ej(output_df)
}
\arguments{
\item{output_df, }{single row of results table from doaggregate - either results_overall or one row of bysite}
}
\description{
Create full EJ + EJ supp index HTML table of indicator rows
}
\keyword{internal}
