% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{fill_tbl_full_subgroups}
\alias{fill_tbl_full_subgroups}
\title{Create full demog subgroup/ language/ health/ community/ etc. HTML table of indicator rows}
\usage{
fill_tbl_full_subgroups(
  output_df,
  extratable_title = "",
  extratable_title_top_row = "ADDITIONAL INFORMATION",
  list_of_sections = NULL,
  extratable_show_ratios_in_report = TRUE,
  hide_missing_rows_for = names_d_language
)
}
\arguments{
\item{output_df}{single row of results table from doaggregate or possibly ejamit(),
either $results_overall or one row of $results_bysite, where colnames are indicators like pop, pctpoor, etc.}

\item{extratable_title}{Text of overall title text for report table, above the actual table not in the table}

\item{extratable_title_top_row}{title text inside the top left cell of the table, in the header row}

\item{list_of_sections}{named list of vectors,
where each name is text phrase that is title of a section of the table,
and each vector is the vector of colnames of output_df that are indicators to show in that section.}

\item{extratable_show_ratios_in_report}{logical, whether to add columns with ratios to US and State overall values}

\item{hide_missing_rows_for}{only for the indicators named in this vector,
leave out rows in table where raw value is NA, as with many of names_d_language}
}
\description{
Create full demog subgroup/ language/ health/ community/ etc. HTML table of indicator rows
}
\seealso{
used by \code{\link[=build_community_report]{build_community_report()}}
}
\keyword{internal}
