% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{fill_tbl_row_subgroups}
\alias{fill_tbl_row_subgroups}
\title{Write a demographic subgroup indicator to an html table row}
\usage{
fill_tbl_row_subgroups(
  output_df,
  Rname,
  longname,
  extratable_show_ratios_in_report
)
}
\arguments{
\item{output_df, }{single row of results table from doaggregate - either results_overall or one row of bysite}

\item{Rname, }{variable name of indicator to pull from results, such as 'pm', 'pctlowinc', 'Demog.Index'}

\item{longname, }{nicer name of indicator to use in table row; can include HTML sub/superscripts}

\item{extratable_show_ratios_in_report}{logical, whether to add columns with ratios to US and State overall values}
}
\description{
Write a demographic subgroup indicator to an html table row
}
\keyword{Write}
\keyword{a}
\keyword{an}
\keyword{demographic}
\keyword{html}
\keyword{indicator}
\keyword{internal}
\keyword{row}
\keyword{subgroup}
\keyword{table}
\keyword{to}
