% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips2name}
\alias{fips2name}
\title{FIPS - Get county or state names from county or state FIPS codes}
\usage{
fips2name(fips, ...)
}
\arguments{
\item{fips}{vector of US Census FIPS codes for
\itemize{
\item States (2 digits once any required leading zeroes are included)
\item Counties (5)
\item City/town/CDP (7)
\item Tracts (11)
\item Blockgroups (12)
Can be string or numeric, with or without leading zeroes.
}}

\item{...}{passed to fips2countyname() to control whether it appends something like , NY or , New York
after county name}
}
\value{
vector of state and/or county names,
where county names optionally have comma and 2-character abbreviation or full state name.
}
\description{
FIPS - Get county or state names from county or state FIPS codes
}
\examples{
  fips2name(fips_counties_from_state_abbrev(c("AK", "LA"))  )
  fips2name(c(22, 02013))  # can have mix where some are a whole state and others are a county.

}
\seealso{
\code{\link[=countyname2fips]{countyname2fips()}}
}
