% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips2pop}
\alias{fips2pop}
\title{Get population counts (ACS EJScreen) by FIPS
Utility to aggregate just population count for each FIPS Census unit}
\usage{
fips2pop(fips)
}
\arguments{
\item{fips}{vector of fips (can be state, county, tract, blockgroup, block).
If block, it estimates using weights like it does when aggregating for a report.
If city/cdp, it returns NA currently since those pop counts are not in blockgroupstats.}
}
\value{
vector of population counts same length as fips vector
}
\description{
Get population counts (ACS EJScreen) by FIPS
Utility to aggregate just population count for each FIPS Census unit
}
