% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips2state_abbrev}
\alias{fips2state_abbrev}
\title{FIPS - Get state abbreviations from any type of FIPS codes}
\usage{
fips2state_abbrev(fips)
}
\arguments{
\item{fips}{vector of FIPS}
}
\value{
vector of abbreviations like "NY", "LA", "DE", etc.
}
\description{
FIPS - Get state abbreviations from any type of FIPS codes
}
\examples{
  cbind(
    stfips    = fips_lead_zero(1:80),
    ST     = fips2state_abbrev(1:80),
    statename = fips2statename(1:80)
  )

}
