% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips2state_fips}
\alias{fips2state_fips}
\title{FIPS - Get state FIPS codes from any type of FIPS codes}
\usage{
fips2state_fips(fips)
}
\arguments{
\item{fips}{vector of FIPS}
}
\value{
vector of State FIPS 2 characters each
}
\description{
FIPS - Get state FIPS codes from any type of FIPS codes
}
\details{
Tells you which State contains each County (or tract or blockgroup or block)
}
\examples{
  fips2state_fips(fips_counties_from_statename(c("Delaware", "Rhode Island")))

}
