% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips2statename}
\alias{fips2statename}
\title{FIPS - Get state names from any type of FIPS codes}
\usage{
fips2statename(fips)
}
\arguments{
\item{fips}{vector of FIPS}
}
\value{
vector of state names
}
\description{
FIPS - Get state names from any type of FIPS codes
}
\examples{
  cbind(fips_lead_zero(1:80), fips2state_abbrev(1:80), fips2statename(1:80))

}
