% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips_counties_from_countynamefull}
\alias{fips_counties_from_countynamefull}
\title{helper function - get county FIPS from exact countyname including, ST abbrev
used by fips_counties_from_countyname()}
\usage{
fips_counties_from_countynamefull(fullname, exact = TRUE)
}
\arguments{
\item{fullname}{exact (case-insensitive) name of
county comma state abbreviation,
like "Johnson County, TX". Ignores case.}

\item{exact}{set to FALSE to use grep, but that can return more than one per input}
}
\value{
the county FIPS (5 digits long with leading zero if needed, as character)
of each, or NA for non matches
}
\description{
helper function - get county FIPS from exact countyname including, ST abbrev
used by fips_counties_from_countyname()
}
\keyword{internal}
