% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips_counties_from_state_abbrev}
\alias{fips_counties_from_state_abbrev}
\title{FIPS - Get ALL county fips in specified states}
\usage{
fips_counties_from_state_abbrev(ST)
}
\arguments{
\item{ST}{vector of state abbreviations like c("NY","GA"), ignoring case}
}
\value{
vector of 5-digit character string county FIPS of all unique counties in those states
}
\description{
FIPS - Get ALL county fips in specified states
}
\examples{
  fips_counties_from_state_abbrev("DE")
  fips_counties_from_state_abbrev("RI", "RI")

}
