% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips_st2eparegion}
\alias{fips_st2eparegion}
\title{FIPS - Get EPA Region number from state FIPS code}
\usage{
fips_st2eparegion(stfips)
}
\arguments{
\item{stfips}{vector of one or more state fips codes (numbers or as strings)}
}
\value{
vector of numbers representing US EPA Regions
}
\description{
FIPS - Get EPA Region number from state FIPS code
}
