% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips_state_from_state_abbrev}
\alias{fips_state_from_state_abbrev}
\title{FIPS - Get state fips for each state abbrev}
\usage{
fips_state_from_state_abbrev(ST)
}
\arguments{
\item{ST}{vector of state abbreviations like c("NY","GA"), ignores case.
Converts any statename to ST in case names were provided instead of ST.}
}
\value{
vector of 2-digit state FIPS codes like c("10", "44", "44"),
same length as input, so including any duplicates
}
\description{
FIPS - Get state fips for each state abbrev
}
\examples{
fips_state_from_state_abbrev(c("DE", "DE", "RI", 'new jersey'))

}
