% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips_state_from_statename}
\alias{fips_state_from_statename}
\title{FIPS - Get state fips for each state name}
\usage{
fips_state_from_statename(statename)
}
\arguments{
\item{statename}{vector of state names like c("New York","Georgia"),
ignoring case.
Converts any ST to statename in case abbreviations were provided instead of name.}
}
\value{
vector of 2-digit state FIPS codes like c("10", "44", "44"),
same length as input, so including any duplicates
}
\description{
FIPS - Get state fips for each state name
}
\examples{
  fips_state_from_statename("Delaware")
  fips_state_from_statename(c("dc", 'district of columbia', 'georgia'))
}
