% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips_states_in_eparegion}
\alias{fips_states_in_eparegion}
\title{FIPS - Get state fips for all States in EPA Region(s)}
\usage{
fips_states_in_eparegion(region)
}
\arguments{
\item{region}{vector of 1 or more EPA Regions (numbers 1 through 10)}
}
\value{
vector of 2-digit state FIPS codes like c("10", "44", "44"),
same length as input, so including any duplicates
}
\description{
FIPS - Get state fips for all States in EPA Region(s)
}
\examples{
  fips_states_in_eparegion(2)
  fips_states_in_eparegion(6)
  fips2state_abbrev(fips_states_in_eparegion(6))
}
