% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{fips_valid}
\alias{fips_valid}
\title{check if FIPS code is valid, meaning it is an actual Census FIPS code for a State, County, City/CDP, etc.}
\usage{
fips_valid(fips)
}
\arguments{
\item{fips}{vector of numeric or character fips. works for state, county, city/cdp, tract, blockgroup or block fips.}
}
\value{
logical vector
}
\description{
check if FIPS code is valid, meaning it is an actual Census FIPS code for a State, County, City/CDP, etc.
}
