% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calc_ejam.R
\name{formula_varname}
\alias{formula_varname}
\title{utility to check formulas and extract variable names they calculate values for}
\usage{
formula_varname(myforms)
}
\arguments{
\item{myforms}{see calc_byformula()}
}
\description{
utility to check formulas and extract variable names they calculate values for
}
