% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formulas_all.R
\docType{data}
\name{formulas_all}
\alias{formulas_all}
\title{formulas_all (DATA) table of formulas to aggregate or calculate indicators}
\format{
An object of class \code{character} of length 186.
}
\usage{
formulas_all
}
\description{
These formulas can describe how each indicator is calculated from
other variables like counts or how it is aggregated as a weighted mean, etc.
}
\details{
Created for EJAM by datacreate_formulas.R script

Can be used by \code{\link[=calc_ejam]{calc_ejam()}} for aggregation or to create a derived custom
indicator for all US blockgroups based on counts obtained from the ACS.
}
\keyword{datasets}
