% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_clean_sic.R
\name{frs_clean_sic}
\alias{frs_clean_sic}
\title{Clean EPA FRS SIC info}
\usage{
frs_clean_sic(
  frs,
  usefulcolumns = c("lat", "lon", "SIC", "LATITUDE83", "LONGITUDE83", "REGISTRY_ID",
    "PRIMARY_NAME", "SIC_CODES", "PGM_SYS_ACRNMS")
)
}
\arguments{
\item{frs}{the frs data object from frs_read()}

\item{usefulcolumns}{optional}
}
\value{
frs data.table
}
\description{
Clean EPA FRS SIC info
}
\details{
works just like \code{\link[=frs_clean]{frs_clean()}}
but for SIC codes instead of NAICS
}
\seealso{
\code{\link[=frs_update_datasets]{frs_update_datasets()}}
}
