% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_download.R
\name{frs_download}
\alias{frs_download}
\title{Download Facility Registry Service (FRS) file national_single.zip}
\usage{
frs_download(
  folder = NULL,
  zfile = "national_single.zip",
  zipbaseurl = "https://ordsext.epa.gov/FLA/www3/state_files/"
)
}
\arguments{
\item{folder}{path Default is NULL which means it is downloaded to a temp folder.}

\item{zfile}{filename}

\item{zipbaseurl}{url}
}
\value{
The full path and file name of the downloaded zip file
}
\description{
Download Facility Registry Service (FRS) file national_single.zip
}
\details{
See \url{https://www.epa.gov/frs/epa-frs-facilities-state-single-file-csv-download}

and \url{https://echo.epa.gov/tools/data-downloads/frs-download-summary}
}
\seealso{
Used by \code{\link[=frs_update_datasets]{frs_update_datasets()}} which uses \code{\link[=frs_get]{frs_get()}}
}
