% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_from_xyz.R
\name{frs_from_naics}
\alias{frs_from_naics}
\title{Use NAICS code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites}
\usage{
frs_from_naics(naics_code_or_name, childrenForNAICS = TRUE, ...)
}
\arguments{
\item{naics_code_or_name}{passed to \code{\link[=naics_from_any]{naics_from_any()}} as the query}

\item{childrenForNAICS}{passed to \code{\link[=naics_from_any]{naics_from_any()}} as the children param of that function}

\item{...}{passed to \code{\link[=naics_from_any]{naics_from_any()}}}
}
\value{
relevant rows of the data.table called frs, which has column names that are
"lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "PGM_SYS_ACRNMS"
}
\description{
Use NAICS code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites
}
\details{
The EPA also provides a \href{https://www.epa.gov/frs/frs-query#industrial}{FRS Facility Industrial Classification Search tool}
where you can find facilities based on NAICS or SIC.

EPA's \href{https://echo.epa.gov/help/facility-search/search-criteria-help#facchar}{ECHO query tools}
also provide search by NAICS or SIC, and by MACT subpart.
}
\examples{
  frs_from_naics("uranium")
  mapfast(frs_from_naics(naics_from_any("nuclear")$code))
  naics_from_any("silver")
  naics_from_name("silver")
  naics_from_any(212222 )
  frs_from_naics(21222)
  regid_from_naics(21222)
  latlon_from_naics(21222)
  
}
\seealso{
\code{\link[=latlon_from_naics]{latlon_from_naics()}} \code{\link[=latlon_from_sic]{latlon_from_sic()}} \code{\link[=frs_from_sic]{frs_from_sic()}} \code{\link[=regid_from_naics]{regid_from_naics()}} \code{\link[=naics_from_any]{naics_from_any()}}
}
