% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_from_xyz.R
\name{frs_from_program}
\alias{frs_from_program}
\title{Use EPA Program acronym like TRIS to see FRS Facility Registry Service data on those EPA-regulated sites}
\usage{
frs_from_program(program)
}
\arguments{
\item{program}{vector of one or more EPA Program names used by FRS}
}
\value{
relevant rows of the data.table called frs, which has column names that are
"lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "PGM_SYS_ACRNMS"
}
\description{
Get data.table based on given FRS Program System CATEGORY.
Find all FRS sites in a program like RCRAINFO, TRIS, or others.
}
\details{
Also see \href{https://www.epa.gov/frs/frs-data-sources}{EPA documentation describing each program code} aka data source.
}
