% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_from_xyz.R
\name{frs_from_regid}
\alias{frs_from_regid}
\title{Use registry ID to see FRS Facility Registry Service data on those EPA-regulated sites}
\usage{
frs_from_regid(regid)
}
\arguments{
\item{regid}{vector of one or more EPA Registry ID codes used by FRS}
}
\value{
relevant rows of the data.table called frs, which has column names that are
"lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "PGM_SYS_ACRNMS"
}
\description{
Use registry ID to see FRS Facility Registry Service data on those EPA-regulated sites
}
\examples{
  frs_from_regid(testids_registry_id)
  frs_from_regid(110000307695)
  frs_from_regid("110000307695")

}
