% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_xyz.R
\name{frs_from_sic}
\alias{frs_from_sic}
\title{Use SIC code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites}
\usage{
frs_from_sic(sic_code_or_name, ...)
}
\arguments{
\item{sic_code_or_name}{passed to \code{\link[=sic_from_any]{sic_from_any()}}}

\item{...}{passed to \code{\link[=sic_from_any]{sic_from_any()}}}
}
\value{
relevant rows of the data.table called frs, which has column names that are
"lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "SIC" "PGM_SYS_ACRNMS"

The EPA also provides a \href{https://www.epa.gov/frs/frs-query#industrial}{FRS Facility Industrial Classification Search tool}
where you can find facilities based on NAICS or SIC.
}
\description{
Use SIC code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites
}
\examples{
  frs_from_sic("glass")
  mapfast(frs_from_sic(sic_from_any("silver")$code))
  sic_from_any("silver")
  sic_from_name("silver")
  sic_from_any('0780')
  frs_from_sic('0780')
  regid_from_sic('0780')
  latlon_from_sic('0780')

}
\seealso{
\code{\link[=regid_from_sic]{regid_from_sic()}} \code{\link[=sic_from_any]{sic_from_any()}} \code{\link[=latlon_from_sic]{latlon_from_sic()}}
}
