% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_from_xyz.R
\name{frs_from_sitename}
\alias{frs_from_sitename}
\title{Use site name text search to see FRS Facility Registry Service data on those EPA-regulated sites}
\usage{
frs_from_sitename(sitenames, ignore.case = TRUE, fixed = FALSE)
}
\arguments{
\item{sitenames}{one or more strings in a vector, which can be regular expressions or query for exact match using fixed=TRUE}

\item{ignore.case}{logical, search is not case sensitive by default (unlike \code{\link[=grepl]{grepl()}} default)}

\item{fixed}{see \code{\link[=grepl]{grepl()}}, if set to TRUE it looks for only exact matches}
}
\value{
relevant rows of the data.table called frs, which has column names that are
"lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "PGM_SYS_ACRNMS"
}
\description{
VERY SLOW search within PRIMARY_NAME of facilities for matching text
}
\examples{
\dontrun{
 # very slow
 x=frs_from_sitename
 nrow(x)
 head(x)
}
  
}
