% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_active_ids.R
\name{frs_inactive_ids}
\alias{frs_inactive_ids}
\title{Download national file and find which IDs are the INACTIVE sites in the FRS}
\usage{
frs_inactive_ids(active = FALSE, ...)
}
\arguments{
\item{active}{If FALSE, default, returns the registry IDs of sites that seem to be inactive,
based on closecodes.}

\item{...}{passed to frs_active_ids()}
}
\value{
vector of FRS IDs that are the clearly inactive sites -- or all other sites --
depending on value of active
}
\description{
Download national file and find which IDs are the INACTIVE sites in the FRS
}
\seealso{
\code{\link[=frs_active_ids]{frs_active_ids()}} \code{\link[=frs_update_datasets]{frs_update_datasets()}}
}
