% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_is_valid.R
\name{frs_is_valid}
\alias{frs_is_valid}
\title{Validate FRS Registry ID table uploaded (just checks colname, mostly)}
\usage{
frs_is_valid(frs_upload)
}
\arguments{
\item{frs_upload}{upload frs registry IDs table converted to data frame (or data.table gets handled too)
with those ids in a column whose name is among allowed aliases that get tried here:
the colname with the FRS regids must be one of REGISTRY_ID, RegistryID, regid, siteid,
checked in that order of preference.}
}
\value{
boolean value (valid or not valid)
}
\description{
Check for proper colname (or what seems to be a valid alias)
}
\details{
note it checks aliases (REGISTRY_ID, RegistryID, regid, siteid) in that order
and once a valid name is found then even if it fails to actually
contain valid ids, the func does not go back and try the rest of the possible aliases,
so if the two cols were regid and siteid and only siteid had valid registry ID values,
this func would fail to figure that out and would say they were invalid.
}
\keyword{internal}
