% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_make_naics_lookup.R
\name{frs_make_naics_lookup}
\alias{frs_make_naics_lookup}
\title{Reformat frs datatable to look up by NAICS}
\usage{
frs_make_naics_lookup(x)
}
\arguments{
\item{x}{data.table frs from \code{\link[=frs_get]{frs_get()}}}
}
\value{
data.table with columns NAICS, REGISTRY_ID, etc.
}
\description{
Reformat frs datatable to look up by NAICS
}
\seealso{
\code{\link[=frs_update_datasets]{frs_update_datasets()}} which uses \code{\link[=frs_get]{frs_get()}} to create frs_by_naics
}
