% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_make_sic_lookup.R
\name{frs_make_sic_lookup}
\alias{frs_make_sic_lookup}
\title{Reformat frs datatable to look up facilities by SIC code}
\usage{
frs_make_sic_lookup(x)
}
\arguments{
\item{x}{data.table frs from \code{\link[=frs_clean_sic]{frs_clean_sic()}}}
}
\value{
data.table with lat, lon, REGISTRY_ID, SIC columns
}
\description{
Reformat frs datatable to look up facilities by SIC code
}
\seealso{
\code{\link[=frs_upate_datasets]{frs_upate_datasets()}} \code{\link[=frs_clean_sic]{frs_clean_sic()}} and the frs_by_sic data.table
}
