% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_unzip.R
\name{frs_unzip}
\alias{frs_unzip}
\title{Unzip Facility Registry Service dataset}
\usage{
frs_unzip(zfile = "national_single.zip", folder = ".", ...)
}
\arguments{
\item{zfile}{zipfile obtained via frs_download}

\item{folder}{optional, folder to look in for zfile}

\item{...}{passed to unzip}
}
\description{
Unzip Facility Registry Service dataset
}
\details{
helper function used by frs_get() to create dataset for EJAM
}
\seealso{
\code{\link[=frs_upate_datasets]{frs_upate_datasets()}} which uses \code{\link[=frs_get]{frs_get()}} which uses \code{\link[=frs_download]{frs_download()}} \code{\link[=frs_unzip]{frs_unzip()}} \code{\link[=frs_read]{frs_read()}} \code{\link[=frs_clean]{frs_clean()}}
}
