% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frsprogramcodes.R
\docType{data}
\name{frsprogramcodes}
\alias{frsprogramcodes}
\title{frsprogramcodes DATA EPA programs listed in Facility Registry Service}
\format{
An object of class \code{data.frame} with 13 rows and 2 columns.
}
\usage{
frsprogramcodes
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{data.frame

                                                              description     code
                                                              
1    National Pollutant Discharge Elimination System (NPDES) (ICIS-NPDES)    NPDES

2  The Integrated Compliance Information System (ICIS) for Air (ICIS-Air)      AIR

3    The Resource Conservation and Recovery Act (RCRA) Information System RCRAINFO

4                                   Risk Management Plan (RMP) facilities      RMP

5                      The Safe Drinking Water Information System (SDWIS)     SFDW

6                              The Superfund Enterprise Management System     SEMS

7                              Clean Air Markets Division Business System   CAMDBS

8                                        Toxics Release Inventory Program     TRIS

9                                        Greenhouse Gas Reporting Program   E-GGRT

10                                             Emissions Inventory System      EIS

11                                           Toxic Substances Control Act     TSCA
}\if{html}{\out{</div>}}
}
\details{
Created by script in /data-raw/

Also see \href{https://www.epa.gov/frs/frs-data-sources}{EPA documentation describing each program code} aka data source.
}
\examples{
\dontrun{
  frs_by_programid[program \%in\% frsprogramcodes$code, .N, by=program]
  
  setkey(frs_by_programid,"program")
  frs_by_programid["TRIS",]
  }
}
\seealso{
\link{frs}
}
\keyword{datasets}
