% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames_etc.R
\name{functions_in_pkg}
\alias{functions_in_pkg}
\title{utility to see which objects in a loaded/attached package are exported functions, internal (unexported) objects, or datasets}
\usage{
functions_in_pkg(
  pkg,
  alphasort_table = FALSE,
  internal_included = TRUE,
  exportedfuncs_included = TRUE,
  data_included = TRUE,
  vectoronly = FALSE
)
}
\arguments{
\item{pkg}{name of package as character like "EJAM"}

\item{alphasort_table}{default is FALSE, to show internal first as a group, then exported funcs, then datasets}

\item{internal_included}{default TRUE includes internal (unexported) objects in the list}

\item{exportedfuncs_included}{default TRUE includes exported functions (non-datasets, actually) in the list}

\item{data_included}{default TRUE includes datasets in the list, as would be seen via data(package=pkg)}

\item{vectoronly}{set to TRUE to just get a character vector of object names instead of the data.frame table output}
}
\value{
data.table with colnames object, exported, data  where exported and data are 1 or 0 for T/F,
unless vectoronly = TRUE in which case it returns a character vector
}
\description{
utility to see which objects in a loaded/attached package are exported functions, internal (unexported) objects, or datasets
}
\details{
See \code{\link[=dupeRfiles]{dupeRfiles()}} for files supporting a shiny app that is not a package, e.g.

See \code{\link[=dupenames]{dupenames()}} for objects that are in R packages.

See \code{\link[=functions_in_pkg]{functions_in_pkg()}}

See \code{\link[=datapack]{datapack()}}
}
\examples{
 # functions_in_pkg("datasets")

}
\seealso{
\code{\link[=ls]{ls()}} \code{\link[=getNamespace]{getNamespace()}} \code{\link[=getNamespaceExports]{getNamespaceExports()}} \code{\link[=loadedNamespaces]{loadedNamespaces()}}
}
\keyword{internal}
