% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames_etc.R
\name{functions_that_use}
\alias{functions_that_use}
\title{utility for developing package - searches for text in each function exported by pkg (or each .R source file in pkg/R)}
\usage{
functions_that_use(text = "stop\\\\(", pkg = "EJAM", ignore_comments = TRUE)
}
\arguments{
\item{text}{something like "EJAM::" or "stop\\(" or "library\\(" or "***"}

\item{pkg}{name of package or path to source package root folder - this

checks only the exported functions of an installed package,
if pkg = some installed package as character string like "EJAM"

checks each .R source FILE NOT each actual function,
if pkg = root folder of source package with subfolder called R with .R source files}

\item{ignore_comments}{logical,
ignore_comments is ignored and treated as if it were TRUE when pkg = some installed package

ignore_comments is used only if pkg = a folder that contains .R files

Note it will fail to ignore comments in .R files that are at the end of the line of actual code like  print(1) # that prints 1}
}
\value{
vector of names of functions or paths to .R files
}
\description{
utility for developing package - searches for text in each function exported by pkg (or each .R source file in pkg/R)
}
\details{
Searches the body and parameter defaults of exported functions.
}
\keyword{internal}
