% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{generate_demog_header}
\alias{generate_demog_header}
\title{Build header for demog. + envt. tables in community report}
\usage{
generate_demog_header()
}
\description{
Build header for demog. + envt. tables in community report
}
\keyword{internal}
