% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{generate_ej_header}
\alias{generate_ej_header}
\title{Build header for EJ index table in community report}
\usage{
generate_ej_header()
}
\description{
Build header for EJ index table in community report
}
\keyword{internal}
