% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{generate_ej_supp_header}
\alias{generate_ej_supp_header}
\title{Build header for EJ supp indexes in community report
!}
\usage{
generate_ej_supp_header()
}
\description{
Build header for EJ supp indexes in community report
!
}
\keyword{internal}
