% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{generate_env_demog_header}
\alias{generate_env_demog_header}
\title{Build header for demog. + envt. tables in community report}
\usage{
generate_env_demog_header(
  title = "Environmental and Residential Population Indicators"
)
}
\arguments{
\item{title}{title}
}
\description{
Build header for demog. + envt. tables in community report
}
\seealso{
used by \code{\link[=build_community_report]{build_community_report()}}
}
\keyword{internal}
