% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{generate_extra_header}
\alias{generate_extra_header}
\title{Build header for extra table in community report}
\usage{
generate_extra_header(title = "Additional Information")
}
\arguments{
\item{title}{title}
}
\description{
Build header for extra table in community report
}
\seealso{
used by \code{\link[=build_community_report]{build_community_report()}}
}
\keyword{internal}
