% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{generate_html_header}
\alias{generate_html_header}
\title{Build HTML header for community report}
\usage{
generate_html_header(
  analysis_title,
  totalpop,
  locationstr,
  in_shiny = FALSE,
  report_title = NULL,
  logo_path = NULL,
  logo_html = NULL
)
}
\arguments{
\item{analysis_title, }{title to use in header of report}

\item{totalpop, }{total population included in location(s) analyzed}

\item{locationstr, }{description of the location(s) analyzed}

\item{in_shiny, }{whether the function is being called in or outside of shiny - affects location of header}

\item{report_title}{generic name of this type of report, to be shown at top, like "EJAM Multisite Report"}

\item{logo_path}{optional relative path to a logo for the upper right of the overall header.
Ignored if logo_html is specified and not NULL, but otherwise uses default or param set in run_app()}

\item{logo_html}{optional HTML for img of logo for the upper right of the overall header.
If specified, it overrides logo_path. If omitted, gets created based on logo_path.}
}
\description{
Build HTML header for community report
}
\seealso{
used by \code{\link[=build_community_report]{build_community_report()}}
}
\keyword{internal}
