% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_global_defaults_or_user_options.R
\name{get_global_defaults_or_user_options}
\alias{get_global_defaults_or_user_options}
\title{utility that reconciles/consolidates user-defined params passed via run_app() and default settings from global_defaults_ files}
\usage{
get_global_defaults_or_user_options(
  user_specified_options = NULL,
  bookmarking_allowed = "url"
)
}
\arguments{
\item{user_specified_options}{passed as arguments to \code{run_app()}}

\item{bookmarking_allowed}{same as \code{?shinyApp} enableBookmarking param}
}
\value{
a list of global defaults or user options that run_app()
uses as the golem_opts parameter in \code{golem::with_golem_options()}
and that later can be retrieved by server or ui via \code{golem::get_golem_options()}
or via \code{global_or_param()} (which both do almost the same thing).
}
\description{
utility that reconciles/consolidates user-defined params passed via run_app() and default settings from global_defaults_ files
}
\details{
This function, called by \code{run_app()},
collects the shiny-app-related default settings that are defined in these places:
\enumerate{
\item parameters passed to \code{run_app()} by a user
\item settings defined in file \code{global_defaults_shiny_public.R}
\item settings defined in file \code{global_defaults_shiny.R}
\item settings defined in file \code{global_defaults_package.R}
}

and consolidates them all as a list, to be available to server/ui.
}
\keyword{internal}
