% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocks_diagnostics.R
\name{getblocks_predict_blocks_per_site}
\alias{getblocks_predict_blocks_per_site}
\title{utility - estimate how large table might be that will be output of getblocksnearby()}
\usage{
getblocks_predict_blocks_per_site(nsites, radius)
}
\arguments{
\item{nsites}{count of sites}

\item{radius}{radius in miles}
}
\value{
a number
}
\description{
utility - estimate how large table might be that will be output of getblocksnearby()
}
\keyword{internal}
