% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocks_diagnostics.R
\name{getblocks_summarize_blocks_per_site}
\alias{getblocks_summarize_blocks_per_site}
\title{utility - How many blocks are near the sites (pop density affects accuracy)}
\usage{
getblocks_summarize_blocks_per_site(x, varname = "ejam_uniq_id")
}
\arguments{
\item{x}{The output of \code{\link[=getblocksnearby]{getblocksnearby()}} like \link{testoutput_getblocksnearby_1000pts_1miles}}

\item{varname}{colname of variable in data.table x that is the one to summarize by}
}
\value{
invisibly, a list of stats, and plot
}
\description{
Number of blocks near avg site, how many sites have only 1 or fewer than 30 blocks nearby, etc.
}
\seealso{
\code{\link[=getblocks_diagnostics]{getblocks_diagnostics()}}
}
\keyword{internal}
