% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocks_diagnostics.R
\name{getblocks_summarize_sites_per_block}
\alias{getblocks_summarize_sites_per_block}
\title{utility - How many sites are near the blocks (site density near residents)}
\usage{
getblocks_summarize_sites_per_block(x, varname = "blockid")
}
\arguments{
\item{x}{The output of \code{\link[=getblocksnearby]{getblocksnearby()}} like \link{testoutput_getblocksnearby_10pts_1miles}}

\item{varname}{colname of variable in data.table x that is the one to summarize by}
}
\value{
invisibly, a list of stats
}
\description{
utility - How many sites are near the blocks (site density near residents)
}
\seealso{
\code{\link[=getblocks_diagnostics]{getblocks_diagnostics()}}
}
\keyword{internal}
