% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getblocksnearbyviaQuadTree_Clustered-DRAFT.R
\name{getblocksnearbyviaQuadTree_Clustered}
\alias{getblocksnearbyviaQuadTree_Clustered}
\title{DRAFT / WAS WORK IN PROGRESS  - find nearby blocks using Quad Tree data structure for speed, CLUSTERED FOR PARALLEL PROCESSING}
\usage{
getblocksnearbyviaQuadTree_Clustered(
  sitepoints,
  radius,
  maxradius,
  avoidorphans,
  CountCPU = 1,
  quadtree
)
}
\arguments{
\item{sitepoints}{data.table with columns LAT, LONG}

\item{radius}{in miles}

\item{maxradius}{miles distance}

\item{avoidorphans}{logical}

\item{CountCPU}{for parallel processing via makeCluster() and \code{\link[doSNOW:registerDoSNOW]{doSNOW::registerDoSNOW()}}}

\item{quadtree}{index of all US blocks like localtree}
}
\description{
Uses packages parallel and snow. \link[parallel:makeCluster]{parallel::makePSOCKcluster} is an enhanced version of \link[snow:snow-startstop]{snow::makeSOCKcluster} in package snow.
It runs Rscript on the specified host(s) to set up a worker process which listens on a socket for expressions to evaluate, and returns the results (as serialized objects).
}
\details{
For all examples, see \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}}

Uses indexgridsize and quaddata  variables that come from global environment (but should pass to this function rather than assume in global env?)
}
\seealso{
\code{\link[=getblocksnearby]{getblocksnearby()}} \code{\link[=getblocksnearbyviaQuadTree]{getblocksnearbyviaQuadTree()}}
}
\keyword{internal}
