% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfrsnearby.R
\name{getfrsnearby}
\alias{getfrsnearby}
\title{DRAFTING A FUNCTION TO COUNT NEARBY SITES OF A CERTAIN TYPE}
\usage{
getfrsnearby(sitepoints, frspoints, radius = 1, quadtree = NULL)
}
\arguments{
\item{sitepoints}{sitepoints}

\item{frspoints}{frspoints}

\item{radius}{radius}

\item{quadtree}{quadtree}
}
\value{
table
}
\description{
DRAFTING A FUNCTION TO COUNT NEARBY SITES OF A CERTAIN TYPE
}
