% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROXIMITY_FUNCTIONS.R
\name{getpointsnearby}
\alias{getpointsnearby}
\title{Find IDs of and distances to all nearby points (e.g., schools, or EPA-regulated facilities, etc.)}
\usage{
getpointsnearby(
  frompoints,
  topoints,
  radius = 3,
  maxradius = 31.07,
  avoidorphans = FALSE,
  retain_unadjusted_distance = TRUE,
  quadtree = NULL,
  quiet = FALSE,
  parallel = FALSE,
  updateProgress = FALSE,
  report_progress_every_n = 1000,
  ...
)
}
\arguments{
\item{frompoints}{used as the sitepoints param of \code{\link[=getblocksnearby]{getblocksnearby()}}.
Can be for example Census blocks (based on internal point of each block).}

\item{topoints}{table of lat lon coordinates of points that may be nearby.
These could be schools, parks, facilities, or any other set of points.}

\item{radius}{passed to \code{\link[=getpointsnearbyviaQuadTree]{getpointsnearbyviaQuadTree()}}}

\item{maxradius}{passed to \code{\link[=getpointsnearbyviaQuadTree]{getpointsnearbyviaQuadTree()}}}

\item{avoidorphans}{passed to \code{\link[=getpointsnearbyviaQuadTree]{getpointsnearbyviaQuadTree()}}}

\item{quadtree}{optional index of topoints
\itemize{
\item if not provided, created by indexpoints()
}}

\item{quiet}{passed to \code{\link[=getpointsnearbyviaQuadTree]{getpointsnearbyviaQuadTree()}}}

\item{...}{passed to \code{\link[=getpointsnearbyviaQuadTree]{getpointsnearbyviaQuadTree()}}}

\item{quaddatatable}{optional table of topoints (in
format provided by internal helper function create_quaddata() as needed).}
}
\value{
sites2points data.table one row per pair of frompoint and nearby topoint,
like output of \code{\link[=getpointsnearbyviaQuadTree]{getpointsnearbyviaQuadTree()}}
}
\description{
Given a table of frompoints (lat lon coordinates),
find IDs of and distances to all nearby points
that could represent e.g., schools, parks, or
EPA-regulated facilities with locations in Facility Registry Services (FRS).
Like \code{\link[=getblocksnearby]{getblocksnearby()}} but for nearby points of any type.
}
\details{
Later steps can aggregate at each frompoint to summarize
\itemize{
\item count schools or facilities, etc., near each frompoint
\item max/min distance for each frompoint, like proximity of nearest, etc.
\item a proximity score for each frompoint (e.g., block) and then each blockgroup
}
}
