% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_input_names.R
\name{input_names_listing}
\alias{input_names_listing}
\title{Utility checking values of input$ that appear in this code}
\usage{
input_names_listing(file = "./R/app_server.R")
}
\arguments{
\item{file}{path to source file to search in}
}
\value{
character vector of ids of inputs like x,y,z if it found input$x input$y input$z
}
\description{
See appsilon pkg shiny.info now
}
\keyword{internal}
