% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{is.island}
\alias{is.island}
\title{which fips, state names, or state abbreviations are island areas}
\usage{
is.island(ST = NULL, statename = NULL, fips = NULL)
}
\arguments{
\item{ST}{optional vector of 2 letter state abbreviations}

\item{statename}{optional vector of statenames like "texas" or "Delaware"}

\item{fips}{optional vector of FIPS codes (first 2 characters get used)}
}
\value{
logical vector of same length as the input
}
\description{
which fips, state names, or state abbreviations are island areas
}
\examples{
  is.island(c("PR", "DE", "AS", NA))
  is.island(statename = c("Guam", "New York", "american samoa", NA))
  is.island(fips = c(21001, 60, "60", "600010000000"))
  tail(cbind(stateinfo2[ , c("statename", "is.island.areas")], is.island(stateinfo2$ST)),10)
}
\seealso{
\code{\link[=latlon_is.islandareas]{latlon_is.islandareas()}}
}
