% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.numeric.text.R
\name{is.numeric.text}
\alias{is.numeric.text}
\title{DRAFT utility to check if vector is numbers stored as text like "01"}
\usage{
\method{is.numeric}{text}(x, trim_ = FALSE)
}
\arguments{
\item{x}{character vector (or it reports FALSE)}
}
\value{
TRUE or FALSE (not a vector as long as x... just length 1).
Always ignores any or all NA values.
}
\description{
DRAFT utility to check if vector is numbers stored as text like "01"
}
\keyword{internal}
