% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.numericish.R
\name{is.numericish}
\alias{is.numericish}
\title{see which columns seem numeric and could be rounded, e.g. - DRAFT NOT FULLY TESTED}
\usage{
is.numericish(
  x,
  only.if.already.numeric = FALSE,
  strip.characters.before.coerce = FALSE
)
}
\arguments{
\item{x}{data.table, data.frame, or vector}

\item{only.if.already.numeric}{logical, if TRUE, only reports TRUE for a column (or element) if  is.numeric() is TRUE for that one}

\item{strip.characters.before.coerce}{logical, if TRUE, tries to remove spaces and percentage signs before trying to coerce to numeric}
}
\value{
logical vector as long as NCOL(x) i.e., is length(x), if x is table, or length(x) if vector
}
\description{
see which columns seem numeric and could be rounded, e.g. - DRAFT NOT FULLY TESTED
}
\details{
Does not report "08" as numeric-ish, so cannot check fips with leading zero this way, e.g.
To do that you could check if is.na(as.numeric(x))
but "08" is not roundable.
}
\seealso{
\code{\link[=is.numeric.text]{is.numeric.text()}}

\code{\link[=table_round]{table_round()}}
}
\keyword{internal}
