% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_islandareas.R
\docType{data}
\name{islandareas}
\alias{islandareas}
\title{islandareas (DATA) table, bounds info on lat lon of US Island Areas}
\format{
An object of class \code{data.frame} with 8 rows and 5 columns.
}
\usage{
islandareas
}
\description{
data.frame of info on approximate lat lon bounding boxes around
American Samoa, Guam, the
Commonwealth of the Northern Mariana Islands (Northern Mariana Islands),
and the United States Virgin Islands.

See also \link{stateinfo} and \link{stateinfo2}

See \href{http://www.census.gov/geo/reference/gtc/gtc_island.html}{Census documentation}

See source package files datacreate_islandareas.R or EJAM/data-raw/islandareas.xlsx

Note the US minor outlying islands are not in that list and are widely dispersed.
They include Midway Islands, etc.
}
\keyword{datasets}
