% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_vectorofcsvpairs.R
\name{latlon2csv}
\alias{latlon2csv}
\title{helper function - combine lat/lon values into csv format}
\usage{
latlon2csv(lat, lon)
}
\arguments{
\item{lat}{vector of latitudes}

\item{lon}{vector of longitudes}
}
\value{
vector of comma-separated pairs (see example)
}
\description{
Combines a vector of latitudes and a vector of longitudes
into one vector of comma-separated pairs like latitude,longitude
}
\examples{
   lat_example = c(30.01,30.26,30.51)
   lon_example = c(-90.61,-90.95,-91.23)
   latloncsv_example = c("30.01,-90.61", "30.26,-90.95", "30.51,-91.23")
   all.equal(latloncsv_example, 
             latlon2csv(lat = lat_example, lon = lon_example)
   )

}
\keyword{internal}
