% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_mactsubpart.R
\name{latlon_from_mactsubpart}
\alias{latlon_from_mactsubpart}
\title{Get point locations for US EPA-regulated facilities that have
sources subject to Maximum Achievable Control Technology (MACT) standards
under the Clean Air Act.}
\usage{
latlon_from_mactsubpart(subpart = "JJJ", include_if_no_latlon = FALSE)
}
\arguments{
\item{subpart}{vector of one or more strings indicating the
Subpart of CFR Title 40 Part 63 that covers the source category
of interest, such as "FFFF" - see for example,
\url{https://www.ecfr.gov/current/title-40/part-63/subpart-FFFF}}

\item{include_if_no_latlon}{logical - many in the database lack lat lon values but have a MACT code}
}
\value{
a data.table with columns named

programid, subpart, title, lat, lon, REGISTRY_ID, program

for US EPA FRS sites with that MACT code.
Or NA if none found.
}
\description{
Get point locations for US EPA-regulated facilities that have
sources subject to Maximum Achievable Control Technology (MACT) standards
under the Clean Air Act.
}
\details{
See \href{https://www.epa.gov/stationary-sources-air-pollution/national-emission-standards-hazardous-air-pollutants-neshap-9}{EPA information about MACT subparts, NESHAPs, etc.}

EPA's \href{https://echo.epa.gov/help/facility-search/search-criteria-help#facchar}{ECHO query tools}
also provide search by NAICS or SIC, and by MACT subpart.
}
\examples{
  mact_table
  latlon_from_mactsubpart("OOOO", include_if_no_latlon = FALSE) # default
  latlon_from_mactsubpart("OOOO", include_if_no_latlon = TRUE)
  
}
