% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_regid.R
\name{latlon_from_regid}
\alias{latlon_from_regid}
\title{Get lat lon (and NAICS) via Facility Registry ID}
\usage{
latlon_from_regid(regid)
}
\arguments{
\item{regid}{Facility Registry Service ID like 110010052520}
}
\value{
data.table with columns
lat,lon,REGISTRY_ID,PRIMARY_NAME,NAICS,PGM_SYS_ACRNMS
}
\description{
Get lat lon (and NAICS) via Facility Registry ID
}
\examples{
 latlon_from_regid(110070874073)
x = latlon_from_regid(
    c(110071293460, 110070874073, 110070538057, 110044340807,
       110030509215, 110019033810, 110056111559, 110056982323)
        )

}
