% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_vectorofcsvpairs.R
\name{latlon_from_vectorofcsvpairs}
\alias{latlon_from_vectorofcsvpairs}
\title{helper function - convert vector of lat,lon pairs to data.frame}
\usage{
latlon_from_vectorofcsvpairs(x)
}
\arguments{
\item{x}{vector of comma-separated pairs of lat,lon values
stored as character strings
such as c("30,-83",  "  32.5,  -86.377325 ")}
}
\value{
data.frame with colnames lat and lon
}
\description{
helper function - convert vector of lat,lon pairs to data.frame
}
\examples{
lat_x = testpoints_10$lat
lon_x = testpoints_10$lon
latlon_pairs = latlon2csv(lat = lat_x, lon = lon_x)
latlon_from_vectorofcsvpairs(latlon_pairs) 
all.equal(testpoints_10[, c("lat", "lon")], latlon_from_vectorofcsvpairs(latlon_pairs))

x = latlon_from_vectorofcsvpairs(c("30,-83",  "  32.5,  -86.377325 "))
x
latlon_is.valid(x)

}
\keyword{internal}
