% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_is.valid.R
\name{latlon_is.islandareas}
\alias{latlon_is.islandareas}
\title{Check lat lon coordinates to see if each is approx. in general area of US Island Areas Guam, USVI, Amer Samoa or N Marianas}
\usage{
latlon_is.islandareas(lat, lon, exact_but_slow_islandareas = FALSE)
}
\arguments{
\item{lat}{vector of latitudes}

\item{lon}{vector of longitudes}

\item{exact_but_slow_islandareas}{optional logical, set it to TRUE to check each point vs
boundaries in \link{states_shapefile} to identify which ones are in Island Areas according to that shapefile.
The default method here is much faster, but just checks if a point is within a bounding box
that should approximate each of the Island Areas, found in the object \link{islandareas}.}
}
\value{
vector of TRUE / FALSE values indicating a given lat lon pair
is approximately in one of the rough bounding boxes that includes the 4 Island Areas.
}
\description{
See \link{islandareas}
}
\examples{
\dontrun{
# this would require the testpoints_1000 data from the EJAM package:
  isles <- which(latlon_is.islandareas(lat = testpoints_1000$lat, lon = testpoints_1000$lon))
  mapfast(testpoints_1000[isles, ]) # c(213,785)
  which(!(latlon_is.usa(lat = testpoints_1000$lat, lon = testpoints_1000$lon)))
}

}
\seealso{
\code{\link[=latlon_is.usa]{latlon_is.usa()}} \code{\link[=latlon_is.islandareas]{latlon_is.islandareas()}} \code{\link[=latlon_is.available]{latlon_is.available()}} \code{\link[=latlon_is.possible]{latlon_is.possible()}}
}
\keyword{internal}
