% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_is.valid.R
\name{latlon_is.possible}
\alias{latlon_is.possible}
\title{Check if lat lon between -180 and +180}
\usage{
latlon_is.possible(lat, lon)
}
\arguments{
\item{lat}{vector of latitudes}

\item{lon}{vector of longitudes}
}
\value{
logical vector, one element per lat lon pair (location)
}
\description{
Check if lat lon between -180 and +180
}
\seealso{
\code{\link[=latlon_is.usa]{latlon_is.usa()}} \code{\link[=latlon_is.islandareas]{latlon_is.islandareas()}} \code{\link[=latlon_is.available]{latlon_is.available()}} \code{\link[=latlon_is.possible]{latlon_is.possible()}}
}
\keyword{internal}
