% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_is.valid.R
\name{latlon_is.usa}
\alias{latlon_is.usa}
\title{Check lat lon coordinates to see if each is approx. in general area of USA excluding Island Areas}
\usage{
latlon_is.usa(lat, lon)
}
\arguments{
\item{lat}{vector of latitudes}

\item{lon}{vector of longitudes}
}
\value{
logical vector, one element per lat lon pair (location)
Indicates the point is approximately in one of the
rough bounding boxes that includes the USA without
the Island Areas Guam, American Samoa, USVI, N Marianas Islands.
}
\description{
Check lat lon coordinates to see if each is approx. in general area of USA excluding Island Areas
}
\seealso{
\code{\link[=latlon_is.usa]{latlon_is.usa()}} \code{\link[=latlon_is.islandareas]{latlon_is.islandareas()}} \code{\link[=latlon_is.available]{latlon_is.available()}} \code{\link[=latlon_is.possible]{latlon_is.possible()}}
}
\keyword{internal}
