% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_join_on_bgid.R
\name{latlon_join_on_bgid}
\alias{latlon_join_on_bgid}
\title{utility to add lat lon columns to data.table by reference, joining on bgid}
\usage{
latlon_join_on_bgid(x)
}
\arguments{
\item{x}{data.table with column called bgid
(as used in bgid2fips or blockgroupstats)}
}
\value{
x with 2 new columns but side effect is it updates x in calling envt
}
\description{
get expanded version of data.table, such as copy(blockgroupstats),
with new lat,lon columns
}
\examples{
# quick map of blockgroups in 1 state, shown as blockgroup centroids
myst <- "NY"
dat <- bgpts[fips2state_abbrev(substr(bgfips,1,2)) == myst, ]
mapfast(dat, radius = 0.1)

# same but popups have all the indicators from EJScreen
myst <- "NY"
dat <- copy(blockgroupstats[ST == myst, ])
# add latlon cols by reference:
latlon_join_on_bgid(dat)
# specify useful labels for the map popups
mapfast(dat, radius = 0.1,
        labels = fixcolnames(names(dat), 'r', 'shortlabel'))

## or add the useful labels to the table 1st
names(dat) <- fixcolnames(names(dat), "r", "shortlabel")
mapfast(dat, radius = 0.1)

}
\keyword{internal}
