% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_linesofcode_.R
\name{linesofcode2}
\alias{linesofcode2}
\title{linesofcode2 - UTILITY - count lines of source code per .R file (not per function) - NOT TESTED}
\usage{
linesofcode2(
  folder = ".",
  packages,
  recursive = TRUE,
  sums = FALSE,
  rfolderonly = FALSE,
  cropfilename = 40,
  croppath = 20,
  showrows = NULL
)
}
\arguments{
\item{folder}{path to folder with .R files}

\item{packages}{optional vector of names of packages of source code}

\item{recursive}{logical, look in subfolders}

\item{sums}{logical, if TRUE, returns sums info, otherwise just prints that to console and returns more info}

\item{rfolderonly}{logical}

\item{cropfilename}{number of character to truncate filename to for display in console}

\item{croppath}{limit path for display}

\item{showrows}{optional}
}
\value{
data.frame of info about files
}
\description{
linesofcode2 - UTILITY - count lines of source code per .R file (not per function) - NOT TESTED
}
\keyword{internal}
