% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_FUNCTIONS.R
\name{map_counties_in_state}
\alias{map_counties_in_state}
\title{Basic map of county outlines within specified state(s)
Not used by shiny app}
\usage{
map_counties_in_state(
  ST = "DE",
  colorcolumn = c("pop", "NAME", "POP_SQMI", "STATE_NAME")[1],
  type = c("leaflet", "mapview")[1]
)
}
\arguments{
\item{ST}{a vector of one or more state abbreviations, like

ST = "ME"  or  ST = c("de", "RI"), or

ST = fips2state_abbrev(fips_state_from_statename(c("Rhode Island", "district of columbia")))

or e.g., all counties in EPA Region 1:

ST = stateinfo$ST\link{stateinfo$REGION == 1}}

\item{colorcolumn}{name of column to use in setting colors of counties on map,
but must be one returned by \code{\link[=shapes_counties_from_countyfips]{shapes_counties_from_countyfips()}} like "STATE_NAME"}

\item{type}{must be "leaflet" or can be "mapview" if installed and loaded}
}
\value{
a map
}
\description{
Basic map of county outlines within specified state(s)
Not used by shiny app
}
\examples{
\dontrun{
map_counties_in_state(ST = c('id', 'mt'))
map_counties_in_state(ST = c('id', 'mt'),
  colorcolumn = "STATE_NAME")
 
map_counties_in_state(ST = c('id', 'mt'), type = "mapview")
map_counties_in_state(ST = c('id', 'mt'), type = "mapview",
  colorcolumn = "STATE_NAME")

 map_counties_in_state(
  ST = c( 'md', 'pa'),
   type = "mapview", colorcolumn = "POP_SQMI") 
}
}
