% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_FUNCTIONS.R
\name{map_ejam_plus_shp}
\alias{map_ejam_plus_shp}
\title{helper - map the merge of ejamit(shapefile = x) output and the shapefile x}
\usage{
map_ejam_plus_shp(
  shp,
  out,
  radius_buffer = NULL,
  circle_color = "#000080",
  launch_browser = FALSE
)
}
\arguments{
\item{shp}{spatial data.frame}

\item{out}{output of ejamit()}

\item{radius_buffer}{optional but can be obtained from out}

\item{circle_color}{optional}

\item{launch_browser}{set TRUE to have it launch browser to show map.}
}
\value{
map html widget
}
\description{
helper - map the merge of ejamit(shapefile = x) output and the shapefile x
}
\details{
used by server, \code{\link[=ejam2map]{ejam2map()}}, and \code{\link[=mapfast]{mapfast()}}
}
\examples{
\dontrun{
shp1 <- shapefile_from_any(file.path(
  testdatafolder(), "shapes/portland_shp.zip"))
ejam2map(ejamit(shapefile = shp1, radius = 0), shp = shp1)
shp2 <- testshapes_2
ejam2map(ejamit(shapefile = shp2, radius = 0), shp = shp2)
ejam2map(ejamit(shapefile = shp2, radius = 2), shp = shp2)

}

}
\keyword{internal}
