% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_FUNCTIONS.R
\name{map_google}
\alias{map_google}
\title{Map - Open Google Maps in browser}
\usage{
map_google(lat, lon, zoom = 12, point = TRUE, launch = TRUE)
}
\arguments{
\item{lat}{\itemize{
\item Anything that can be handled by \code{\link[=sitepoints_from_any]{sitepoints_from_any()}}.
Leave unspecified to interactively browse to a .xlsx file that has lat,lon columns,
or lat can be a data.frame with lat,lon column names in which case longitude should not be provided,
such as lat = testpoints_10\link{1,}, or lat and lon can be separately provided as vectors.
}}

\item{lon}{longitude, or omit this parameter to provide points as the first parameter.}

\item{zoom}{zoomed out value could be 3 or 5, zoomed in default is 12}

\item{point}{logical optional, passed to \code{\link[=url_map_google]{url_map_google()}}}

\item{launch}{logical, whether to launch browser}
}
\value{
opens a browser window with Google Maps centered on the specified lat, lon
}
\description{
Map - Open Google Maps in browser
}
\examples{
# map_google(testpoints_10[1,])

}
