% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_FUNCTIONS.R
\name{map_shapes_plot}
\alias{map_shapes_plot}
\title{Map - polygons - Use base R plot() to map polygons}
\usage{
map_shapes_plot(shapes, main = "Selected Census Units", ...)
}
\arguments{
\item{shapes}{like from shapes_counties_from_countyfips(fips_counties_from_state_abbrev("DE"))}

\item{main}{title for map}

\item{...}{passed to plot()}
}
\value{
Just draws map using plot()
}
\description{
Map - polygons - Use base R plot() to map polygons
}
