% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_community_download.R
\name{map_single_location}
\alias{map_single_location}
\title{helper - creates leaflet map for summary report}
\usage{
map_single_location(
  row_index = NULL,
  inshiny = FALSE,
  input_circleweight_in,
  react_sanitized_bt_rad_buff,
  react_submitted_upload_method,
  react_data_uploaded,
  react_data_processed
)
}
\arguments{
\item{row_index}{which site in the table is the one of interest}

\item{inshiny}{logical}

\item{input_circleweight_in}{from shiny}

\item{react_sanitized_bt_rad_buff}{from shiny}

\item{react_submitted_upload_method}{from shiny}

\item{react_data_uploaded}{from shiny, needed only if shapefile was input}

\item{react_data_processed}{from shiny, results like output of ejamit(), but only 1 row of 1 table gets used}
}
\description{
helper - creates leaflet map for summary report
}
\keyword{internal}
