% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_FUNCTIONS.R
\name{mapfast_gg}
\alias{mapfast_gg}
\title{Map - points - ggplot2 map of points in the USA - very basic map}
\usage{
mapfast_gg(
  mydf = data.frame(lat = 40, lon = -100)[0, ],
  dotsize = 1,
  ptcolor = "black",
  xlab = "Longitude",
  ylab = "Latitude",
  ...
)
}
\arguments{
\item{mydf}{data.frame with columns named lat and lon}

\item{dotsize}{optional, size of dot representing a point}

\item{ptcolor}{optional, color of dot}

\item{xlab}{optional, text for x label}

\item{ylab}{optional, text for y label}

\item{...}{optional, passed to \code{\link[ggplot2:labs]{ggplot2::labs()}}}
}
\value{
a ggplot() object
}
\description{
Map - points - ggplot2 map of points in the USA - very basic map
}
\examples{
\dontrun{
  mapfast_gg(EJAM::testpoints_10)

  pts <- read.table(textConnection(
  "lat lon
  39.5624775 -119.7410994
  42.38748056 -94.61803333"
  ),
  header = TRUE,
  as.is = TRUE
  )
  mapfast_gg(pts)
  # str(pts) # lon, not long
  }

}
