% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_metadata4pins.R
\docType{data}
\name{metadata4pins}
\alias{metadata4pins}
\title{metadata about data object, for pins}
\format{
An object of class \code{data.frame} with 14 rows and 4 columns.
}
\usage{
metadata4pins
}
\description{
metadata about data object, for pins
}
\details{
metadata such as description of dataset, for pins board data in EJAM package
}
\keyword{datasets}
