% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_metadata_add.R
\name{metadata_add}
\alias{metadata_add}
\title{helper function for package to set metadata attributes of a dataset}
\usage{
metadata_add(x)
}
\arguments{
\item{x}{dataset (or any object) whose metadata (stored as attributes) you want to update or create
EJAM, EJScreen, and other dataset versions and release dates are tracked in DESCRIPTION}
}
\value{
returns x but with new or altered attributes
}
\description{
Together with the metadata_mapping script, this can be used
annually to update the metadata for datasets in a package.
It just makes it easier to set a few metadata attributes similarly
for a number of data elements, for example,
to add new or update existing attributes.
}
\details{
This utility would be used in scripts in EJAM/data-raw/ to
add metadata to objects like x before use_data(x, overwrite=T)
}
\examples{
  # metadata_check() # internal function
  x <- data.frame(a=1:10,b=1001:1010)
  # x <- metadata_add(x) # internal function
  attributes(x)

}
\seealso{
metadata_check()
}
\keyword{internal}
