% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meters_per_mile.R
\docType{data}
\name{meters_per_mile}
\alias{meters_per_mile}
\title{how many meters are in one mile (for conversions between units)}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
meters_per_mile
}
\description{
1609.344 meters per mile
}
\keyword{datasets}
