% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_FUNCTIONS.R
\name{naics_categories}
\alias{naics_categories}
\title{NAICS - See the names of industrial categories and their NAICS code}
\usage{
naics_categories(digits = 2, dataset = EJAM::NAICS)
}
\arguments{
\item{digits}{default is 2, for 2-digits NAICS, the top level, but could be up to 6.}

\item{dataset}{Should default to the dataset called NAICS, installed with this package.
see \link{NAICS}  Check attr(NAICS, 'year')}
}
\value{
matrix with 1 column of 2-digit codes and rownames that look like
"22 - Utilities" etc.
}
\description{
Easy way to list the 2-digit NAICS (17 categories), or other level
}
\details{
Also see \url{https://www.naics.com/search/}

There are this many NAICS codes roughly by number of digits in the code:

table(nchar(NAICS))

2    3    4    5    6

17   99  311  709 1057

See \url{https://www.census.gov/naics/}
}
\examples{
 naics_categories()
}
\seealso{
\link{naics_from_any}  \link{NAICS}
}
