% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_FUNCTIONS.R
\name{naics_findwebscrape}
\alias{naics_findwebscrape}
\title{for query term, show list of roughly matching NAICS, scraped from web}
\usage{
naics_findwebscrape(query)
}
\arguments{
\item{query}{text like "gasoline" or "copper smelting"}
}
\value{
data.frame of info on what was found, naics and title
}
\description{
This finds more than just \code{\link[=naics_from_any]{naics_from_any()}} does, since that needs an exact match
but this looks at naics.com website which lists various aliases for a sector.
}
\examples{
 # naics_from_any("copper smelting")
 # naics_from_any("copper smelting", website_scrape=TRUE)
 # browseURL(naics_from_any("copper smelting", website_url=TRUE) )

  url_naics.com("copper smelting")
  \dontrun{
  naics_findwebscrape("copper smelting")
  browseURL(url_naics.com("copper smelting"))
  browseURL(naics_url_of_code(326))
  }

}
\seealso{
\code{\link[=naics_from_any]{naics_from_any()}}  \code{\link[=url_naics.com]{url_naics.com()}}
}
