% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_FUNCTIONS.R
\name{naics_from_code}
\alias{naics_from_code}
\title{NAICS - search for industry names by NAICS code(s), 2-6 digits long each}
\usage{
naics_from_code(mycodes, children = FALSE)
}
\arguments{
\item{mycodes}{vector of numeric NAICS codes. see \url{https://naics.com}}

\item{children}{logical, if TRUE, also return all the subcategories - where NAICS starts with the same digits}
}
\value{
a subset of the \link{naicstable} data.table (not just the codes column)
}
\description{
See \code{\link[=naics_from_any]{naics_from_any()}} which uses this
}
\seealso{
\code{\link[=naics_subcodes_from_code]{naics_subcodes_from_code()}} \code{\link[=naics_from_code]{naics_from_code()}}  \code{\link[=naics_from_name]{naics_from_name()}}  \code{\link[=naics_from_any]{naics_from_any()}}
}
\keyword{internal}
