% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_FUNCTIONS.R
\name{naics_from_name}
\alias{naics_from_name}
\title{NAICS - Search for industry names and NAICS codes by query string}
\usage{
naics_from_name(mynames, children = FALSE, ignore.case = TRUE, fixed = FALSE)
}
\arguments{
\item{mynames}{query string, vector of NAICS industry names or any regular expression or partial words. See \url{https://naics.com}}

\item{children}{logical, if TRUE, also return all the subcategories - where NAICS starts with the same digits}

\item{ignore.case}{see \code{\link[=grepl]{grepl()}}}

\item{fixed}{should it be an exact match? see \code{\link[=grepl]{grepl()}}}
}
\value{
a subset of the \link{naicstable} data.table (not just the codes column)
}
\description{
query by parts of words, etc. in the industry name.
}
\details{
See \code{\link[=naics_from_any]{naics_from_any()}} which uses this
}
\examples{
 data.table::fintersect(naics_from_any( "manufac"), naics_from_any("chem"))
}
\seealso{
\code{\link[=naics_findwebscrape]{naics_findwebscrape()}} \code{\link[=naics_subcodes_from_code]{naics_subcodes_from_code()}} \code{\link[=naics_from_code]{naics_from_code()}}  \code{\link[=naics_from_name]{naics_from_name()}}  \code{\link[=naics_from_any]{naics_from_any()}}
}
\keyword{internal}
