% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naics_is.valid.R
\name{naics_is.valid}
\alias{naics_is.valid}
\title{validate industry NAICS codes}
\usage{
naics_is.valid(code)
}
\arguments{
\item{code}{vector of one or more
numeric or character codes like c(22, 111, 4239, 423860)}
}
\value{
logical vector, TRUE means valid
}
\description{
validate industry NAICS codes
}
\examples{
  naics_is.valid(c(22, "022", " 22", 111, "4239", 423860))
  # table(naics_is.valid(frs_by_naics$NAICS)) / NROW(frs_by_naics)

}
