% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_FUNCTIONS.R
\name{naics_subcodes_from_code}
\alias{naics_subcodes_from_code}
\title{NAICS - find subcategories of the given overall NAICS industry code(s)}
\usage{
naics_subcodes_from_code(mycodes)
}
\arguments{
\item{mycodes}{NAICS codes vector, of 2 to 6 digits each. See \url{https://naics.com}}
}
\value{
a subset of the \link{naicstable} data.table (not just the codes column)
}
\description{
Given 3-digit NAICS code, for example, get all NAICS that start with those digits.
}
\details{
similar idea was naics2children() but this is more robust
See \code{\link[=naics_from_any]{naics_from_any()}} which uses this
}
\examples{
  naics_categories()

}
\seealso{
\code{\link[=naics_subcodes_from_code]{naics_subcodes_from_code()}} \code{\link[=naics_from_code]{naics_from_code()}}  \code{\link[=naics_from_name]{naics_from_name()}}  \code{\link[=naics_from_any]{naics_from_any()}}
}
