% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAICS_FUNCTIONS.R
\name{naics_url_of_code}
\alias{naics_url_of_code}
\title{NAICS - Get URL for page with info about industry sector(s) by NAICS}
\usage{
naics_url_of_code(naics)
}
\arguments{
\item{naics}{vector of one or more NAICS codes, like 11,"31-33",325}
}
\value{
vector of URLs as strings like https://www.naics.com/six-digit-naics/?v=2017&code=22
}
\description{
See (https://naics.com) for more information on NAICS codes
}
\seealso{
\code{\link[=naics_from_any]{naics_from_any()}} \code{\link[=naics_findwebscrape]{naics_findwebscrape()}}
}
