% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_all.R
\docType{data}
\name{names_all}
\alias{names_all}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 562.
}
\usage{
names_all
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
