% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_all_r.R
\docType{data}
\name{names_all_r}
\alias{names_all_r}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 559.
}
\usage{
names_all_r
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
