% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_community.R
\docType{data}
\name{names_community}
\alias{names_community}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 7.
}
\usage{
names_community
}
\description{
a list of variable names for internal use in EJAM
}
\details{
To see other groups of variables: unique(map_headernames$varlist)
}
\seealso{
\code{\link[=varinfo]{varinfo()}}  \link{map_headernames}  \link{names_d} \link{names_e}
}
\keyword{internal}
