% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_pct_as_fraction_blockgroupstats.R
\docType{data}
\name{names_pct_as_fraction_blockgroupstats}
\alias{names_pct_as_fraction_blockgroupstats}
\title{which indicators are percentages stored as 0-1 not 0-100, in blockgroupstats}
\format{
An object of class \code{character} of length 39.
}
\usage{
names_pct_as_fraction_blockgroupstats
}
\description{
which indicators are percentages stored as 0-1 not 0-100, in blockgroupstats
}
\keyword{internal}
