% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_these.R
\docType{data}
\name{names_these}
\alias{names_these}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 30.
}
\usage{
names_these
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
