% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_names_whichlist.R
\name{names_whichlist_multi}
\alias{names_whichlist_multi}
\title{See which lists of names the given indicator names are in}
\usage{
names_whichlist_multi(x, ...)
}
\arguments{
\item{x}{vector of names (query terms)}

\item{...}{passed to names_whichlist()}
}
\value{
a list of sets of names
}
\description{
See which lists of names the given indicator names are in
}
\seealso{
\code{\link[=varinfo]{varinfo()}}
}
\keyword{internal}
