% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_names_whichlist.R
\name{names_whichlist_multi_key}
\alias{names_whichlist_multi_key}
\title{See which key lists of names the given indicator names are in}
\usage{
names_whichlist_multi_key(x, ...)
}
\arguments{
\item{x}{vector of names}

\item{...}{passed to names_whichlist_multi()}
}
\value{
vector maybe
}
\description{
See which key lists of names the given indicator names are in
}
\seealso{
\code{\link[=varinfo]{varinfo()}}
}
\keyword{internal}
