% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_names_wts.R
\docType{data}
\name{names_wts}
\alias{names_wts}
\title{a list of variable names for internal use in EJAM}
\format{
An object of class \code{character} of length 1.
}
\usage{
names_wts
}
\description{
a list of variable names for internal use in EJAM
}
\keyword{datasets}
