% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_namez.R
\docType{data}
\name{namez}
\alias{namez}
\title{namez (DATA) list of lists of indicator names (complete list in 1 object)}
\description{
namez (DATA) list of lists of indicator names (complete list in 1 object)
}
\details{
Not yet implemented/used in EJAM as of 7/2024
}
\seealso{
\code{\link[=varinfo]{varinfo()}}  \link{map_headernames} \link{names_d} \link{names_e}
}
