% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_offline.R
\name{offline}
\alias{offline}
\title{utility to check if internet connection is not available}
\usage{
offline(url = "r-project.org")
}
\arguments{
\item{url}{optional URL checked, using \code{\link[curl:nslookup]{curl::nslookup()}}}
}
\value{
logical (TRUE or FALSE), TRUE if offline, FALSE if can connect to r-project.org
}
\description{
utility to check if internet connection is not available
}
\seealso{
offline_warning() and offline_cat() utilities unexported undocumented,
either returns the same as offline() does,
while also providing, if offline, warning() or cat() info to console
}
\keyword{internal}
