% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barplot_ratios.R
\name{plot_barplot_ratios_ez}
\alias{plot_barplot_ratios_ez}
\title{Same as ejam2barplot() but ejam2barplot() handles a sitenumber parameter}
\usage{
plot_barplot_ratios_ez(
  out,
  varnames = c(names_d_ratio_to_avg, names_d_subgroups_ratio_to_avg),
  main = "Demographics at the Analyzed Locations Compared to US Overall",
  single_location = FALSE,
  row_index = NULL,
  ...
)
}
\arguments{
\item{out}{the list of tables that is the output of ejamit() or a related function}

\item{varnames}{vector of indicator names that are ratios to avg, like
c(names_d_ratio_to_avg , names_d_subgroups_ratio_to_avg)
but could be c(names_d_ratio_to_state_avg , names_d_subgroups_ratio_to_state_avg)}

\item{main}{title of plot - must change to note it vs. State if not comparing to US avg.}

\item{single_location}{set to TRUE if using row_index to view one site,
set to FALSE to view overall results from out$results_overall}

\item{row_index}{the number of the row to use from out$results_bysite}

\item{...}{
  Arguments passed on to \code{\link[=ejam2barplot]{ejam2barplot}}
  \describe{
    \item{\code{ejamitout}}{like from \code{\link[=ejamit]{ejamit()}}}
    \item{\code{sitenumber}}{default is all sites from ejamitout$results_overall, and
if an integer, it is the site number to show from ejamitout$results_bysite}
  }}
}
\description{
Same as ejam2barplot() but ejam2barplot() handles a sitenumber parameter
}
