% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot_pctiles.R
\name{plot_boxplot_pctiles}
\alias{plot_boxplot_pctiles}
\title{Boxplots comparing a few indicators showing how each varies across sites
Visualize mean median etc. for each of several percentile indicators}
\usage{
plot_boxplot_pctiles(
  out,
  vars = names_d_state_pctile,
  ylab = "Percentile in State",
  ranked = TRUE,
  ...
)
}
\arguments{
\item{out}{The output from ejamit() such as
testoutput_ejamit_10pts_1miles}

\item{vars}{Typically would be one of these:
names_d_pctile, names_d_state_pctile,
names_d_subgroups_pctile, names_d_subgroups_state_pctile,
names_e_pctile, names_e_state_pctile,
and possibly ratios or others, but this is designed to plot pctiles.}

\item{ylab}{inferred from vars normally}

\item{ranked}{set FALSE to avoid sorting x axis on size of wtdmeans}

\item{...}{passed to boxplot()}
}
\value{
prints means etc. and plots
}
\description{
Boxplots comparing a few indicators showing how each varies across sites
Visualize mean median etc. for each of several percentile indicators
}
\examples{
\dontrun{
bplot = plot_boxplot_pctiles
bplot(out, names_d_state_pctile)
bplot(out, names_d_subgroups_state_pctile)
bplot(out, names_e_state_pctile)
bplot(out, names_d_pctile)
bplot(out, names_d_subgroups_pctile)
bplot(out, names_e_pctile)
}
}
