% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_bycounty.R
\name{plot_bycounty}
\alias{plot_bycounty}
\title{plot comparison of counties in 1 state, for 1 indicator (variable)}
\usage{
plot_bycounty(
  x,
  myvars = x$variable[1],
  myvarnames = NULL,
  mystate = NULL,
  labeltype = NULL,
  acsinfo = NULL,
  yr = 2022
)
}
\arguments{
\item{x}{table that is output of acs_bycounty() i.e., output of get_acs()
from tidycensus pkg that downloads ACS Census Bureau data via API.}

\item{myvars}{optional .extracted from x, one (or more?) ACS5 variables like "B03002_001"}

\item{myvarnames}{optional friendlier names of myvars}

\item{mystate}{name of state}

\item{labeltype}{as from scales package.  for continuous scales:
label_bytes(), label_number_auto(), label_number_si(), label_ordinal(),
label_parse(), label_percent(), label_pvalue(), label_scientific()}

\item{acsinfo}{large table of metadata as from load_variables function
from tidycensus pkg}

\item{yr}{like 2022, end of 5 year ACS 2018-2022}
}
\value{
plot
}
\description{
plot comparison of counties in 1 state, for 1 indicator (variable)
}
