% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lorenz_popcount_by_site - DRAFT.R
\name{plot_lorenz_distance_by_dcount}
\alias{plot_lorenz_distance_by_dcount}
\title{lorenz plot bybg_people (cumulative share of x vs cum share of y) - DRAFT/EXPERIMENTAL
COUNT OF SITES (or PEOPLE?) BY BIN}
\usage{
plot_lorenz_distance_by_dcount(bybg_people, varname)
}
\arguments{
\item{bybg_people}{from ejamit()$results_bybg_people}

\item{varname}{************ tbd}
}
\value{
a ggplot
}
\description{
lorenz plot bybg_people (cumulative share of x vs cum share of y) - DRAFT/EXPERIMENTAL
COUNT OF SITES (or PEOPLE?) BY BIN
}
