% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lorenz_popcount_by_site - DRAFT.R
\name{plot_lorenz_popcount_by_site}
\alias{plot_lorenz_popcount_by_site}
\title{lorenz plot bysite (cumulative share of x vs cum share of y) - DRAFT/EXPERIMENTAL
COMPARES TWO subsets OF SITES (or people??)}
\usage{
plot_lorenz_popcount_by_site(bysite, radius)
}
\arguments{
\item{bysite}{from ejamit()$results_bysite}

\item{radius}{miles ************ tbd}
}
\value{
a ggplot
}
\description{
lorenz plot bysite (cumulative share of x vs cum share of y) - DRAFT/EXPERIMENTAL
COMPARES TWO subsets OF SITES (or people??)
}
