% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridgeline_ratios.R
\name{plot_ridgeline_ratios_ez}
\alias{plot_ridgeline_ratios_ez}
\title{Make ridgeline plot of ratios of demographic score to its average}
\usage{
plot_ridgeline_ratios_ez(
  out,
  varnames = c(names_d_ratio_to_avg, names_d_subgroups_ratio_to_avg),
  maxratio = 5
)
}
\arguments{
\item{out}{like from ejamit()}

\item{varnames}{vector of colnames in out$results_bysite, the ratio variables}

\item{maxratio}{cap on ratios to show so plot looks better (all values above cap get replaced by cap before plotting)}
}
\description{
Make ridgeline plot of ratios of demographic score to its average
}
\examples{
 out <- testoutput_ejamit_1000pts_1miles
 plot_ridgeline_ratios_ez(out)

}
