% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_st_v_us.R
\name{plot_st_v_us}
\alias{plot_st_v_us}
\title{Barplot of ratios of 1 State's indicator averages vs US overall}
\usage{
plot_st_v_us(ST = "CA", varnames = names_these)
}
\arguments{
\item{ST}{state abbreviation like "NY"}

\item{varnames}{vector of character names of raw indicator variables that are among
names(statestats), like "pm" or "pctlowinc" or a vector like
names_d or names_e}
}
\value{
similar to \code{\link[=plot_barplot_ratios]{plot_barplot_ratios()}}
}
\description{
Barplot of ratios of 1 State's indicator averages vs US overall
}
\examples{
plot_st_v_us("CA", names_these)

}
