% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popup_from_uploadedpoints.R
\name{popup_from_uploadedpoints}
\alias{popup_from_uploadedpoints}
\title{Map popups - make simple popups for map to show info about uploaded points}
\usage{
popup_from_uploadedpoints(mypoints, n = "all")
}
\arguments{
\item{mypoints}{data.frame (or tibble?) with lat and lon columns preferably}

\item{n}{Show the first n columns of mypoints, in popup. "all" means all of them.}
}
\value{
popups vector to be used in leaflet maps
}
\description{
Map popups - make simple popups for map to show info about uploaded points
}
\seealso{
\code{\link[=popup_from_any]{popup_from_any()}}
}
\keyword{internal}
