% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popup_from_ejscreen.R
\name{popup_print}
\alias{popup_print}
\title{Helper function to view popup info in an interactive session - easier format to view}
\usage{
popup_print(x, linkregex = "<a href.*>(.*)<.*", linksimple = "\\\\1")
}
\arguments{
\item{x}{output of \code{\link[=popup_from_ejscreen]{popup_from_ejscreen()}}}

\item{linkregex}{see source}

\item{linksimple}{see source}
}
\description{
Helper function to view popup info in an interactive session - easier format to view
}
\seealso{
\code{\link[=popup_from_ejscreen]{popup_from_ejscreen()}}
}
\keyword{internal}
