% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{pre_comma}
\alias{pre_comma}
\title{utility - keep text before last or 1st comma (get "Waco" from "Waco, TX")
not used}
\usage{
pre_comma(x, lastcomma = TRUE, if_no_comma_do_nothing = TRUE, trim = FALSE)
}
\arguments{
\item{x}{string vector}

\item{lastcomma}{logical}

\item{if_no_comma_do_nothing}{logical}

\item{trim}{logical}
}
\value{
vector like x
}
\description{
utility - keep text before last or 1st comma (get "Waco" from "Waco, TX")
not used
}
\keyword{internal}
