% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROXIMITY_FUNCTIONS.R
\name{proximity.score.in.miles}
\alias{proximity.score.in.miles}
\title{proximity.score.in.miles - convert EJScreen proximity scores to miles per site instead of sites per kilometer
Shows US percentiles if no arguments used}
\usage{
proximity.score.in.miles(scoresdf = NULL)
}
\arguments{
\item{scoresdf}{data.frame of simple proximity scores like for tsdf, rmp, npl
but not traffic.score or npdes one since those are weighted and not just count per km}
}
\description{
proximity.score.in.miles - convert EJScreen proximity scores to miles per site instead of sites per kilometer
Shows US percentiles if no arguments used
}
