% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROXIMITY_FUNCTIONS.R
\name{proxistat_via_getblocks}
\alias{proxistat_via_getblocks}
\title{proximity score tool DRAFT - trying to use getblocksnearby() normally (pts to blocks)
but then filling in that info for rest of blocks in US}
\usage{
proxistat_via_getblocks(pts, countradius = 5, maxradius = 31)
}
\arguments{
\item{pts}{data.table with lat lon column names}

\item{countradius}{distance within in which nearby sites are counted to create proximity score.
In miles, and default is 5 km (5000 / meters_per_mile = 3.106856 miles)
which is the EJScreen zone for proximity scores based on counts.}

\item{maxradius}{max distance in miles to search for nearest single facility,
if none found within countradius. EJScreen seems to use 1,000 km as the max to search,
since the lowest scores for proximity scores of RMP, TSDF, or NPL are ROUGHLY 0.001,
(exactly 0.000747782)
meaning approx. 1/1000 km and km_per_mile = 1.609344 = meters_per_mile / 1000
so 1000 km is 1000 / 1.609344 = 621.3712 miles.
However, the exact min value implies 1337.288 kilometers, or 830.9523 miles?}
}
\value{
data.table of block groups, with proximityscore, bgfips, lat, lon, etc.
}
\description{
proximity score tool DRAFT - trying to use getblocksnearby() normally (pts to blocks)
but then filling in that info for rest of blocks in US
}
