% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_or_xl.R
\name{read_csv_or_xl}
\alias{read_csv_or_xl}
\title{Read table of data from .csv or .xlsx Excel file}
\usage{
read_csv_or_xl(fname = NULL, show_col_types = FALSE, rowsize_warn = 30 * 1000)
}
\arguments{
\item{fname}{full path to folder and filename}

\item{show_col_types}{FALSE makes it print less to console as it reads using readr::read_csv()}

\item{rowsize_warn}{Give warning if there are more than this many rows in the table}
}
\value{
data.frame with contents of table it read
}
\description{
Read simple table from csv or xls or xlsx.
For excel format, must be simple table on first tab,
one row for header (column names),
data itself starting in first cell of second row, like A2, and
all other rows and columns must be empty.
}
\details{
If in RStudio, interactively can select file from local drive. Also used by EJAM shiny app for file uploads.
}
\keyword{internal}
