% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_naics.R
\name{regid_from_naics}
\alias{regid_from_naics}
\title{Find registry ids of EPA-regulated facilities in FRS by NAICS code (industrial category)
Like latlon_from_naics() but returns only regid}
\usage{
regid_from_naics(naics, children = TRUE, id_only = TRUE, ...)
}
\arguments{
\item{naics}{a vector of naics codes, or
a data.table with column named code, as with output of \code{\link[=naics_from_any]{naics_from_any()}}}

\item{children}{optional logical. Must set to TRUE to get facilities whose NAICS
starts with provided naics (or naics based on provided title) rather than
only exact matches. Many facilities have only a longer more specific NAICS code
listed in the FRS, such as a 6-digit code, so if the category (e.g., 4-digit)
is queried then children = TRUE has to be specified to find all the sites
within that overall category.}

\item{id_only}{optional, only for backward compatibility}

\item{...}{passed to \code{\link[=naics_from_any]{naics_from_any()}}}
}
\value{
vector of registry ID values of facilities in EPA FRS
that are listed there as being in this/these NAICS,
like \code{\link[=latlon_from_naics]{latlon_from_naics()}} but with id_only = TRUE
}
\description{
Find registry ids of EPA-regulated facilities in FRS by NAICS code (industrial category)
Like latlon_from_naics() but returns only regid
}
\details{
Finding the right NAICS/SIC and finding all the right
sites is complicated. See discussion of \code{\link[=latlon_from_naics]{latlon_from_naics()}}.
}
\seealso{
\code{\link[=latlon_from_naics]{latlon_from_naics()}}
}
