% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{report_residents_within_xyz}
\alias{report_residents_within_xyz}
\title{Build text for report: Residents within( X miles of)( any of) the (N) point(s)/polygon(s)/Census unit(s)}
\usage{
report_residents_within_xyz(
  text1 = "Residents within ",
  radius = NULL,
  unitsingular = "mile",
  area_in_square_miles = NULL,
  nsites = 1,
  sitenumber = NULL,
  ejam_uniq_id = NULL,
  sitetype = c("latlon", "fips", "shp", "frs", "epa_program_up", "echo", "naics", "sic",
    "mact", "epa_program_sel")[1]
)
}
\arguments{
\item{text1}{text to start the phrase, like "Residents within "}

\item{radius}{The distance from each place, normally in miles (which can be 0),
or custom text like "seven kilometers from" in which case
it should end with words like "a safe distance from" or
"the vicinity of" or "proximity to" or "near"
-- but may need to specify custom text1 also.}

\item{unitsingular}{'mile' by default, but can use 'kilometer' etc.
Ignored if radius is not a number.}

\item{area_in_square_miles}{number if available, area in square miles, added as a second line}

\item{nsites}{number of places or text in lieu of number}

\item{sitenumber}{if the 1 site is from a list of sites, can say which one (1:N)}

\item{ejam_uniq_id}{if the 1 site is from a list of sites, can say which ID}

\item{sitetype}{can be 'latlon', 'fips', 'shp', or
some singular custom text like "Georgia location"
but should be something that can be made plural by just adding "s" so ending with "site"
works better than ending with "... facility" since that would print as "facilitys" here.}
}
\description{
Build text for report: Residents within( X miles of)( any of) the (N) point(s)/polygon(s)/Census unit(s)
}
\seealso{
used in app_server to create locationstr for  \code{\link[=build_community_report]{build_community_report()}}
}
\keyword{internal}
