% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_community_report.R
\name{report_setup_temp_files}
\alias{report_setup_temp_files}
\title{helper - copies template, css to tempdir for render of summary report
helper - copies .Rmd (template), .css from Rmd_folder to a temp dir subfolder for rendering}
\usage{
report_setup_temp_files(
  Rmd_name = "community_report_template.Rmd",
  Rmd_folder = "report/community_report/"
)
}
\arguments{
\item{Rmd_name}{.Rmd filename the package uses}

\item{Rmd_folder}{folder the package stores the template in}
}
\description{
helper - copies template, css to tempdir for render of summary report
helper - copies .Rmd (template), .css from Rmd_folder to a temp dir subfolder for rendering
}
\details{
used by \code{ejam2report()} only now? was also copying  logo .png ? but not now?
}
\keyword{internal}
