% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{report_xmilesof}
\alias{report_xmilesof}
\title{helper for \code{\link[=report_residents_within_xyz]{report_residents_within_xyz()}}}
\usage{
report_xmilesof(radius, unitsingular = "mile")
}
\arguments{
\item{radius}{distance from site}

\item{unitsingular}{use default units}
}
\description{
helper for \code{\link[=report_residents_within_xyz]{report_residents_within_xyz()}}
}
\keyword{internal}
