% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.summarize.helpers.R
\name{rowMaxs2}
\alias{rowMaxs2}
\title{Get maximum of each row of data.frame}
\usage{
rowMaxs2(df, na.rm = TRUE)
}
\arguments{
\item{df}{data.frame or matrix.}

\item{na.rm}{default is TRUE. not tested for FALSE}
}
\value{
max of each row
}
\description{
Get maximum of each row of data.frame
}
\seealso{
colMaxs2()
}
\keyword{internal}
