% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames_etc.R
\name{setdiff_yx}
\alias{setdiff_yx}
\title{UTILITY - see what is in y not x}
\usage{
setdiff_yx(x, y)
}
\description{
utility just like setdiff except for y,x instead of x,y
}
\examples{
  setdiff(   1:4, 3:8)
  setdiff_yx(1:4, 3:8) # makes it easy to check both without 
}
\keyword{internal}
