% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_xyz.R
\name{shape_buffered_from_shapefile_points}
\alias{shape_buffered_from_shapefile_points}
\title{shape_buffered_from_shapefile_points - add buffer around shape (points, here)}
\usage{
shape_buffered_from_shapefile_points(
  shapefile_points,
  radius.miles,
  crs = 4269,
  ...
)
}
\arguments{
\item{shapefile_points}{spatial object like areas at high risk or areas with facilities to be analyzed}

\item{radius.miles}{width of buffer to add to shapefile_points
(in case dist is a units object, it should be
convertible to arc_degree if x has geographic coordinates,
and to st_crs(x)$units otherwise)}

\item{crs}{used in st_transform()  default is crs = 4269 or Geodetic CRS NAD83}

\item{...}{passed to st_buffer()}
}
\value{
same format as \code{\link[sf:geos_unary]{sf::st_buffer()}} returns
}
\description{
shape_buffered_from_shapefile_points - add buffer around shape (points, here)
}
\details{
Just a wrapper for \code{\link[sf:geos_unary]{sf::st_buffer()}}
}
\examples{
map_shapes_leaflet(
  shape_buffered_from_shapefile_points(
    shapefile_from_sitepoints(testpoints_100), 
    radius.miles = 3
  )
) 
# (ignoring projections for this example)
# compare to
mapfast(testpoints_100)

}
\seealso{
\code{\link[=get_blockpoints_in_shape]{get_blockpoints_in_shape()}} \code{\link[=shapefile_from_sitepoints]{shapefile_from_sitepoints()}} \code{\link[=shape_buffered_from_shapefile_points]{shape_buffered_from_shapefile_points()}}
}
