% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blockpoints_in_shape.R
\name{shapefile2blockpoints}
\alias{shapefile2blockpoints}
\title{Find all Census blocks in a polygon, using internal point of block}
\usage{
shapefile2blockpoints(
  polys,
  addedbuffermiles = 0,
  blocksnearby = NULL,
  dissolved = FALSE,
  safety_margin_ratio = 1.1,
  crs = 4269,
  updateProgress = NULL
)
}
\arguments{
\item{polys}{Spatial data as from sf::st_as_sf(), with
points as from \code{\link[=shapefile_from_sitepoints]{shapefile_from_sitepoints()}},
or a table of points with lat,lon columns that will first be converted here using that function,
or polygons}

\item{addedbuffermiles}{width of optional buffering to add to the points (or edges), in miles}

\item{blocksnearby}{optional table of blocks with blockid, etc (from which lat,lon can be looked up in blockpoints dt)}

\item{dissolved}{If TRUE, use sf::st_union(polys) to find unique blocks inside any one or more of polys}

\item{safety_margin_ratio}{multiplied by addedbuffermiles, how far to search for
blocks nearby using getblocksnearby(), before using those found to do the intersection via sf::}

\item{crs}{used in st_as_sf() and st_transform() and shape_buffered_from_shapefile_points(), crs = 4269 or Geodetic CRS NAD83}

\item{updateProgress}{optional Shiny progress bar to update}
}
\value{
same as \code{\link[=get_blockpoints_in_shape]{get_blockpoints_in_shape()}}
}
\description{
Find all Census blocks in a polygon, using internal point of block
}
