% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_xyz.R
\name{shapefile_from_filepaths}
\alias{shapefile_from_filepaths}
\title{Read shapefile from disk based on the filenames given}
\usage{
shapefile_from_filepaths(
  filepaths = NULL,
  cleanit = TRUE,
  crs = 4269,
  layer = NULL,
  ...
)
}
\arguments{
\item{filepaths}{vector of full paths with filenames (types .shp, .shx, .dbf, and .prj) as strings}

\item{cleanit}{set to FALSE if you want to skip validation and dropping invalid rows}

\item{crs}{if cleanit = TRUE, crs is passed to shapefile_clean()
default is crs = 4269 or Geodetic CRS NAD83
Also can check this via x <- sf::st_crs(sf::st_read()); x$input}

\item{layer}{optional name of layer to read}

\item{...}{passed to \code{\link[sf:st_read]{sf::st_read()}}}
}
\value{
a shapefile object using \code{\link[sf:st_read]{sf::st_read()}}
}
\description{
Read shapefile from disk based on the filenames given
}
\seealso{
\code{\link[=shapefile_from_folder]{shapefile_from_folder()}}
}
