% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_xyz.R
\name{shapefile_from_gdb}
\alias{shapefile_from_gdb}
\title{read .gdb geodatabase file via sf::st_read()}
\usage{
shapefile_from_gdb(fname, layer = NULL, ...)
}
\arguments{
\item{fname}{path and filename of .gdb file}

\item{layer}{optional name of layer, see \code{\link[sf:st_read]{sf::st_read()}}}

\item{...}{passed to \code{\link[sf:st_read]{sf::st_read()}}}
}
\value{
like output of \code{\link[sf:st_read]{sf::st_read()}} but with ejam_uniq_id column 1:NROW()
}
\description{
read .gdb geodatabase file via sf::st_read()
}
\examples{
  # npl <- sf::st_read("~/../Desktop/NPL/NPL_Boundaries.gdb")
  # npl <- shapefile_from_gdb("~/../Desktop/NPL/NPL_Boundaries.gdb",
  #   layer = "SITE_BOUNDARIES_SF")
  # npl <- shapefile_from_gdbzip("~/../Desktop/NPL/NPL_Boundaries.zip")
  # mapview::mapview(npl[x$STATE_CODE == "CA", ])

}
