% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_xyz.R
\name{shapefile_from_sitepoints}
\alias{shapefile_from_sitepoints}
\title{Convert table of lat,lon points/sites into sf:: shapefile}
\usage{
shapefile_from_sitepoints(sitepoints, crs = 4269, ...)
}
\arguments{
\item{sitepoints}{a data.table or data.frame with columns called lat,lon}

\item{crs}{used in st_as_sf() default is crs = 4269 or Geodetic CRS NAD83}

\item{...}{passed to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}
}
\value{
A shapefile via \code{\link[sf:st_as_sf]{sf::st_as_sf()}}
}
\description{
Creates a simple feature (sf) dataframe from points
}
\seealso{
\code{\link[=get_blockpoints_in_shape]{get_blockpoints_in_shape()}} \code{\link[=shapefile_from_sitepoints]{shapefile_from_sitepoints()}} \code{\link[=shape_buffered_from_shapefile_points]{shape_buffered_from_shapefile_points()}}
}
