% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefix.R
\name{shapefix}
\alias{shapefix}
\title{shapefix cleans a spatial data.frame, flags invalid rows, add id if missing, etc.}
\usage{
shapefix(shp, crs = 4269)
}
\arguments{
\item{shp}{simple feature data.frame}

\item{crs}{coordinate reference system, default is 4269}
}
\value{
returns all rows of shp, but adds columns "valid" and "invalid_msg"
and adds attributes shiny can use to update some reactives,
and standardizes "geometry" as the sfc column name.
}
\description{
a way for app_server, and ejamit() via shapefile_from_any(),
to both use this one function to do the same thing
whether or not in a reactive context
}
\keyword{internal}
