% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_from_fips.R
\name{shapes_places_from_placefips}
\alias{shapes_places_from_placefips}
\title{Get shapefiles/ boundaries of census places like cities}
\usage{
shapes_places_from_placefips(fips)
}
\arguments{
\item{fips}{vector of 7-digit City/town/CDP codes as in \link{censusplaces$fips}}
}
\value{
spatial data.frame for mapping
}
\description{
Get shapefiles/ boundaries of census places like cities
}
\seealso{
\code{\link[=shapes_from_fips]{shapes_from_fips()}}
}
\keyword{internal}
