% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_from_fips.R
\name{shapes_state_from_statefips}
\alias{shapes_state_from_statefips}
\title{Get boundaries of State(s) for mapping}
\usage{
shapes_state_from_statefips(fips)
}
\arguments{
\item{fips}{vector of one or more State FIPS codes}
}
\value{
spatial data.frame of boundaries
}
\description{
Get boundaries of State(s) for mapping
}
\seealso{
\code{\link[=shapes_from_fips]{shapes_from_fips()}}
}
\keyword{internal}
