% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_from_fips.R
\name{shapes_tract_from_tractfips}
\alias{shapes_tract_from_tractfips}
\title{Get tract boundaries, via API, to map them}
\usage{
shapes_tract_from_tractfips(
  fips,
  outFields = c("FIPS", "STATE_ABBR", "SQMI"),
  myservice =
    "https://services.arcgis.com/P3ePLMYs2RVChkJx/ArcGIS/rest/services/USA_Boundaries_2022/FeatureServer/4/query"
)
}
\arguments{
\item{fips}{one or more FIPS codes as 11-character strings in a vector}

\item{outFields}{can be "*" for all, or can be
just a vector of variables that particular service provides, like FIPS, SQMI, POPULATION_2020, etc.}

\item{myservice}{URL of feature service to get shapes from.}
}
\value{
spatial object via \code{\link[sf:st_read]{sf::st_read()}} # sf-data.frame, not sf-tibble like \code{\link[sf:st_read]{sf::read_sf()}}
}
\description{
Get tract boundaries, via API, to map them
}
\details{
This is useful mostly for small numbers of tracts.
The EJScreen map services provide other ways to map tracts and see EJScreen data.
}
\seealso{
\code{\link[=shapes_from_fips]{shapes_from_fips()}}
}
\keyword{internal}
