% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SIC.R
\docType{data}
\name{SIC}
\alias{SIC}
\title{SIC (DATA) named list of all SIC code numbers and category name for each}
\format{
An object of class \code{character} of length 1117.
}
\usage{
SIC
}
\description{
SIC (DATA) named list of all SIC code numbers and category name for each
}
\details{
Also see \href{https://siccode.com}{SIC info}
}
\seealso{
\link{SIC} \link{sictable} \code{\link[=sic_categories]{sic_categories()}}
}
\keyword{datasets}
