% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_xyz.R
\name{sic_from_any}
\alias{sic_from_any}
\title{General way to search for industry names and NAICS codes}
\usage{
sic_from_any(
  query,
  children = FALSE,
  ignore.case = TRUE,
  fixed = FALSE,
  website_scrape = FALSE,
  website_url = FALSE
)
}
\arguments{
\item{query}{query string(s) and/or number(s), vector of NAICS codes or industry names or any regular expression or partial words}

\item{children}{logical, if TRUE, also return all the subcategories - where NAICS starts with the same digits}

\item{ignore.case}{see \code{\link[=grepl]{grepl()}}}

\item{fixed}{should it be an exact match? see \code{\link[=grepl]{grepl()}}}

\item{website_scrape}{whether to scrape info from the NAICS website to return a table of codes and names that match (web query uses synonyms so gets more hits)}

\item{website_url}{whether to return the URL of the webpage with info on the NAICS (web query uses synonyms so gets more hits)}
}
\value{
a subset of the \link{sictable} data.table (not just the codes column)
}
\description{
Find industry names and codes by searching for queried code(s) or text
}
\seealso{
\code{\link[=sic_subcodes_from_code]{sic_subcodes_from_code()}} \code{\link[=sic_from_code]{sic_from_code()}}  \code{\link[=sic_from_name]{sic_from_name()}}  \code{\link[=sic_from_any]{sic_from_any()}}
}
