% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_xyz.R
\name{sic_from_code}
\alias{sic_from_code}
\title{Search for industry names by SIC code(s), 4 digits each}
\usage{
sic_from_code(mycodes, children = FALSE)
}
\arguments{
\item{mycodes}{vector of character SIC codes. see \url{https://siccode.com}}

\item{children}{logical, if TRUE, also return all the subcategories - where SIC starts with the same digits}
}
\value{
a subset of the \link{sictable} data.table (not just the codes column)
}
\description{
Search for industry names by SIC code(s), 4 digits each
}
\seealso{
\code{\link[=sic_subcodes_from_code]{sic_subcodes_from_code()}} \code{\link[=sic_from_code]{sic_from_code()}}  \code{\link[=sic_from_name]{sic_from_name()}}
}
