% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_xyz.R
\name{sic_from_name}
\alias{sic_from_name}
\title{Search for industry names and SIC codes by query string}
\usage{
sic_from_name(mynames, children = FALSE, ignore.case = TRUE, fixed = FALSE)
}
\arguments{
\item{mynames}{query string, vector of SIC industry names or any regular expression or partial words. See \url{https://siccode.com}}

\item{children}{logical, if TRUE, also return all the subcategories - where SIC starts with the same digits}

\item{ignore.case}{see \code{\link[=grepl]{grepl()}}}

\item{fixed}{should it be an exact match? see \code{\link[=grepl]{grepl()}}}
}
\value{
a subset of the \link{sictable} data.table (not just the codes column)
}
\description{
query by parts of words, etc. in the industry name.
}
\examples{
 data.table::fintersect(sic_from_any( "glass"), sic_from_any("paint"))
}
\seealso{
\code{\link[=sic_subcodes_from_code]{sic_subcodes_from_code()}} \code{\link[=sic_from_code]{sic_from_code()}}  \code{\link[=sic_from_name]{sic_from_name()}}  \code{\link[=sic_from_any]{sic_from_any()}}
}
