% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sic_xyz.R
\name{sic_subcodes_from_code}
\alias{sic_subcodes_from_code}
\title{Find subcategories of the given overall SIC industry code(s)}
\usage{
sic_subcodes_from_code(mycodes)
}
\arguments{
\item{mycodes}{SIC codes vector, of 2 to 4 digits each. See \url{https://siccode.com}}
}
\value{
a subset of the \link{sictable} data.table (not just the codes column)
}
\description{
Given 3-digit SIC code, for example, get all SIC that start with those digits.
}
\details{
similar idea was naics2children() but this is more robust
See \code{\link[=sic_from_any]{sic_from_any()}} which uses this
}
\examples{
  # codes starting with '07'
  sic_subcodes_from_code('07')
  # codes starting with '078'
  sic_subcodes_from_code('078')
}
\seealso{
\code{\link[=sic_subcodes_from_code]{sic_subcodes_from_code()}} \code{\link[=sic_from_code]{sic_from_code()}}  \code{\link[=sic_from_name]{sic_from_name()}}  \code{\link[=sic_from_any]{sic_from_any()}}
}
