% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sictable.R
\docType{data}
\name{sictable}
\alias{sictable}
\title{sictable (DATA) data.table of SIC code(s) and industry names for each EPA-regulated site}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1117 rows and 6 columns.
}
\usage{
sictable
}
\description{
data.table of SIC code(s) and industry names for each EPA-regulated site in Facility Registry Service
Also has the 2,3, and 4-digit SIC that this code falls under, where relevant for given length
}
\details{
This is similar to the data file EJAM::SIC but in a more useful format and newer functions work with it. Also see \href{https://siccode.com}{SIC info}
}
\seealso{
\link{SIC} \link{sictable} \code{\link[=sic_categories]{sic_categories()}}
}
\keyword{datasets}
