% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_anything.R
\name{sites_from_file}
\alias{sites_from_file}
\title{helper - given filename, figure out type and return list of input params for ejamit()
Do not actually read file but get list of sitepoints, fips, shapefile args to pass to ejamit()}
\usage{
sites_from_file(file)
}
\arguments{
\item{file}{a file name (with path) to look at}
}
\value{
named list, with sitepoints, fips, shapefile as names
}
\description{
helper - given filename, figure out type and return list of input params for ejamit()
Do not actually read file but get list of sitepoints, fips, shapefile args to pass to ejamit()
}
\keyword{internal}
