% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{sitetype2text}
\alias{sitetype2text}
\title{helper to convert sitetype code ("latlon") to text describing it (" specified point")}
\usage{
sitetype2text(sitetype = NULL)
}
\arguments{
\item{sitetype}{code like (if lowercase) latlon, shp, fips, fips_place, frs, echo, naics, sic, mact, epa_program_sel, epa_program_up as used in server
or some of which come from ejamit()$sitetype like latlon, fips, or shp}
}
\value{
phrase to use in report header or excel notes tab
}
\description{
helper to convert sitetype code ("latlon") to text describing it (" specified point")
}
\keyword{internal}
