% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedreport.R
\name{speedmessage}
\alias{speedmessage}
\title{estimate how long it will take to get buffer batch results}
\usage{
speedmessage(n, perhourslow = 1000, perhourfast = 12000, perhourguess = 6000)
}
\arguments{
\item{n}{number of points to buffer at}

\item{perhourslow}{n per hour if slow (conservative estimate of time needed)}

\item{perhourfast}{n per hour if fast}

\item{perhourguess}{n per hour best guess}
}
\description{
estimate how long it will take to get buffer batch results
}
\seealso{
\code{\link[=speedreport]{speedreport()}}
}
\keyword{internal}
