% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedreport.R
\name{speedreport}
\alias{speedreport}
\title{helper function that reports on how long buffering took}
\usage{
speedreport(start, end, n)
}
\arguments{
\item{start}{start time}

\item{end}{end time}

\item{n}{how many buffers were completed}
}
\value{
text string summarizing the speed
}
\description{
helper function that reports on how long buffering took
}
\seealso{
\code{\link[=speedmessage]{speedmessage()}}
}
\keyword{internal}
