% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_speedtest.R
\name{speedtable_expand}
\alias{speedtable_expand}
\title{Utility used by speedtest() and speedtable_summarize()}
\usage{
speedtable_expand(speedtable)
}
\arguments{
\item{speedtable}{must have columns called  points, miles, and perhr}
}
\description{
Utility used by speedtest() and speedtable_summarize()
}
\keyword{internal}
