% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_speedtest.R
\name{speedtable_summarize}
\alias{speedtable_summarize}
\title{utility used by speedtest()}
\usage{
speedtable_summarize(speedtable)
}
\arguments{
\item{speedtable}{from speedtest(), with columns named points and perhr}
}
\description{
utility used by speedtest()
}
\seealso{
\code{\link[=speedtest]{speedtest()}}
}
\keyword{internal}
