% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_speedtest.R
\name{speedtest_plot}
\alias{speedtest_plot}
\title{utility to plot output of speedtest(), rate of points analyzed per hour}
\usage{
speedtest_plot(x, ltype = "b", plotfile = NULL, secondsperthousand = FALSE)
}
\arguments{
\item{x}{table from speedtest()}

\item{ltype}{optional type of line for plot}

\item{plotfile}{optional path and filename of .png image file to save}
}
\value{
side effect is a plot. returns x but with seconds column added to it
}
\description{
utility to plot output of speedtest(), rate of points analyzed per hour
}
\seealso{
\code{\link[=speedtest]{speedtest()}}
}
\keyword{internal}
