% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{st2statename}
\alias{st2statename}
\title{utility to convert between statename and ST abbreviation}
\usage{
st2statename(ST)
}
\arguments{
\item{ST}{vector of state abbreviations like "GA"
(but can include state names)}
}
\value{
returns vector of state names as long as ST vector,
with NA for elements that are neither statename nor ST
}
\description{
utility to convert between statename and ST abbreviation
}
\examples{
st2statename(c("TX", 'dc', "Illinois"))
  
}
\keyword{internal}
