% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_from_latlon.R
\name{state_from_blockid_table}
\alias{state_from_blockid_table}
\title{state_from_blockid_table was used in some special cases e.g., in testpoints_n()}
\usage{
state_from_blockid_table(dt_with_blockid)
}
\arguments{
\item{dt_with_blockid}{}
}
\value{
vector of ST info like AK, CA, DE, etc.
}
\description{
given data.table with blockid column, get state abbreviation of each - not used?
}
\examples{
x = sample(blockpoints$blockid, 3)
state_from_blockid_table(blockpoints[blockid \%in\% x, ])[]
mapfast(blockpoints[blockid \%in\% x, ])

table(state_from_blockid_table(testoutput_getblocksnearby_10pts_1miles))
# unique(state_from_latlon(testpoints_10)$ST) # slow

all.equal(state_from_blockid(x), state_from_blockid_table(blockpoints[blockid \%in\% x, ]))

}
\keyword{internal}
