% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_per_site_for_doaggregate.R
\name{state_per_site_for_doaggregate}
\alias{state_per_site_for_doaggregate}
\title{identify the State each site is in, for doaggregate()}
\usage{
state_per_site_for_doaggregate(s2b, s2st)
}
\arguments{
\item{s2b}{like testoutput_getblocksnearby_100pts_1miles, or output of getblocksnearby()}

\item{s2st}{like testpoints_10, like input to ejamit() or to getblocksnearby()}
}
\value{
data.table
}
\description{
identify the State each site is in, for doaggregate()
}
\examples{
\dontrun{

 # cannot quickly id ST if a site spans 2+ states
 # not this is an unexported function:
 tail(state_from_s2b_bysite(testoutput_getblocksnearby_100pts_1miles))
 
 # using the closest block can id the wrong state:
 tail(state_from_nearest_block_bysite(testoutput_getblocksnearby_100pts_1miles))
 
 # getting the true state is slow if some sites span 2+ states:
 tail(
   state_per_site_for_doaggregate(
     testoutput_getblocksnearby_100pts_1miles, 
     testpoints_100
   ))
}

}
\keyword{internal}
