% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stateinfo.R
\docType{data}
\name{stateinfo}
\alias{stateinfo}
\title{stateinfo (DATA) data.frame of state abbreviations and state names (50+DC+PR; not AS, GU, MP, VI, UM)}
\format{
An object of class \code{data.frame} with 52 rows and 5 columns.
}
\usage{
stateinfo
}
\description{
52 rows and a few variables: ST is the 2-letter abbreviation,
statename is the State name, etc.
}
\details{
Created for EJAM by datacreate_stateinfo.R script

Also see \href{https://www.census.gov/programs-surveys/decennial-census/decade/2020/planning-management/release/2020-island-areas-data-products.html}{Census details}

column names: 'ST'  'statename'  'ftpname'   'FIPS.ST'   'REGION'

Some datasets lack PR. (72)

Many datasets lack these:  AS, GU, MP, VI (codes '60' '66' '69' '78')

Almost all datasets lack UM. (74)

\if{html}{\out{<div class="sourceCode">}}\preformatted{    72 PR                 Puerto Rico
    
    66 GU                        Guam 
    
    69 MP    Northern Mariana Islands  
    
    78 VI         U.S. Virgin Islands 
    
    74 UM U.S. Minor Outlying Islands  
}\if{html}{\out{</div>}}
}
\seealso{
\link{stateinfo2} for more columns
}
\keyword{datasets}
