% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{statename2st}
\alias{statename2st}
\title{utility to convert between statename and ST abbreviation}
\usage{
statename2st(statename)
}
\arguments{
\item{statename}{vector of state names (but can include state abbreviations)}
}
\value{
returns vector of ST abbreviations as long as statename vector,
with NA for elements that are neither statename nor ST
}
\description{
utility to convert between statename and ST abbreviation
}
\examples{
 EJAM:::statename2st(c("TX", 'dc', "Illinois"))
  
}
\keyword{internal}
