% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS_FUNCTIONS.R
\name{states_as_sites}
\alias{states_as_sites}
\title{FIPS - Analyze US States as if they were sites, to get EJ indicators summary}
\usage{
states_as_sites(fips)
}
\arguments{
\item{fips}{State FIPS vector, like c("01", "02") or
fips_state_from_state_abbrev(c("DE", "RI"))}
}
\value{
provides table similar to the output of getblocksnearby(),
data.table with one row per blockgroup in these states, or
all pairs of states fips - bgid, and ejam_uniq_id (1 through N) assigned to each state
but missing blockid and distance so not ready for doaggregate().
}
\description{
FIPS - Analyze US States as if they were sites, to get EJ indicators summary
}
\details{
This function provides one row per blockgroup.
\code{\link[=getblocksnearby_from_fips]{getblocksnearby_from_fips()}} provides one row per block.
See more below under "Value"
}
\examples{
  s2b <- states_as_sites(fips_state_from_state_abbrev(c("DE", "RI")))

}
