% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_states_shapefile.R
\docType{data}
\name{states_shapefile}
\alias{states_shapefile}
\title{This is used to figure out which state contains each point (facility/site).}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 56 rows and 15 columns.
}
\usage{
states_shapefile
}
\description{
This is used to figure out which state contains each point (facility/site).
}
\details{
This is used by \code{\link[=state_from_latlon]{state_from_latlon()}} to find which state is associated with each point
that the user wants to analyze. That is needed to report indicators in
the form of State-specific percentiles
(e.g., a score that is at the 80th percentile within Texas).
It is created by the package via a script at EJAM/data-raw/datacreate_states_shapefile.R
which downloads the data from Census Bureau.
}
\seealso{
seealso \code{\link[=state_from_latlon]{state_from_latlon()}} \code{\link[=get_blockpoints_in_shape]{get_blockpoints_in_shape()}}
}
\keyword{datasets}
