% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statestats_query.R
\name{statestats_means}
\alias{statestats_means}
\title{statestats_means - convenient way to see STATE MEANS of ENVIRONMENTAL and DEMOGRAPHIC indicators}
\usage{
statestats_means(
  ST = unique(EJAM::statestats$REGION),
  varnames = c(EJAM::names_e, EJAM::names_d, EJAM::names_d_subgroups_nh),
  PCTILES = "mean",
  dig = 4
)
}
\arguments{
\item{ST}{vector of state abbreviations, or USA}

\item{varnames}{names of columns in lookup table, like "proximity.rmp"}

\item{PCTILES}{vector of percentiles 0-100 and/or "mean"}

\item{dig}{digits to round to}
}
\description{
statestats_means - convenient way to see STATE MEANS of ENVIRONMENTAL and DEMOGRAPHIC indicators
}
