% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statestats_query.R
\name{statestats_means_bystates}
\alias{statestats_means_bystates}
\title{statestats_means - convenient way to see STATE MEANS of indicators for a list of states (that can have repeats)}
\usage{
statestats_means_bystates(
  ST = unique(EJAM::statestats$REGION),
  varnames = names_these,
  PCTILES = "mean"
)
}
\arguments{
\item{ST}{vector of 2-char ST abbrevs, or all values can be "USA" to get duplicate rows like found in ejamit()$results_bysite\link{, names_d_avg}}

\item{varnames}{vector of colnames in statestats table (indicator names)}

\item{PCTILES}{"mean"}
}
\value{
data.frame of state averages for those, one row per ST provided (can have repeats) and colnames are varnames.
}
\description{
Given a vector of 2-char ST abbrevs, and vector of colnames in statestats table (indicator names),
return data.frame of state averages
}
\keyword{internal}
