% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_gt_.R
\name{table4gt_from_scorevectors}
\alias{table4gt_from_scorevectors}
\title{DRAFT EXPERIMENTAL - attempt to make table more flexible / any indicators}
\usage{
table4gt_from_scorevectors(
  varnames_r = names_e,
  varnames_shown = fixcolnames(varnames_r, "r", "long"),
  value = as.vector(usastats_means(varnames_r)),
  state_avg = NULL,
  state_pctile = NULL,
  usa_avg = NULL,
  usa_pctile = NULL,
  state_ratio = NULL,
  usa_ratio = NULL,
  ST = "NY"
)
}
\arguments{
\item{varnames_r}{vector of variable names like names_d}

\item{varnames_shown}{vector like fixcolnames(names_d,'r','short')}

\item{value}{indicator values for a place or overall}

\item{state_avg}{indicator values average in State}

\item{state_pctile}{indicator values as State percentiles}

\item{usa_avg}{indicator values US average}

\item{usa_pctile}{indicator values as US percentiles}

\item{state_ratio}{indicator values as ratio to State average}

\item{usa_ratio}{indicator values as ratio to US average}

\item{ST}{State abbreviation like "NY"}
}
\value{
data.frame ready for table_gt_format_step2 ???
}
\description{
Based on just indicator names and a value for each, it tries to fill in the rest of a summary table's data.
and formats this as a data.frame ready for the next step
}
\seealso{
\code{\link[=table_gt_from_ejamit]{table_gt_from_ejamit()}} \code{\link[=table_gt_from_ejamit_overall]{table_gt_from_ejamit_overall()}} \code{\link[=table_gt_from_ejamit_1site]{table_gt_from_ejamit_1site()}} \code{\link[=table_validated_ejamit_row]{table_validated_ejamit_row()}} \code{\link[=table_gt_format_step1]{table_gt_format_step1()}} \code{\link[=table_gt_format_step2]{table_gt_format_step2()}}
}
\keyword{internal}
