% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_gt_.R
\name{table_gt_format_step2}
\alias{table_gt_format_step2}
\title{Format a table of demog or envt scores, percentiles, etc. to look similar to EJScreen report tables}
\usage{
table_gt_format_step2(
  df,
  type = c("demog", "envt")[1],
  my_cell_color = "#dce6f0",
  my_border_color = "#aaaaaa",
  digits_default = 2
)
}
\arguments{
\item{df}{A data frame from table_gt_format_step1

which is just a specific format of key EJAM results.

It has these columns (but it still works if the first two are omitted
and user-provided indicators are used - it just names them indicator 1, indicator 2, etc.):

varnames_r, varnames_shown, value, state_avg, state_pctile, usa_avg, usa_pctile

and one row per indicator, where varnames_shown are longer indicator names for use in report.

The sort order in this df is ignored!  Instead, the variables are shown in the same order as
shown in EJScreen reports, as recorded in map_headernames and checked here via varinfo(varnames_r, "reportsort"), etc.

Uses gt R package for formatting.}

\item{type}{string - must be demog or envt}

\item{my_cell_color}{color for table cell fill backgrounds,  can be given as string ('blue') or hex code ('#0070c0')}

\item{my_border_color}{color for table borders and boundaries, can be given as string ('blue') or hex code ('#0070c0')}

\item{digits_default}{number of digits to round to if not specified for a given indicator
(rounding info is drawn from map_headernames$decimals)}
}
\value{
a gt-style table with formatting to closely match EJScreen standard report formatting
}
\description{
Format a table of demog or envt scores, percentiles, etc. to look similar to EJScreen report tables
}
\seealso{
\code{\link[=table_gt_from_ejamit]{table_gt_from_ejamit()}}
}
