% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_gt_.R
\name{table_gt_from_ejamit}
\alias{table_gt_from_ejamit}
\title{Create a gt-format table of results from EJAM}
\usage{
table_gt_from_ejamit(ejamitoutput = NULL, type = c("demog", "envt")[1])
}
\arguments{
\item{ejamitoutput}{list of EJAM results formatted as in testoutput_ejamit_100pts_1miles,
as would be the output of ejamit()}

\item{type}{Must be "demog" or "envt" -- Creates one of these at a time}
}
\value{
Provides table in gt format from the R package called gt
}
\description{
Uses the list of results of ejamit()
}
\details{
See the R package called gt. Also see code that creates html tables from html template
and code that creates formatted spreadsheets like \code{\link[=table_xls_format]{table_xls_format()}}
}
\examples{
 table_gt_from_ejamit(testoutput_ejamit_100pts_1miles)

}
