% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_gt_.R
\name{table_gt_from_ejamit_1site}
\alias{table_gt_from_ejamit_1site}
\title{Create a formatted table of results for 1 site from EJAM}
\usage{
table_gt_from_ejamit_1site(...)
}
\arguments{
\item{...}{passed to \code{\link[=table_gt_from_ejamit_overall]{table_gt_from_ejamit_overall()}}}
}
\description{
Uses 1 row from the results_bysite part of ejamit() output
}
\examples{
 table_gt_from_ejamit_1site(testoutput_ejamit_100pts_1miles$results_bysite[ 1, ])
}
