% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_rounding_info.R
\name{table_rounding_info}
\alias{table_rounding_info}
\title{How many decimal places to round to for given variable(s)}
\usage{
table_rounding_info(var, varnametype = "rname")
}
\arguments{
\item{var}{vector of variable names such as c("pctlowinc", "pm") or c(names_d, names_d_subgroups)}

\item{varnametype}{which column of map_headernames to use when looking for var, like "rname" or "api" or "long"}
}
\value{
named vector same size as var, with var as names.
}
\description{
How many decimal places to round to for given variable(s)
}
\examples{
  table_rounding_info("pm")
  table_round(8.252345, "pm")
  table_round(8, "pm")
  
  cbind(table_rounding_info(names_all_r), fixcolnames(names_all_r, "r", "long"))
 
}
\seealso{
\code{\link[=table_signif_round_x100]{table_signif_round_x100()}} \code{\link[=table_signif]{table_signif()}} \code{\link[=table_round]{table_round()}} \code{\link[=table_x100]{table_x100()}}
}
\keyword{internal}
