% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_signif.R
\name{table_signif}
\alias{table_signif}
\title{Round numbers in a table, each column to appropriate number of significant digits}
\usage{
table_signif(dat, digits = NULL)
}
\arguments{
\item{dat}{data.frame or data.table of numbers}

\item{digits}{vector as long as number of columns in dat,
or use default which is to get the number of significant digits from
varinfo(colnames(dat), 'sigfigs')$sigfigs which gets it from
map_headernames dataset of metadata on EJAM/EJScreen indicators.}
}
\value{
table same size as dat
}
\description{
Round numbers in a table, each column to appropriate number of significant digits
}
\examples{
out <- testoutput_ejamit_10pts_1miles
mytable <- out$results_bysite[1:2, ..names_these]
table_signif_round_x100(mytable)
# same as this:
table_signif(
  table_round(
    table_x100(
      mytable, names_pct_as_fraction_ejamit
    )
  )
)


}
\seealso{
\code{\link[=table_signif_round_x100]{table_signif_round_x100()}} \code{\link[=table_signif]{table_signif()}} \code{\link[=table_round]{table_round()}} \code{\link[=table_x100]{table_x100()}}
}
\keyword{internal}
