% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_signif_round_x100.R
\name{table_signif_round_x100}
\alias{table_signif_round_x100}
\title{Clean table of EJAM numbers: signif digits, rounding, scaling as 0-100\%
Does table_signif() and table_round() and fix_pctcols_x100() in one call.}
\usage{
table_signif_round_x100(x, cnames = names_pct_as_fraction_ejamit)
}
\arguments{
\item{x}{data.frame or data.table}

\item{cnames}{use default when formatting output like ejamit()$results_bysite}
}
\value{
table of same shape as x
}
\description{
Clean table of EJAM numbers: signif digits, rounding, scaling as 0-100\%
Does table_signif() and table_round() and fix_pctcols_x100() in one call.
}
\examples{
out <- testoutput_ejamit_10pts_1miles$results_bysite
table_signif_round_x100(
  out[1:2, ..names_these]
)

}
\seealso{
\code{\link[=table_signif_round_x100]{table_signif_round_x100()}} \code{\link[=table_signif]{table_signif()}} \code{\link[=table_round]{table_round()}} \code{\link[=table_x100]{table_x100()}}
}
\keyword{internal}
