% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2table_tall.R
\name{table_tall_from_overall}
\alias{table_tall_from_overall}
\title{Format the results_overall part of the output of ejamit() or doaggregate()}
\usage{
table_tall_from_overall(results_overall, longnames = NULL)
}
\arguments{
\item{results_overall}{data.table of 1 row, from output of ejamit() or doaggregate()}

\item{longnames}{vector of names of variables in results_overall,
from output of ejamit() or doaggregate()}
}
\value{
data.frame with one indicator per row
}
\description{
Take a quick look at results in the RStudio console
}
\examples{
 table_tall_from_overall(testoutput_ejamit_10pts_1miles$results_overall)
 table_tall_from_overall(x$results_bysite[1, ])
 
}
\keyword{internal}
