% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_xls_from_ejam.R
\name{table_xls_from_ejam}
\alias{table_xls_from_ejam}
\title{Format the results of ejamit() for excel and optionally save .xlsx file}
\usage{
table_xls_from_ejam(
  ejamitout,
  fname = NULL,
  save_now = TRUE,
  overwrite = TRUE,
  launchexcel = FALSE,
  interactive_console = TRUE,
  ok2plot = TRUE,
  in.testing = FALSE,
  in.analysis_title = "EJAM analysis",
  react.v1_summary_plot = NULL,
  radius_or_buffer_in_miles = NULL,
  buffer_desc = "Selected Locations",
  radius_or_buffer_description =
    "Miles radius of circular buffer (or distance used if buffering around polygons)",
  hyperlink_colnames = c("EJScreen Report", "EJScreen Map", "ECHO report"),
  site_method = "",
  ...
)
}
\arguments{
\item{ejamitout}{output of \code{\link[=ejamit]{ejamit()}}}

\item{fname}{optional name or full path and name of file to save locally, like "out.xlsx"}

\item{save_now}{optional logical, whether to save as a .xlsx file locally or just return workbook object
that can later be written to .xlsx file using \code{\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook()}}}

\item{overwrite}{optional logical, passed to \code{\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook()}}}

\item{launchexcel}{optional logical, passed to \code{\link[=table_xls_format]{table_xls_format()}}, whether to launch browser to see spreadsheet immediately}

\item{interactive_console}{optional - should set to FALSE when used in code or server. If TRUE,
prompts RStudio user interactively asking where to save the downloaded file}

\item{ok2plot}{optional logical, passed to  \code{\link[=table_xls_format]{table_xls_format()}}, whether safe to try and plot or set FALSE if debugging plot problems}

\item{in.testing}{optional logical}

\item{in.analysis_title}{optional title as character string}

\item{react.v1_summary_plot}{optional - a plot object}

\item{radius_or_buffer_in_miles}{optional radius in miles}

\item{buffer_desc}{description of location to use in labels, like "Selected Locations"}

\item{radius_or_buffer_description}{optional text phrase describing places analyzed}

\item{hyperlink_colnames}{optional names of columns with URLs}

\item{site_method}{site selection method, such as NAICS, FRS, SHP, latlon,
optional site method parameter used to create a more specific title with create_filename}

\item{...}{optional additional parameters passed to \code{\link[=table_xls_format]{table_xls_format()}}, such as
heatmap_colnames, heatmap_cuts, heatmap_colors, etc.}
}
\value{
returns a workbook object for use by openxlsx::saveWorkbook(wb_out, pathname)
or returns just the full path/file name of where it was saved if save_now = TRUE
}
\description{
Almost identical to \code{\link[=ejam2excel]{ejam2excel()}}
}
\examples{
\dontrun{
  table_xls_from_ejam(testoutput_ejamit_10pts_1miles)
  }
}
\keyword{internal}
