% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablefixed.R
\name{tablefixed}
\alias{tablefixed}
\title{Table of counts of integer values zero through maxbin}
\usage{
tablefixed(x, maxbin = NULL)
}
\arguments{
\item{x}{vector of integers, like counts, that can include 0}

\item{maxbin}{highest integer among x, or number of bins}
}
\value{
summary table
}
\description{
Like tabulate or table, sort of, but includes zero unlike tabulate,
and lets you ensure results include every integer 0 through maxbin,
so you can, for example, easily combine tables of counts where some
did not include all integers.
}
\details{
There is likely a more efficient way to do this in some existing package,
but this is useful and fast enough.

When using a dataset like EJScreen with 13 indicators of interest,
and counting how many of the 13 are above various cutpoints,
there may be zero rows that have exactly 8 above some cutoff, for example.

This function makes it easier to combine those tables into a summary
where 0-13 are in each table while table() would only return integers
that came up in a given case (for one cutoff).
}
\seealso{
colcounter_summary()
}
