% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_test_address_table.R
\docType{data}
\name{test_address_table}
\alias{test_address_table}
\alias{test_address_parts1}
\alias{test_addresses2}
\alias{test_address_table_9}
\alias{test_addresses_9}
\alias{test_address_table_goodnames}
\alias{test_address_table_withfull}
\title{datasets for trying address-related functions}
\format{
An object of class \code{data.frame} with 2 rows and 6 columns.
}
\usage{
test_address_table
}
\description{
datasets for trying address-related functions (test_address_parts1, test_addresses2, test_address_table_9, test_addresses_9, test_address_table, test_address_table_goodnames, test_address_table_withfull)
}
\keyword{datasets}
