% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_test_regid.R
\docType{data}
\name{test_regid}
\alias{test_regid}
\title{test_regid (DATA) test data, vector of EPA FRS Registry ID numbers}
\description{
test_regid (DATA) test data, vector of EPA FRS Registry ID numbers
}
\details{
For testing, e.g.,

frs_from_regid(test_regid)

mapfast(  frs_from_regid(test_regid)  )
}
