% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_test_xtrac.R
\docType{data}
\name{test_xtrac}
\alias{test_xtrac}
\title{for internal use}
\format{
An object of class \code{numeric} of length 25.
}
\usage{
test_xtrac
}
\description{
for internal use
}
\keyword{datasets}
