% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{testdata}
\alias{testdata}
\title{utility to show dir_tree of available files in testdata folders
See list of samples of input files to try in EJAM, and output examples from EJAM functions}
\usage{
testdata(installed = TRUE)
}
\arguments{
\item{installed}{If you are a developer who has the local source package,
you can set this parameter to FALSE if you want to work with the
local source package version of the testdata folders
rather than the locally installed version.}
}
\value{
path to local testdata folder comes with the EJAM package
}
\description{
utility to show dir_tree of available files in testdata folders
See list of samples of input files to try in EJAM, and output examples from EJAM functions
}
\keyword{internal}
