% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{testdatafolder}
\alias{testdatafolder}
\title{utility to show path to testdata folders
see folder that has samples of input files to try in EJAM, and output examples from EJAM functions}
\usage{
testdatafolder(installed = TRUE)
}
\arguments{
\item{installed}{If you are a developer who has the local source package,
you can set this parameter to FALSE if you want to work with the
local source package version of the testdata folders
rather than the locally installed version.}
}
\value{
path to local testdata folder comes with the EJAM package
}
\description{
utility to show path to testdata folders
see folder that has samples of input files to try in EJAM, and output examples from EJAM functions
}
\keyword{internal}
