% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testids_registry_id.R
\docType{data}
\name{testids_registry_id}
\alias{testids_registry_id}
\title{test data, EPA Facility Registry ID numbers to try using}
\format{
An object of class \code{numeric} of length 8.
}
\usage{
testids_registry_id
}
\description{
test data, EPA Facility Registry ID numbers to try using
}
\details{
Just for convenience, installed with the package
}
\keyword{datasets}
