% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_fips_blockgroups.R
\docType{data}
\name{testinput_fips_blockgroups}
\alias{testinput_fips_blockgroups}
\title{testinput_fips_blockgroups dataset}
\format{
An object of class \code{character} of length 14.
}
\usage{
testinput_fips_blockgroups
}
\description{
Census FIPS codes vector as example of input, e.g., ejamit(fips = testinput_fips_blockgroups)
}
\keyword{datasets}
