% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_fips_cities.R
\docType{data}
\name{testinput_fips_cities}
\alias{testinput_fips_cities}
\title{testinput_fips_cities dataset}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
testinput_fips_cities
}
\description{
Census FIPS codes vector as example of input, e.g., ejamit(fips = testinput_fips_cities)
}
\keyword{datasets}
