% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_fips_counties.R
\docType{data}
\name{testinput_fips_counties}
\alias{testinput_fips_counties}
\title{testinput_fips_counties dataset}
\format{
An object of class \code{character} of length 3.
}
\usage{
testinput_fips_counties
}
\description{
Census FIPS codes vector as example of input, e.g., ejamit(fips = testinput_fips_counties)
}
\keyword{datasets}
