% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_fips_states.R
\docType{data}
\name{testinput_fips_states}
\alias{testinput_fips_states}
\title{testinput_fips_states dataset}
\format{
An object of class \code{character} of length 2.
}
\usage{
testinput_fips_states
}
\description{
Census FIPS codes vector as example of input, e.g., ejamit(fips = testinput_fips_states)
}
\keyword{datasets}
