% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_fips_tracts.R
\docType{data}
\name{testinput_fips_tracts}
\alias{testinput_fips_tracts}
\title{testinput_fips_tracts dataset}
\format{
An object of class \code{character} of length 8.
}
\usage{
testinput_fips_tracts
}
\description{
Census FIPS codes vector as example of input, e.g., ejamit(fips = testinput_fips_tracts)
}
\keyword{datasets}
