% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_mact.R
\docType{data}
\name{testinput_mact}
\alias{testinput_mact}
\title{testinput_mact dataset}
\format{
An object of class \code{character} of length 1.
}
\usage{
testinput_mact
}
\description{
MACT subpart code as example of input, e.g., latlon_from_mactsubpart(testinput_mact)
}
\keyword{datasets}
