% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_naics.R
\docType{data}
\name{testinput_naics}
\alias{testinput_naics}
\title{testinput_naics dataset}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
testinput_naics
}
\description{
NAICS codes vector as example of input, e.g., latlon_from_naics(testinput_naics)
}
\keyword{datasets}
