% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_program_name.R
\docType{data}
\name{testinput_program_name}
\alias{testinput_program_name}
\title{testinput_program_name dataset}
\format{
An object of class \code{character} of length 2.
}
\usage{
testinput_program_name
}
\description{
EPA program name codes vector as example of input, e.g., latlon_from_program(testinput_program_name)
}
\keyword{datasets}
