% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_shapes_2.R
\docType{data}
\name{testinput_shapes_2}
\alias{testinput_shapes_2}
\title{testinput_shapes_2 dataset}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2 rows and 18 columns.
}
\usage{
testinput_shapes_2
}
\description{
Sample shapefile/polygon data as spatial data.frame
}
\seealso{
same as \link{testshapes_2}
}
\keyword{datasets}
