% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testinput_sic.R
\docType{data}
\name{testinput_sic}
\alias{testinput_sic}
\title{testinput_sic dataset}
\format{
An object of class \code{character} of length 3.
}
\usage{
testinput_sic
}
\description{
SIC codes vector as example of input to e.g., latlon_from_sic(testinput_sic)
}
\keyword{datasets}
