% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testoutput_doaggregate_1000pts_1miles.R
\docType{data}
\name{testoutput_doaggregate_1000pts_1miles}
\alias{testoutput_doaggregate_1000pts_1miles}
\title{test output of doaggregate()}
\format{
An object of class \code{list} of length 5.
}
\usage{
testoutput_doaggregate_1000pts_1miles
}
\description{
test output of doaggregate()
}
\details{
This is the output of doaggregate(testoutput_getblocksnearby_1000pts_1miles, sites2states_or_latlon = testpoints_1000, radius = 1, include_ejindexes = TRUE)
}
\seealso{
\code{\link[=doaggregate]{doaggregate()}} \code{\link[=ejamit]{ejamit()}} \link{testoutput_getblocksnearby_1000pts_1miles} \link{testpoints_1000}
}
\keyword{datasets}
