% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testoutput_ejamit_100pts_1miles.R
\docType{data}
\name{testoutput_ejamit_100pts_1miles}
\alias{testoutput_ejamit_100pts_1miles}
\title{test output of ejamit()}
\format{
An object of class \code{list} of length 8.
}
\usage{
testoutput_ejamit_100pts_1miles
}
\description{
test output of ejamit()
}
\details{
This is the output of ejamit(testpoints_100, radius = 1, include_ejindexes = TRUE)
}
\seealso{
\code{\link[=doaggregate]{doaggregate()}} \code{\link[=ejamit]{ejamit()}} \link{testoutput_doaggregate_100pts_1miles} and \link{testpoints_100}
}
\keyword{datasets}
