% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testoutput_ejamit_fips_counties.R
\docType{data}
\name{testoutput_ejamit_fips_counties}
\alias{testoutput_ejamit_fips_counties}
\title{testoutput_ejamit_fips_counties dataset}
\format{
An object of class \code{list} of length 8.
}
\usage{
testoutput_ejamit_fips_counties
}
\description{
This is the output of ejamit(fips = testinput_fips_counties)
}
\seealso{
\code{\link[=ejamit]{ejamit()}} \code{\link[=testdata]{testdata()}}
}
\keyword{datasets}
