% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testoutput_ejamit_shapes_2.R
\docType{data}
\name{testoutput_ejamit_shapes_2}
\alias{testoutput_ejamit_shapes_2}
\title{testoutput_ejamit_shapes_2 dataset}
\format{
An object of class \code{list} of length 8.
}
\usage{
testoutput_ejamit_shapes_2
}
\description{
This is the output of ejamit(shapefile = testinput_shapes_2, radius = 0)
}
\seealso{
\code{\link[=ejamit]{ejamit()}} \code{\link[=testdata]{testdata()}}
}
\keyword{datasets}
