% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testoutput_ejscreenit_10pts_1miles.R
\docType{data}
\name{testoutput_ejscreenit_10pts_1miles}
\alias{testoutput_ejscreenit_10pts_1miles}
\title{test output of ejscreenit(), using the EJScreen API}
\description{
test output of ejscreenit(), using the EJScreen API
}
\details{
This is the output of

ejscreenit(

testpoints_10, radius = 1,

nosave = T, nosee = T, interactiveprompt = F, calculate_ratios = T
)

testoutput_ejscreenit_10pts_1miles$table

Also see

testoutput_ejamit_10pts_1miles$results_bysite
}
