% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testoutput_ejscreenit_50.R
\docType{data}
\name{testoutput_ejscreenit_50}
\alias{testoutput_ejscreenit_50}
\title{test data examples of output from \code{\link[=ejscreenit]{ejscreenit()}} using testpoints_50, radius = 1}
\format{
An object of class \code{list} of length 3.
}
\usage{
testoutput_ejscreenit_50
}
\description{
test data examples of output from \code{\link[=ejscreenit]{ejscreenit()}} using testpoints_50, radius = 1
}
\details{
Just for convenience, installed with the package.
A list of outputs, named 'table', 'map', and 'plot'
}
\seealso{
\link{testpoints_5} \link{testpoints_50} \link{testpoints_500}

\link{testoutput_ejscreenit_5} \link{testoutput_ejscreenit_50} \link{testoutput_ejscreenit_500}
}
\keyword{datasets}
