% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/data_testoutput_ejscreenRESTbroker_1pts_1miles.R
\docType{data}
\name{testoutput_ejscreenRESTbroker_1pts_1miles}
\alias{testoutput_ejscreenRESTbroker_1pts_1miles}
\title{test data, output from this function}
\format{
An object of class \code{response} of length 10.
}
\usage{
testoutput_ejscreenRESTbroker_1pts_1miles
}
\description{
test data, output from this function
}
\details{
Just for convenience, installed with the package,
the equivalent of results of
\code{ejscreenRESTbroker(lon = testpoints_5$lon[1], lat = testpoints_5$lat[1], radius = 1)}
}
\seealso{
\link{testpoints_5} \link{testpoints_50} \link{testpoints_500}

\link{testoutput_ejscreenit_5} \link{testoutput_ejscreenit_50} \link{testoutput_ejscreenit_500}
}
\keyword{datasets}
