% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testoutput_getblocksnearby_10pts_1miles.R
\docType{data}
\name{testoutput_getblocksnearby_10pts_1miles}
\alias{testoutput_getblocksnearby_10pts_1miles}
\title{test output of getblocksnearby(), and is an input to doaggregate()}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1914 rows and 6 columns.
}
\usage{
testoutput_getblocksnearby_10pts_1miles
}
\description{
test output of getblocksnearby(), and is an input to doaggregate()
}
\details{
This is the output of getblocksnearby(testpoints_10, radius = 1)
}
\seealso{
\code{\link[=getblocksnearby]{getblocksnearby()}}  \code{\link[=doaggregate]{doaggregate()}}  \link{testpoints_10}
}
\keyword{datasets}
