% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testpoints_10.R
\docType{data}
\name{testpoints_10}
\alias{testpoints_10}
\title{test points data.frame with columns sitenumber, lat, lon}
\format{
An object of class \code{data.frame} with 10 rows and 3 columns.
}
\usage{
testpoints_10
}
\description{
test points data.frame with columns sitenumber, lat, lon
}
\keyword{datasets}
