% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testpoints_100_dt.R
\docType{data}
\name{testpoints_100_dt}
\alias{testpoints_100_dt}
\title{test points data.frame with columns sitenumber, lat, lon}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 100 rows and 4 columns.
}
\usage{
testpoints_100_dt
}
\description{
test points data.frame with columns sitenumber, lat, lon
}
\keyword{datasets}
