% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testpoints_500.R
\docType{data}
\name{testpoints_500}
\alias{testpoints_500}
\title{test points data.frame with columns sitenumber, lat, lon}
\format{
An object of class \code{data.frame} with 500 rows and 4 columns.
}
\usage{
testpoints_500
}
\description{
test points data.frame with columns sitenumber, lat, lon
}
\keyword{datasets}
