% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_testpoints_bad.R
\docType{data}
\name{testpoints_bad}
\alias{testpoints_bad}
\title{test points data.frame with columns note, lat, lon}
\description{
test points data.frame with columns note, lat, lon
}
\details{
examples of test points for testing functions that need lat lon,
with some invalid or tricky cases like rural (no blocks nearby), outside US,
on edge of two states, NA values, etc.
}
