% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dupenames_etc.R
\name{unshared}
\alias{unshared}
\title{UTILITY - see what is only in x or y but not both - just like setdiff except for y,x and also x,y

setdiff2 aka unshared just shows which elements are in one and only one of the sets x and y}
\usage{
unshared(x, y)
}
\description{
UTILITY - see what is only in x or y but not both - just like setdiff except for y,x and also x,y

setdiff2 aka unshared just shows which elements are in one and only one of the sets x and y
}
\examples{
unshared(1:4, 3:8)

}
\keyword{internal}
