% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part2.R
\name{url_4table}
\alias{url_4table}
\title{URL functions - Create URLs in columns, for EJAM}
\usage{
url_4table(
  lat,
  lon,
  radius = NULL,
  regid = NULL,
  fips,
  wherestr = "",
  namestr = NULL,
  shapefile = NULL,
  as_html = TRUE,
  ...
)
}
\arguments{
\item{lat}{vector of latitudes}

\item{lon}{vector of longitudes}

\item{radius}{vector of values for radius in miles}

\item{regid}{optional vector of FRS registry IDs if available to use to create links
to detailed ECHO facility reports}

\item{fips}{vector of FIPS codes if relevant, to use instead of lat lon,
Passed to \code{\link[=url_ejscreen_report]{url_ejscreen_report()}} as areaid
Note that nearly half of all county fips codes are impossible to distinguish from
5-digit zipcodes because the same numbers are used for both purposes.}

\item{wherestr}{optional because inferred from fips if provided.
Passed to \code{\link[=url_ejscreenmap]{url_ejscreenmap()}} and can be name of city, county, state like
from fips2name(201090), or "new rochelle, ny" or "AK"
or even a zip code, but NOT a fips code! (for FIPS, use the fips parameter instead).
Note that nearly half of all county fips codes are impossible to distinguish from
5-digit zipcodes because the same numbers are used for both purposes.}

\item{namestr}{passed to \code{\link[=url_ejscreen_report]{url_ejscreen_report()}}}

\item{shapefile}{not implemented}

\item{as_html}{logical, optional.
passed to \code{\link[=url_ejscreen_report]{url_ejscreen_report()}} and \code{\link[=url_ejscreenmap]{url_ejscreenmap()}}}

\item{...}{passed to \code{\link[=url_ejscreen_report]{url_ejscreen_report()}} such as areaid="0201090" for a city fips}
}
\value{
list of data.frames to append to the list of data.frames created by
\code{\link[=ejamit]{ejamit()}} or \code{\link[=doaggregate]{doaggregate()}},

\verb{list(results_bysite = results_bysite, }
\verb{    results_overall = results_overall,}
\verb{      newcolnames=newcolnames)}
}
\description{
Could start to use this in server and ejamit(), and already used in table_xls_format()
}
\details{
used in \code{\link[=table_xls_format]{table_xls_format()}}
}
\seealso{
\code{\link[=url_ejscreen_report]{url_ejscreen_report()}} \code{\link[=url_ejscreenmap]{url_ejscreenmap()}} \code{\link[=url_echo_facility_webpage]{url_echo_facility_webpage()}} \code{\link[=urls_clusters_and_sort_cols]{urls_clusters_and_sort_cols()}}
}
\keyword{internal}
