% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_locate_by_id.R
\name{url_by_id}
\alias{url_by_id}
\title{Get the URLs to use to query FRS API to find EPA facilities by ID}
\usage{
url_by_id(id, type = "frs", ...)
}
\arguments{
\item{id}{vector of one or more character strings with pgm_sys_id or registry_id
values (all need to be the same type, as defined by type parameter).
Program ids are like "VA0088986" and frs ids are like "110015787683"}

\item{type}{one word, applies to all. default is frs but can be program or the word other.}

\item{...}{appended to the end of the URL as-is, useful if type is other, for example}
}
\value{
vector of URLs as strings, same length as id parameter
}
\description{
This uses an API to find sites, but it is faster to look in a table
if that FRS dataset is already loaded in an app, for example.
}
\details{
For details on FRS API,
see https://www.epa.gov/frs/frs-rest-services
and examples at https://www.epa.gov/frs/frs-rest-services#ex1
and more at https://www.epa.gov/frs/frs-rest-services#appendixa
For example:
https://frs-public.epa.gov/ords/frs_public2/frs_rest_services.get_facilities?pgm_sys_id=VA0088986
https://frs-public.epa.gov/ords/frs_public2/frs_rest_services.get_facilities?registry_id=110010912496
Note: API URL for internal use at EPA appears to be different than public one?
}
\examples{
 \dontrun{
    url_by_id(testids_registry_id)
    browseURL(url_by_id(testids_registry_id)[1])
}

}
\seealso{
locate_by_id()
}
\keyword{internal}
