% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part2.R
\name{url_countyhealthrankings}
\alias{url_countyhealthrankings}
\title{URL functions - Get URLs of useful report(s) on Counties from countyhealthrankings.org}
\usage{
url_countyhealthrankings(fips, year = 2023)
}
\arguments{
\item{fips}{vector of fips codes of counties, 5 characters each, like "10003"}

\item{year}{2023}
}
\value{
vector of URLs
}
\description{
URL functions - Get URLs of useful report(s) on Counties from countyhealthrankings.org
}
\examples{
 url_countyhealthrankings(fips_counties_from_state_abbrev("DE"))
 # browseURL(url_countyhealthrankings(fips_counties_from_state_abbrev("DE"))[1])

}
