% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part1.R
\name{url_ejscreen_acs_report}
\alias{url_ejscreen_acs_report}
\title{Get URLs of EJScreen older ACS reports - will be phased out}
\usage{
url_ejscreen_acs_report(lon, lat, radius, as_html = FALSE, linktext)
}
\arguments{
\item{lon}{one or more longitudes}

\item{lat}{one or more latitudes}

\item{radius}{miles radius}

\item{as_html}{Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example}

\item{linktext}{used as text for hyperlinks, if supplied and as_html=TRUE}
}
\value{
URL(s)
}
\description{
Get URL(s) for EJScreen ACS report on residents near given point(s)
}
\details{
NOT USED BY APP NOW THAT COMMUNITY REPORT EXISTS
}
\seealso{
\code{\link[=url_ejscreen_report]{url_ejscreen_report()}}  \code{\link[=url_ejscreen_acs_report]{url_ejscreen_acs_report()}}   \code{\link[=url_ejscreenmap]{url_ejscreenmap()}}
\code{\link[=url_echo_facility_webpage]{url_echo_facility_webpage()}} \code{\link[=url_frs_report]{url_frs_report()}}  \code{\link[=url_enviromapper]{url_enviromapper()}}  \code{\link[=url_envirofacts_data]{url_envirofacts_data()}}
}
\keyword{internal}
