% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part1.R
\name{url_ejscreen_report}
\alias{url_ejscreen_report}
\title{Get URLs of EJScreen reports}
\usage{
url_ejscreen_report(
  lat = "",
  lon = "",
  radius = "",
  as_html = FALSE,
  linktext,
  mobile = FALSE,
  areatype = "",
  areaid = "",
  namestr = "",
  shapefile = NULL,
  wkid = 4326,
  unit = 9035,
  f = "report",
  interactiveprompt = FALSE
)
}
\arguments{
\item{lat}{one or more latitudes (or a table with lat, lon columns, or filepath with that, or omit to interactively select file)}

\item{lon}{one or more longitudes (or omitted -- see lat parameter details)}

\item{radius}{miles radius}

\item{as_html}{Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example}

\item{linktext}{used as text for hyperlinks, if supplied and as_html=TRUE}

\item{mobile}{If TRUE, provides URL for the mobile browser version, not desktop version}

\item{areatype}{passed as areatype= in API, inferred if not provided but areaid is provided}

\item{areaid}{fips codes if used,  passed as areaid=  in API, can be FIPS for blockgroups, tracts, counties.}

\item{namestr}{The character string text to show on the report as the name of the place}

\item{shapefile}{not implemented}

\item{wkid}{default is 4326 -WGS84 - World Geodetic System 1984, used in GPS - see (https://epsg.io/4326)}

\item{unit}{default is 9035 which means miles; for kilometers use 9036}

\item{f}{can be "report" or "pjson" or "json"}

\item{interactiveprompt}{passed to \code{\link[=sitepoints_from_any]{sitepoints_from_any()}}}
}
\value{
URL(s)
}
\description{
Get URL(s) for EJScreen standard report on residents near given point(s)
}
\details{
Also see \code{\link[=ejscreenRESTbroker]{ejscreenRESTbroker()}} and \code{\link[=ejscreenapi1]{ejscreenapi1()}}

and (https://ejscreen.epa.gov/mapper/EJAPIinstructions.pdf)

and (https://ejscreen.epa.gov/mapper/ejscreenapi1.html)
}
\seealso{
\code{\link[=url_ejscreen_report]{url_ejscreen_report()}}  \code{\link[=url_ejscreen_acs_report]{url_ejscreen_acs_report()}}   \code{\link[=url_ejscreenmap]{url_ejscreenmap()}}
\code{\link[=url_echo_facility_webpage]{url_echo_facility_webpage()}} \code{\link[=url_frs_report]{url_frs_report()}}  \code{\link[=url_enviromapper]{url_enviromapper()}}  \code{\link[=url_envirofacts_data]{url_envirofacts_data()}}
}
