% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part1.R
\name{url_enviromapper}
\alias{url_enviromapper}
\title{Get URLs of EnviroMapper reports}
\usage{
url_enviromapper(lon, lat, as_html = FALSE, linktext, zoom = 13)
}
\arguments{
\item{lon}{one longitude}

\item{lat}{one latitude}

\item{as_html}{Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example}

\item{linktext}{used as text for hyperlinks, if supplied and as_html=TRUE}

\item{zoom}{initial map zoom extent, with smaller numbers if zoomed in closer}
}
\value{
URL of one webpage (that launches the mapping tool)
}
\description{
Get URL(s) for EnviroMapper web-based tool, to open map at specified point location(s)
}
\details{
EnviroMapper lets you view EPA-regulated facilities and other information on a map, given the lat,lon
}
\seealso{
\code{\link[=url_ejscreen_report]{url_ejscreen_report()}}  \code{\link[=url_ejscreen_acs_report]{url_ejscreen_acs_report()}}   \code{\link[=url_ejscreenmap]{url_ejscreenmap()}}
\code{\link[=url_echo_facility_webpage]{url_echo_facility_webpage()}} \code{\link[=url_frs_report]{url_frs_report()}}  \code{\link[=url_enviromapper]{url_enviromapper()}}  \code{\link[=url_envirofacts_data]{url_envirofacts_data()}}
}
