% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part1.R
\name{url_frs_report}
\alias{url_frs_report}
\title{Get URLs of FRS reports}
\usage{
url_frs_report(regid, as_html = FALSE, linktext)
}
\arguments{
\item{regid}{one or more EPA FRS Registry IDs.}

\item{as_html}{Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example}

\item{linktext}{used as text for hyperlinks, if supplied and as_html=TRUE}
}
\value{
URL(s)
}
\description{
Get URL(s) for reports on facilities from EPA FRS (facility registry service)
}
\examples{
\dontrun{
browseURL(url_frs_report(testids_registry_id)[1])
}
url_frs_report(testids_registry_id)

}
\seealso{
\code{\link[=url_ejscreen_report]{url_ejscreen_report()}}  \code{\link[=url_ejscreen_acs_report]{url_ejscreen_acs_report()}}   \code{\link[=url_ejscreenmap]{url_ejscreenmap()}}
\code{\link[=url_echo_facility_webpage]{url_echo_facility_webpage()}} \code{\link[=url_frs_report]{url_frs_report()}}  \code{\link[=url_enviromapper]{url_enviromapper()}}  \code{\link[=url_envirofacts_data]{url_envirofacts_data()}}
}
