% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part2.R
\name{url_get_eparest_chunked_by_id}
\alias{url_get_eparest_chunked_by_id}
\title{URL functions - url_get_eparest_chunked_by_id - experimental/ work in progress: in chunks, get ACS data or Block weights nearby via EPA API}
\usage{
url_get_eparest_chunked_by_id(objectIds, chunksize = 200, ...)
}
\arguments{
\item{objectIds}{see API}

\item{chunksize}{see API}

\item{...}{passed to url_getacs_epaquery()}
}
\value{
a table
}
\description{
URL functions - url_get_eparest_chunked_by_id - experimental/ work in progress: in chunks, get ACS data or Block weights nearby via EPA API
}
\keyword{internal}
