% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part2.R
\name{url_getacs_epaquery}
\alias{url_getacs_epaquery}
\title{URL functions - url_getacs_epaquery - experimental/ work in progress: get ACS data via EPA API (for <200 places)}
\usage{
url_getacs_epaquery(
  objectIds = 1:3,
  servicenumber = 7,
  outFields = NULL,
  returnGeometry = FALSE,
  justurl = FALSE,
  ...
)
}
\arguments{
\item{objectIds}{see API}

\item{servicenumber}{see API}

\item{outFields}{see API. eg "STCNTRBG","TOTALPOP","PCT_HISP",}

\item{returnGeometry}{see API}

\item{justurl}{if TRUE, returns url instead of default making API request}

\item{...}{passed to url_getacs_epaquery_chunked()}
}
\value{
table
}
\description{
uses ACS2019 rest services ejscreen ejquery MapServer 7

Documentation of format and examples of input parameters:

\url{https://geopub.epa.gov/arcgis/sdk/rest/index.html#/Query_Map_Service_Layer/02ss0000000r000000/}
}
\examples{
 url_getacs_epaquery(justurl=TRUE)

}
\keyword{internal}
