% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part1.R
\name{url_linkify}
\alias{url_linkify}
\title{utility to make html link from URL}
\usage{
url_linkify(url, text)
}
\arguments{
\item{url}{string that is URL}

\item{text}{string that is label}
}
\value{
url_linkify('epa.gov','EPA') returns \verb{"<a href=\\"epa.gov\\", target=\\"_blank\\">EPA</a>"}
}
\description{
Convert URL to HTML link that opens in new tab
}
\keyword{internal}
