% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_google.R
\name{url_map_google}
\alias{url_map_google}
\title{Map - get URL(s) that would display map(s) on maps.google.com}
\usage{
url_map_google(lat, lon, zoom = 13, point = TRUE)
}
\arguments{
\item{lat}{\itemize{
\item Anything that can be handled by \code{\link[=sitepoints_from_any]{sitepoints_from_any()}}.
Leave unspecified to interactively browse to a .xlsx file that has lat,lon columns,
or lat can be a data.frame with lat,lon column names in which case longitude should not be provided,
such as lat = testpoints\link{1,}, or lat and lon can be separately provided as vectors.
}}

\item{lon}{longitude, or omit this parameter to provide points as the first parameter.}

\item{zoom}{zoomed out value could be 3 or 5, zoomed in default is 12}

\item{point}{logical, if TRUE, URL will have a marker at the point
and zoom parameter is ignored. Otherwise just the map.}
}
\value{
URL(s) vector one per point
}
\description{
Map - get URL(s) that would display map(s) on maps.google.com
}
\seealso{
\code{\link[=map_google]{map_google()}}
}
\keyword{internal}
