% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part2.R
\name{url_naics.com}
\alias{url_naics.com}
\title{URL functions - url_naics.com - Get URL for page with info about industry sectors by text query term}
\usage{
url_naics.com(query, as_html = FALSE, linktext)
}
\arguments{
\item{query}{string query term like "gasoline" or "copper smelting"}

\item{as_html}{Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example}

\item{linktext}{used as text for hyperlinks, if supplied and as_html=TRUE}
}
\value{
URL as string
}
\description{
See (https://naics.com) for more information on NAICS codes
}
