% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part1.R
\name{url_online}
\alias{url_online}
\title{utility - check if URL available, such as if an API is online or offline}
\usage{
url_online(url)
}
\arguments{
\item{url}{the URL to check}
}
\value{
TRUE or FALSE (but NA if no internet connection seems to be available at all)
}
\description{
utility - check if URL available, such as if an API is online or offline
}
\keyword{internal}
