% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URL_FUNCTIONS_part1.R
\name{urls_clusters_and_sort_cols}
\alias{urls_clusters_and_sort_cols}
\title{Add Links to ejscreenapi output, to launch EJScreen report for given point(s)}
\usage{
urls_clusters_and_sort_cols(results_table)
}
\arguments{
\item{results_table}{from \code{\link[=ejscreenapi]{ejscreenapi()}} function for example}
}
\value{
the input table but with extra columns
}
\description{
Add or update, and reorder, columns with results
}
\details{
Was used for \code{\link[=ejscreenapi_plus]{ejscreenapi_plus()}} not for \code{\link[=ejamit]{ejamit()}}, to create weblinks to maps, for output table.
\enumerate{
\item Adds weblinks in column.
\item Adds a column to flag sites that are close to other sites, and
\item Puts certain columns first.
}
}
\seealso{
\code{\link[=url_4table]{url_4table()}} for the EJAM version. \code{\link[=url_ejscreenmap]{url_ejscreenmap()}} \code{\link[=distance_near_eachother]{distance_near_eachother()}}
}
\keyword{internal}
