% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_usastats.R
\docType{data}
\name{usastats}
\alias{usastats}
\title{usastats (DATA) data.frame of 100 percentiles and means}
\format{
An object of class \code{data.frame} with 102 rows and 67 columns.
}
\usage{
usastats
}
\description{
data.frame of 100 percentiles and means (about 100 rows)
in the USA overall, across all locations (e.g., block groups in \link{blockgroupstats})
for a set of indicators such as percent low income.
Each column is one indicator (or specifies the percentile).

This should be similar to the lookup tables in the gdb on the FTP site of EJScreen,
except it also has data for the demographic race/ethnicity subgroups.

For details on how the table was made, see source package files
EJAM/data-raw/datacreate_usastats2.32_add_dsubgroups.R and
EJAM/data-raw/datacreate_usastats2.32.R

See also \link{statestats}
}
\keyword{datasets}
