% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statestats_query.R
\name{usastats_means}
\alias{usastats_means}
\title{usastats_means - convenient way to see US MEANS of ENVIRONMENTAL and DEMOGRAPHIC indicators}
\usage{
usastats_means(
  varnames = c(EJAM::names_e, EJAM::names_d, EJAM::names_d_subgroups_nh),
  PCTILES = NULL,
  dig = 4
)
}
\arguments{
\item{varnames}{names of columns in lookup table, like "proximity.rmp"}

\item{PCTILES}{vector of percentiles 0-100 and/or "mean"}

\item{dig}{how many digits to round to}
}
\description{
usastats_means - convenient way to see US MEANS of ENVIRONMENTAL and DEMOGRAPHIC indicators
}
