% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_community_download.R
\name{v1_summary_plot_report}
\alias{v1_summary_plot_report}
\title{helper creates bar/box/ridgeline plot for summary report}
\usage{
v1_summary_plot_report(
  row_index = NULL,
  input_plotkind_1pager,
  input_Custom_title_for_bar_plot_of_indicators = NULL,
  react_cur_button,
  react_sanitized_bt_rad_buff,
  react_data_processed,
  react_ratio.to.us.d
)
}
\arguments{
\item{row_index}{from shiny app, which row of table is the site of interest (should be related to react_cur_button)}

\item{input_plotkind_1pager}{from shiny app, from input$plotkind_1pager like "bar", "box", "ridgeline"}

\item{input_Custom_title_for_bar_plot_of_indicators}{from shiny app}

\item{react_cur_button}{from shiny app, such as "button_1" for sitenumber 1, or NULL (should be related to row_index)}

\item{react_sanitized_bt_rad_buff}{from shiny app, radius in miles}

\item{react_data_processed}{from shiny app, input table of places like input to ejamit()}

\item{react_ratio.to.us.d}{from shiny app}
}
\description{
helper creates bar/box/ridgeline plot for summary report
}
\keyword{internal}
