% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_xls_format.R
\name{varname2color_ejam}
\alias{varname2color_ejam}
\title{helper function - for color coding excel sheet columns}
\usage{
varname2color_ejam(varname, varnameinfo)
}
\arguments{
\item{varname}{things like us.avg.pctlowinc}

\item{varnameinfo}{must be left as default currently}
}
\value{
vector of colors
}
\description{
Convert R variable name of indicator to appropriate color for header row in Excel
}
\seealso{
\code{\link[=varinfo]{varinfo()}} \code{\link[=varname2vartype_ejam]{varname2vartype_ejam()}} \code{\link[=varname2varcategory_ejam]{varname2varcategory_ejam()}} \code{\link[=vartype_cat2color_ejam]{vartype_cat2color_ejam()}}
}
\keyword{internal}
