% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_xls_format.R
\name{varname2varcategory_ejam}
\alias{varname2varcategory_ejam}
\title{helper function - given indicator names, look up what category each is}
\usage{
varname2varcategory_ejam(varname, varnameinfo)
}
\arguments{
\item{varname}{vector of 1 or more names like "pctlowinc" as in unique(map_headernames$rname)}

\item{varnameinfo}{data.frame with info on type of each variable}
}
\value{
vector same size as varname
}
\description{
helper function - given indicator names, look up what category each is
}
\details{
tells if variable is "Demographic" "Environmental" "EJ Index" or "other"
as from dput(unique(map_headernames$varcategory))
}
\seealso{
\code{\link[=varinfo]{varinfo()}} \code{\link[=vartype_cat2color_ejam]{vartype_cat2color_ejam()}} \code{\link[=varname2color_ejam]{varname2color_ejam()}}
}
\keyword{internal}
