% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_xls_format.R
\name{varname2vartype_ejam}
\alias{varname2vartype_ejam}
\title{helper function - given indicator names, look up what type each is}
\usage{
varname2vartype_ejam(varname, varnameinfo)
}
\arguments{
\item{varname}{vector of 1 or more names}

\item{varnameinfo}{data.frame with info on type of each variable}
}
\value{
vector same size as varname
}
\description{
helper function - given indicator names, look up what type each is
}
\details{
The types are things like raw data count for indicator, average, percentile, etc.
}
\seealso{
\code{\link[=varinfo]{varinfo()}} \code{\link[=vartype_cat2color_ejam]{vartype_cat2color_ejam()}} \code{\link[=varname2color_ejam]{varname2color_ejam()}}
}
\keyword{internal}
