% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_xls_format.R
\name{vartype_cat2color_ejam}
\alias{vartype_cat2color_ejam}
\title{helper function - assign fill color to shade excel cells by indicator type and category}
\usage{
vartype_cat2color_ejam(vartype = raw, varcategory = "other")
}
\arguments{
\item{vartype}{must be one found in dput(unique(map_headernames$vartype))
like "usratio", "stateratio", "usraw", "stateraw",
"uspctile", "statepctile", "usavg", "stateavg", etc.
NA if not found.}

\item{varcategory}{must be one of "Demographic"   "Environmental" "EJ Index" "other"
as from dput(unique(map_headernames$varcategory))}
}
\value{
vector of colors like c('lightblue', 'gray') matching length of vartype
}
\description{
Use color shading to make spreadsheet easier to use, grouping the indicators
}
\seealso{
\code{\link[=varinfo]{varinfo()}} \code{\link[=varname2vartype_ejam]{varname2vartype_ejam()}}  \code{\link[=varname2varcategory_ejam]{varname2varcategory_ejam()}} \code{\link[=varname2color_ejam]{varname2color_ejam()}}
}
\keyword{internal}
