% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_x_anyother.R
\docType{data}
\name{x_anyother}
\alias{x_anyother}
\title{for internal use}
\format{
An object of class \code{character} of length 141.
}
\usage{
x_anyother
}
\description{
for internal use
}
\keyword{internal}
