% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_xls_varname2color.R
\name{xls_varname2color}
\alias{xls_varname2color}
\title{helper function - for color coding excel sheet columns}
\usage{
xls_varname2color(varname, varnameinfo)
}
\arguments{
\item{varname}{things like avg.pctlowinc}

\item{varnameinfo}{data.frame with info on type of each variable}
}
\value{
vector of colors
}
\description{
helper function - for color coding excel sheet columns
}
\seealso{
\code{\link[=xls_varname2vartype]{xls_varname2vartype()}} \code{\link[=xls_vartype2color]{xls_vartype2color()}} \code{\link[=xls_varname2color]{xls_varname2color()}}
}
\keyword{internal}
