% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_xls_varname2color.R
\name{xls_varname2vartype}
\alias{xls_varname2vartype}
\title{helper function - given indicator names, look up what type each is}
\usage{
xls_varname2vartype(varname, varnameinfo)
}
\arguments{
\item{varname}{vector of 1 or more names}

\item{varnameinfo}{data.frame with info on type of each variable, like map_headernames}
}
\value{
vector same size as varname
}
\description{
helper function - given indicator names, look up what type each is
}
\details{
The types are things like raw data count for indicator, average, percentile, etc.
Variable names are stored in column of map_headernames called newnames_ejscreenapi
Types are stored in column of map_headernames called jsondoc_vartype
}
\seealso{
\code{\link[=xls_varname2vartype]{xls_varname2vartype()}} \code{\link[=xls_vartype2color]{xls_vartype2color()}} \code{\link[=xls_varname2color]{xls_varname2color()}}
}
\keyword{internal}
