% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_blockgroupstats.R
\docType{data}
\name{blockgroupstats}
\alias{blockgroupstats}
\title{blockgroupstats (DATA) EJScreen demographic and environmental indicators for Census block groups}
\description{
The EJScreen dataset (demographic, environmental indicators).

For EJ Indexes, see \link{bgej}
}
\details{
\itemize{
\item Version 2.32.0 of EJAM / EJScreen uses ACS data from 2018-2022.
}

Each year this should be re-created as for the latest version.
See attributes(blockgroupstats)

It is a data.table of US Census blockgroups (not blocks).
With PR, and Island Areas

See \url{https://www.epa.gov/ejscreen}

Column names include bgfips, bgid (for join to blockwt$bgid), pop, pctlowinc, etc.
}
