% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_community_report.R
\name{build_community_report}
\alias{build_community_report}
\title{Generate EJScreen Multisite (EJAM) Summary Report in HTML within shiny app}
\usage{
build_community_report(
  output_df,
  analysis_title,
  totalpop,
  locationstr,
  include_ejindexes = FALSE,
  in_shiny = FALSE,
  filename = NULL
)
}
\arguments{
\item{output_df, }{single row of results table from doaggregate - either results_overall or one row of bysite}

\item{analysis_title, }{title to use in header of report}

\item{totalpop, }{total population included in location(s) analyzed}

\item{locationstr, }{description of the location(s) analyzed}

\item{include_ejindexes, }{whether to build tables for EJ and EJ supp. indexes}

\item{in_shiny, }{whether the function is being called in or outside of shiny - affects location of header}

\item{filename, }{path to file to save HTML content to; if null, returns as string (used in Shiny app)}
}
\description{
Creates a 2 page report on overall results or for one site, with demographic and environmental indicators, and EJ Indexes if needed.
}
\details{
For the same function for use in RStudio,
see \code{\link[=ejam2report]{ejam2report()}} which relies on \code{build_community_report()}

This function gets called by
app_server,
ejam2report(), and
community_report_template.Rmd
}
\keyword{internal}
