% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EJAM-package.R
\docType{package}
\name{EJAM}
\alias{EJAM}
\alias{EJAM-package}
\title{EJAM - Environmental Justice Analysis Multisite tool}
\description{
EJAM provides a way to run hundreds or thousands of EJScreen-style reports all at once, very quickly.
It is also an R package, with functions and data for very fast analysis of areas defined by
proximity to points like regulated facilities, areas defined as shapefiles/polygons,
or areas specified as Census units (e.g., Counties).
}
\section{\strong{What is EJAM and How to do EJ Analysis with the EJAM R Package}}{
\href{https://usepa.github.io/EJAM/index.html}{EJAM package documentation} provides
an overview of EJAM, a complete reference document on all functions, and
articles (vignettes) introducing key functions and how to use them.
}

\section{\strong{Data Sources -- Demographic and Environmental Indicators}}{
\href{https://www.epa.gov/ejscreen/understanding-ejscreen-results}{EJScreen documentation} provides
information about the sources, definitions, and vintage of the
demographic and environmental data and indicators used by
EJScreen/EJAM.

Each year, EJScreen/EJAM are updated to use newer demographic and environmental data,
and some indicators may be added, dropped, or revised during those updates.
For details on these updates and changes to datasets and indicators,
please see \href{https://www.epa.gov/ejscreen}{EJScreen} and in particular note the
information about updates and changes here:
\href{https://www.epa.gov/ejscreen/ejscreen-change-log}{EJScreen Change Log}.
Demographics are from the American Community Survey 5-year summary file data that are updated by
the United States Census Bureau annually and then used in EJScreen/EJAM.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://usepa.github.io/EJAM}
  \item \url{https://github.com/USEPA/EJAM}
}

}
\keyword{internal}
