% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots_ratios.R
\name{ejam2boxplot_ratios}
\alias{ejam2boxplot_ratios}
\title{Make boxplot of ratios to US averages}
\usage{
ejam2boxplot_ratios(
  ejamitout,
  radius,
  varnames = c(names_d, names_d_subgroups)
)
}
\arguments{
\item{ejamitout}{output from an EJAM analysis, like from \code{\link[=ejamit]{ejamit()}}}

\item{radius}{buffer radius used for an analysis}

\item{varnames}{currently only works with names_d and names_d_subgroups}
}
\value{
ggplot object
}
\description{
Make boxplot of ratios to US averages
}
\examples{
ejam2boxplot_ratios(testoutput_ejamit_1000pts_1miles, radius=1)
}
