% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2means.R
\name{ejam2means}
\alias{ejam2means}
\title{ejam2means - quick look at averages, via ejamit() results}
\usage{
ejam2means(ejamitout, vars = names_these, w = dt$pop)
}
\arguments{
\item{ejamitout}{as from ejamit()}

\item{vars}{colnames in ejamitout$results_bysite}

\item{w}{weights}
}
\value{
means in a useful format
}
\description{
ejam2means - quick look at averages, via ejamit() results
}
\examples{
# pctiles
ejam2quantiles(out,  vars = names_e_state_pctile, topx_insteadof_pctile = F) 
# ratios  
ejam2quantiles(out, vars = names_e_ratio_to_state_avg,  topx_insteadof_pctile = T)
# see the quantiles and mean together
t(rbind(
  ejam2quantiles(out, vars = names_e_ratio_to_state_avg, probs = 0.5, topx_insteadof_pctile = T), 
  AvgPerson = ejam2means(out, vars = names_e_ratio_to_state_avg)))
# these tell you the same thing:
out$results_summarized$keystats[rownames(out$results_summarized$keystats) \%in\% names_e_ratio_to_state_avg, ]
ejam2means(out, vars = names_e_ratio_to_state_avg)

}
