% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2means.R
\name{ejam2quantiles}
\alias{ejam2quantiles}
\title{ejam2quantiles - draft - info about quantiles of places or people}
\usage{
ejam2quantiles(
  ejamitout,
  vars = names_e_state_pctile,
  w = dt$pop,
  probs = c(50, 90)/100,
  topx_insteadof_pctile = T
)
}
\arguments{
\item{ejamitout}{as from ejamit()}

\item{vars}{colnames in ejamitout$results_bysite}

\item{w}{weights}

\item{probs}{as written is always the pctile, but default is to report on the top X\%, or 100-probs}

\item{topx_insteadof_pctile}{logical}
}
\value{
info about quantiles of places or people
}
\description{
ejam2quantiles - draft - info about quantiles of places or people
}
\keyword{internal}
