% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejam2report.R
\name{ejam2report}
\alias{ejam2report}
\title{View HTML Report on EJAM Results (Overall or at 1 Site)}
\usage{
ejam2report(
  ejamitout = testoutput_ejamit_10pts_1miles,
  sitenumber = NULL,
  analysis_title = "EJAM Report",
  submitted_upload_method = c("latlon", "SHP", "FIPS")[1],
  data_up_shp = NA,
  return_html = FALSE,
  launch_browser = TRUE
)
}
\arguments{
\item{ejamitout}{output as from \code{\link[=ejamit]{ejamit()}}, list with a data.table called \code{results_bysite}
if sitenumber parameter is used, or a data.table called \code{results_overall} otherwise}

\item{sitenumber}{If a number is provided, the report is about
\code{ejamout$results_bysite[sitenumber, ]} and if no number is provided (param is NULL)
then the report is about \code{ejamout$results_overall}}

\item{analysis_title}{optional title of analysis}

\item{submitted_upload_method}{\code{"latlon"} is the default and other options are not implemented yet.}

\item{data_up_shp}{not implemented yet, for upload method SHP}

\item{return_html}{set TRUE to have function return HTML object instead of URL of local file}

\item{launch_browser}{set TRUE to have it launch browser and show report.}
}
\value{
URL of temp html file or object depending on return_html,
and has side effect of launching browser to view it depending on return_html
}
\description{
Get URL for and view in browser a 2-page summary report similar to the
EJScreen Community Report
}
\details{
This relies on \code{\link[=build_community_report]{build_community_report()}} as used in web app
for viewing report on 1 site from a list of sites.
}
\examples{
#out <- ejamit(testpoints_10, radius = 3, include_ejindexes = T)
out <- testoutput_ejamit_10pts_1miles

ejam2report(out)
table_gt_from_ejamit_overall(out$results_overall)
table_tall_from_overall(out$results_overall)

x <- ejam2report(out, sitenumber = 1)
table_gt_from_ejamit_1site(out$results_bysite[1, ])
browseURL(x)

}
