% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frs.R
\name{frs}
\alias{frs}
\title{frs (DATA) EPA Facility Registry Service table of regulated sites}
\description{
This is a data.table snapshot version of the EPA FRS.
You can look up sites by REGISTRY_ID in \link{frs}, and get their location, etc.
}
\details{
This dataset can be updated by a package maintainer by using
frs_update_datasets() (which is not an exported function)

The definitions of active/inactive here are not quite the
same as used in ECHO. See attributes(frs) to see date created, etc.

Also, EJScreen has maps of EPA-regulated facilities of a few program types,
as provided here: \url{https://www.epa.gov/ejscreen/ejscreen-map-descriptions#sites-reporting-to-epa}
and for a table of acronym definitions
see https://www.epa.gov/sites/default/files/2021-05/frs_program_abbreviations_and_names.xlsx
and \link{epa_programs_defined}
\itemize{
\item Count of    all REGISTRY_ID rows:   Approx 7 million
\item Count of unique REGISTRY_ID values: Approx 4-5 million
\item Clearly inactive unique IDs:        Approx 1-2 million
\item Assumed   active unique IDs:        Approx 3 million
\item frs rows total:            Approx 2-3 million
\item frs_by_programid rows:     Approx 3-4 million
\item frs_by_naics rows:         Approx 700k
\item frs_by_sic rows:           Approx 700k
}

Classes ‘data.table’ and 'data.frame'

colnames
\itemize{
\item \link{1,} "lat"
\item \link{2,} "lon"
\item \link{3,} "REGISTRY_ID"
\item \link{4,} "PRIMARY_NAME"
\item \link{5,} "NAICS"
\item \link{6,} "SIC"
\item \link{7,} "PGM_SYS_ACRNMS"
}
}
\seealso{
\link{epa_programs} \link{epa_programs_defined} \link{frs_by_programid}  \link{frs_by_naics} \link{frs_by_sic}
}
