% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_report_helper_funs.R
\name{generate_html_header}
\alias{generate_html_header}
\title{Build HTML header for community report}
\usage{
generate_html_header(
  analysis_title,
  totalpop,
  locationstr,
  in_shiny = FALSE,
  report_title = NULL
)
}
\arguments{
\item{analysis_title, }{title to use in header of report}

\item{totalpop, }{total population included in location(s) analyzed}

\item{locationstr, }{description of the location(s) analyzed}

\item{in_shiny, }{whether the function is being called in or outside of shiny - affects location of header}

\item{report_title}{generic name of this type of report, to be shown at top, like "EJScreen-EJAM Multisite Report" or "EJAM Multisite Report"}
}
\description{
Build HTML header for community report
}
\keyword{internal}
