% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROXIMITY_FUNCTIONS.R
\name{getfrsnearby}
\alias{getfrsnearby}
\title{Find all EPA-regulated facilities nearby each specified point and distances}
\usage{
getfrsnearby(
  frompoints,
  radius = 3,
  maxradius = 31.07,
  avoidorphans = FALSE,
  quadtree = NULL,
  quiet = FALSE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{frompoints}{used as the sitepoints param of \code{\link[=getblocksnearby]{getblocksnearby()}}}

\item{radius}{passed to \code{\link[=getblocksnearby]{getblocksnearby()}}}

\item{maxradius}{passed to \code{\link[=getblocksnearby]{getblocksnearby()}}}

\item{avoidorphans}{passed to \code{\link[=getblocksnearby]{getblocksnearby()}}}

\item{quadtree}{index that should be created by indexpoints}

\item{quiet}{passed to \code{\link[=getblocksnearby]{getblocksnearby()}}}

\item{parallel}{passed to \code{\link[=getblocksnearby]{getblocksnearby()}}}

\item{...}{passed to \code{\link[=getblocksnearby]{getblocksnearby()}}}

\item{quaddatatable}{optional table of topoints (in
format provided by internal helper function create_quaddata() as needed).}
}
\value{
sites2points data.table one row per pair of frompoint and nearby frs point,
like output of \code{\link[=getblocksnearby]{getblocksnearby()}}
}
\description{
Given a table of frompoints (lat lon coordinates),
find IDs of and distances to all nearby points
that represent EPA-regulated facilities with locations in Facility Registry Services (FRS).
Like \code{\link[=getblocksnearby]{getblocksnearby()}} but for regulated facilities in US EPA FRS
}
\details{
Later steps can aggregate at each frompoint to summarize
\itemize{
\item count facilities near each frompoint
\item max/min distance for each frompoint, like proximity of nearest, etc.
\item a proximity score for each block and then each blockgroup
}
}
