% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlon_from_address.R
\name{latlon_from_address_table}
\alias{latlon_from_address_table}
\title{get lat,lon from table that contains USPS addresses}
\usage{
latlon_from_address_table(x)
}
\arguments{
\item{x}{data.frame or can be missing if interactive}
}
\value{
same as output of \code{\link[=latlon_from_address]{latlon_from_address()}}
}
\description{
get lat,lon from table that contains USPS addresses
}
\examples{
address_from_table(test_address_table)

## fname <- system.file("testdata/address/street_address_9.xlsx", package = "EJAM")
## test_addresses_9b <- address_from_table(fname)
\dontrun{

# This requires first attaching the AOI package.

pts <- latlon_from_address(test_addresses_9[1:2])
## out <- ejamit(pts, radius = 1)
## ejam2report(out)

latlon_from_address_table(test_address_table)
latlon_from_address_table(test_address_table_withfull)
## *** NOTE IT FAILS IF A COLUMN WITH STREET NAME ONLY IS CALLED "address"
##   instead of that storing the full address.
}
fixcolnames_infer(currentnames = test_address_parts1)
fixcolnames_infer(currentnames = names(test_address_table))

}
