% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridgeline_ratios.R
\name{plot_ridgeline_ratios}
\alias{plot_ridgeline_ratios}
\title{Make ridgeline plot of ratios of demographic score to its average}
\usage{
plot_ridgeline_ratios(ratio.to.us.d.bysite, shortlabels = NULL)
}
\arguments{
\item{ratio.to.us.d.bysite}{named list of a few ratios to plot (data.frame)}

\item{shortlabels}{names to use for plot - should be same length as named list ratio.to.us.d.overall}
}
\description{
Make ridgeline plot of ratios of demographic score to its average
}
\examples{
 out <- testoutput_ejamit_1000pts_1miles
 ratio.to.us.d.bysite <- out$results_bysite[ ,  c(
   ..names_d_ratio_to_avg, ..names_d_subgroups_ratio_to_avg
   )]
 # plot_ridgeline_ratios(ratio.to.us.d.bysite)
 # cap the ratio, for better plot
 x <- as.matrix(ratio.to.us.d.bysite)
 x[x > 5] <- 5
 plot_ridgeline_ratios(data.table::data.table(x))

}
