% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_report_community_download.R
\name{report_setup_temp_files}
\alias{report_setup_temp_files}
\title{helper - copies css, logo, and .Rmd template to tempdir for render of report}
\usage{
report_setup_temp_files(
  Rmd_name = "community_report_template.Rmd",
  Rmd_folder = "report/community_report/"
)
}
\arguments{
\item{Rmd_name}{.Rmd filename the package uses}

\item{Rmd_folder}{folder the package stores the template in}
}
\description{
helper - copies css, logo, and .Rmd template to tempdir for render of report
}
\keyword{internal}
