% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_run_app.R
\name{run_app}
\alias{run_app}
\alias{app_run_EJAM}
\title{Launch the Shiny Application in RStudio}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = "server",
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{...}{arguments to pass to golem_opts.
This can be used to pass some of the variables that otherwise would be set to
defaults defined in global.R, and currently can include the following:

\if{html}{\out{<div class="sourceCode">}}\preformatted{run_app(
  default_default_miles=3.1, 
  default_max_miles=31, 
  default_max_mb_upload=100, 
  default_default_miles_shapefile = 0.5, 
  advanced=TRUE, 
  testing=TRUE, 
  shiny.testmode=TRUE, 
  # and this will be implemented at some point: 
  sitepoints = mytable such as testpoints_100
  ## or maybe
  ## sitepoints = 'latlondata.xlsx'
)
}\if{html}{\out{</div>}}

This uses an internal function global_or_param()}
}
\value{
An object that represents the app. Printing the object or passing it to \code{\link[=runApp]{runApp()}} will run the app.
}
\description{
Launches shiny web app from RStudio - \code{\link[=app_run_EJAM]{app_run_EJAM()}} and \code{\link[=run_app]{run_app()}} are the same
}
\details{
run_app( isPublic = TRUE ) will launch the public-facing version.
}
\seealso{
\code{\link[=app_run_EJAM]{app_run_EJAM()}}
}
