% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes_from_fips.R
\name{shapes_from_fips}
\alias{shapes_from_fips}
\title{Download shapefiles based on FIPS codes of States, Counties, Cities/CDPs, Tracts, or Blockgroups}
\usage{
shapes_from_fips(fips)
}
\arguments{
\item{fips}{vector of one or more Census FIPS codes such as from \code{\link[=name2fips]{name2fips()}}}
}
\value{
spatial data.frame with one row per fips (assuming any fips are valid)
}
\description{
Download shapefiles based on FIPS codes of States, Counties, Cities/CDPs, Tracts, or Blockgroups
}
\examples{
 fipslist = list(
  statefips = name2fips(c('DE', 'RI')),
  countyfips = fips_counties_from_state_abbrev(c('DE')),
  cityfips = name2fips(c('chelsea,MA', 'st. john the baptist parish, LA')),
  tractfips = substr(blockgroupstats$bgfips[300:301], 1, 12),
  bgfips = blockgroupstats$bgfips[300:301]
  )
  shp <- list()
  \dontrun{
   for (i in seq_along(fipslist)) {
    shp[[i]] <- shapes_from_fips(fipslist[[i]])
    print(shp[[i]])
    # mapfast(shp[[i]])
   }
  }
  

}
